Retrieve a single analysis result set by its identifier

Retrieve a single analysis result set by its identifier

.getResultSets(
  resultSet = NA_character_,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

resultSet

An expression analysis result set numerical identifier.

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

Varies

Examples

gemma.R:::.getResultSets(523099)
#>               Probe NCBIid    GeneSymbol
#>              <char> <char>        <char>
#>     1:   1453255_at  72401       Slc43a1
#>     2: 1432455_a_at  73244        Prl8a1
#>     3: 1437709_x_at 234875         Ttc13
#>     4:   1456216_at                     
#>     5: 1438337_x_at 320454     Tmprss11g
#>    ---                                  
#> 45033:   1427375_at 108943       Trmt10a
#> 45034:   1422999_at  53859       Map3k14
#> 45035:   1458474_at                     
#> 45036:   1435812_at                     
#> 45037:   1431915_at  75800 4930442E04Rik
#>                                                 GeneName pvalue
#>                                                   <char>  <num>
#>     1:                solute carrier family 43, member 1 0.6822
#>     2:         prolactin family 8, subfamily a, member 1 0.3379
#>     3:                tetratricopeptide repeat domain 13 0.1769
#>     4:                                                   0.4880
#>     5:                transmembrane protease, serine 11g 0.8458
#>    ---                                                         
#> 45033:                        tRNA methyltransferase 10A 0.5115
#> 45034: mitogen-activated protein kinase kinase kinase 14 0.8922
#> 45035:                                                   0.0751
#> 45036:                                                   0.0725
#> 45037:                        RIKEN cDNA 4930442E04 gene 0.5512
#>        corrected_pvalue   rank contrast_16452_16455_log2fc
#>                   <num>  <num>                       <num>
#>     1:                1 0.5902                     -0.0826
#>     2:                1 0.2551                      0.2536
#>     3:                1 0.1232                     -0.1846
#>     4:                1 0.3920                      0.3658
#>     5:                1 0.7808                     -0.0657
#>    ---                                                    
#> 45033:                1 0.4153                     -0.1661
#> 45034:                1 0.8400                     -0.0597
#> 45035:                1 0.0502                      0.1284
#> 45036:                1 0.0483                     -0.3665
#> 45037:                1 0.4539                     -0.1580
#>        contrast_16452_16455_tstat contrast_16452_16455_pvalue
#>                             <num>                       <num>
#>     1:                    -0.3684                      0.7153
#>     2:                     1.4925                      0.1466
#>     3:                    -1.2397                      0.2253
#>     4:                     1.6907                      0.1019
#>     5:                    -0.8012                      0.4297
#>    ---                                                       
#> 45033:                    -0.8684                      0.3925
#> 45034:                    -0.2951                      0.7701
#> 45035:                     0.8711                      0.3910
#> 45036:                    -2.5688                      0.0158
#> 45037:                    -0.9948                      0.3283
#>        contrast_16452_16456_log2fc contrast_16452_16456_tstat
#>                              <num>                      <num>
#>     1:                      0.1881                     0.7916
#>     2:                     -0.0361                    -0.2001
#>     3:                     -0.1967                    -1.2457
#>     4:                      0.3087                     1.3456
#>     5:                     -0.0025                    -0.0290
#>    ---                                                       
#> 45033:                     -0.0848                    -0.4178
#> 45034:                      0.1502                     0.6998
#> 45035:                      0.2874                     1.8381
#> 45036:                     -0.1914                    -1.2646
#> 45037:                     -0.1699                    -1.0090
#>        contrast_16452_16456_pvalue contrast_16453_16455_log2fc
#>                              <num>                       <num>
#>     1:                      0.4352                      0.1561
#>     2:                      0.8429                      0.1084
#>     3:                      0.2231                      0.1335
#>     4:                      0.1891                      0.0596
#>     5:                      0.9771                     -0.0621
#>    ---                                                        
#> 45033:                      0.6793                      0.0796
#> 45034:                      0.4898                     -0.0250
#> 45035:                      0.0766                      0.1272
#> 45036:                      0.2163                     -0.1772
#> 45037:                      0.3215                      0.1005
#>        contrast_16453_16455_tstat contrast_16453_16455_pvalue
#>                             <num>                       <num>
#>     1:                     0.6967                      0.4917
#>     2:                     0.6380                      0.5286
#>     3:                     0.8967                      0.3775
#>     4:                     0.2757                      0.7848
#>     5:                    -0.7570                      0.4553
#>    ---                                                       
#> 45033:                     0.4160                      0.6806
#> 45034:                    -0.1233                      0.9027
#> 45035:                     0.8629                      0.3955
#> 45036:                    -1.2417                      0.2245
#> 45037:                     0.6330                      0.5318
#>        contrast_16453_16456_log2fc contrast_16453_16456_tstat
#>                              <num>                      <num>
#>     1:                      0.0620                     0.2765
#>     2:                     -0.0644                    -0.3790
#>     3:                      0.1483                     0.9959
#>     4:                     -0.0502                    -0.2320
#>     5:                     -0.0013                    -0.0155
#>    ---                                                       
#> 45033:                      0.2207                     1.1537
#> 45034:                      0.1441                     0.7121
#> 45035:                      0.0692                     0.4696
#> 45036:                      0.0725                     0.5079
#> 45037:                      0.0298                     0.1878
#>        contrast_16453_16456_pvalue contrast_16454_16455_log2fc
#>                              <num>                       <num>
#>     1:                      0.7842                      0.0619
#>     2:                      0.7075                     -0.0482
#>     3:                      0.3277                      0.1791
#>     4:                      0.8182                      0.0543
#>     5:                      0.9877                      0.0420
#>    ---                                                        
#> 45033:                      0.2583                      0.0257
#> 45034:                      0.4823                     -0.1852
#> 45035:                      0.6422                      0.1622
#> 45036:                      0.6154                     -0.0316
#> 45037:                      0.8523                     -0.0032
#>        contrast_16454_16455_tstat contrast_16454_16455_pvalue
#>                             <num>                       <num>
#>     1:                     0.2761                      0.7845
#>     2:                    -0.2836                      0.7788
#>     3:                     1.2031                      0.2389
#>     4:                     0.2508                      0.8038
#>     5:                     0.5118                      0.6128
#>    ---                                                       
#> 45033:                     0.1341                      0.8942
#> 45034:                    -0.9155                      0.3677
#> 45035:                     1.1003                      0.2805
#> 45036:                    -0.2213                      0.8265
#> 45037:                    -0.0202                      0.9840
#>        contrast_16454_16456_log2fc contrast_16454_16456_tstat
#>                              <num>                      <num>
#>     1:                      0.2563                     1.1436
#>     2:                     -0.2939                    -1.7296
#>     3:                     -0.0020                    -0.0137
#>     4:                     -0.1233                    -0.5700
#>     5:                      0.0355                     0.4329
#>    ---                                                       
#> 45033:                      0.2718                     1.4208
#> 45034:                     -0.0218                    -0.1078
#> 45035:                     -0.1615                    -1.0957
#> 45036:                      0.2080                     1.4580
#> 45037:                     -0.1912                    -1.2040
#>        contrast_16454_16456_pvalue
#>                              <num>
#>     1:                      0.2624
#>     2:                      0.0946
#>     3:                      0.9892
#>     4:                      0.5732
#>     5:                      0.6683
#>    ---                            
#> 45033:                      0.1663
#> 45034:                      0.9149
#> 45035:                      0.2824
#> 45036:                      0.1558
#> 45037:                      0.2386