Return all supported filter properties

Some functions such as get_datasets and get_platforms_by_ids include a filter argument that allows creation of more complex queries. This function returns a list of supported properties to be used in those filters

filter_properties()

Value

A list of data.tables that contain supported properties and their data types

Examples

filter_properties()
#> $dataset
#>                                  properties    type
#>                                      <char>  <char>
#>   1:                    accession.accession  string
#>   2:             accession.accessionVersion  string
#>   3:      accession.externalDatabase.ftpUri  string
#>   4:          accession.externalDatabase.id integer
#>   5: accession.externalDatabase.lastUpdated  string
#>  ---                                               
#> 130:                    taxon.isGenesUsable boolean
#> 131:                           taxon.ncbiId integer
#> 132:                   taxon.scientificName  string
#> 133:                  taxon.secondaryNcbiId integer
#> 134:                               troubled boolean
#>                             description
#>                                  <char>
#>   1:                               <NA>
#>   2:                               <NA>
#>   3:                               <NA>
#>   4:                               <NA>
#>   5:                               <NA>
#>  ---                                   
#> 130:                               <NA>
#> 131:                               <NA>
#> 132:                               <NA>
#> 133:                               <NA>
#> 134: alias for curationDetails.troubled
#> 
#> $platform
#>                                          properties    type
#>                                              <char>  <char>
#>   1:               advertisedNumberOfDesignElements integer
#>   2: alternativeTo.advertisedNumberOfDesignElements integer
#>   3:          alternativeTo.compositeSequences.size integer
#>   4:               alternativeTo.curationDetails.id integer
#>   5:      alternativeTo.curationDetails.lastUpdated  string
#>  ---                                                       
#> 126:                                   taxon.ncbiId integer
#> 127:                           taxon.scientificName  string
#> 128:                          taxon.secondaryNcbiId integer
#> 129:                                 technologyType  string
#> 130:                                       troubled boolean
#>                                 description
#>                                      <char>
#>   1:                                   <NA>
#>   2:                                   <NA>
#>   3:                                   <NA>
#>   4:                                   <NA>
#>   5:                                   <NA>
#>  ---                                       
#> 126:          alias for primaryTaxon.ncbiId
#> 127:  alias for primaryTaxon.scientificName
#> 128: alias for primaryTaxon.secondaryNcbiId
#> 129:                                   <NA>
#> 130:     alias for curationDetails.troubled
#> 
#> $resultSet
#>                                                  properties    type description
#>                                                      <char>  <char>      <lgcl>
#>  1:                          analysis.experimentAnalyzed.id integer          NA
#>  2:                        analysis.experimentAnalyzed.name  string          NA
#>  3:                                             analysis.id integer          NA
#>  4:                       analysis.numberOfElementsAnalyzed integer          NA
#>  5:                                analysis.resultSets.size integer          NA
#>  6:     analysis.subsetFactorValue.characteristics.category  string          NA
#>  7:  analysis.subsetFactorValue.characteristics.categoryUri  string          NA
#>  8: analysis.subsetFactorValue.characteristics.evidenceCode  string          NA
#>  9:           analysis.subsetFactorValue.characteristics.id integer          NA
#> 10:         analysis.subsetFactorValue.characteristics.name  string          NA
#> 11:         analysis.subsetFactorValue.characteristics.size integer          NA
#> 12:        analysis.subsetFactorValue.characteristics.value  string          NA
#> 13:     analysis.subsetFactorValue.characteristics.valueUri  string          NA
#> 14:                           analysis.subsetFactorValue.id integer          NA
#> 15:                  baselineGroup.characteristics.category  string          NA
#> 16:               baselineGroup.characteristics.categoryUri  string          NA
#> 17:              baselineGroup.characteristics.evidenceCode  string          NA
#> 18:                        baselineGroup.characteristics.id integer          NA
#> 19:                      baselineGroup.characteristics.name  string          NA
#> 20:                      baselineGroup.characteristics.size integer          NA
#> 21:                     baselineGroup.characteristics.value  string          NA
#> 22:                  baselineGroup.characteristics.valueUri  string          NA
#> 23:                     baselineGroup.experimentalFactor.id integer          NA
#> 24:                   baselineGroup.experimentalFactor.name  string          NA
#> 25:                   baselineGroup.experimentalFactor.type  string          NA
#> 26:                                        baselineGroup.id integer          NA
#> 27:                            baselineGroup.measurement.id integer          NA
#> 28:                        baselineGroup.measurement.kindCV  string          NA
#> 29:                     baselineGroup.measurement.otherKind  string          NA
#> 30:                baselineGroup.measurement.representation  string          NA
#> 31:                          baselineGroup.measurement.type  string          NA
#> 32:                         baselineGroup.measurement.value  string          NA
#> 33:                                                      id integer          NA
#> 34:                                     numberOfGenesTested integer          NA
#> 35:                                    numberOfProbesTested integer          NA
#> 36:                                            results.size integer          NA
#>                                                  properties    type description
#>