Dataset data

Retrieves the data for the given dataset

getDatasetData(
  dataset = NA_character_,
  filter = FALSE,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset

Required, part of the URL path. Can either be the dataset ID or its short name (e.g. GSE1234). Retrieval by ID is more efficient. Only datasets that user has access to will be available

filter

Optional, defaults to empty. Filtering can be done on any* property or nested property that the appropriate object class defines or inherits (and that is mapped by hibernate). [These do not correspond to the properties of the objects returned by the API calls.].description-imp Class definitions:

Datasets: javaDoc gitHub

Platforms: javaDoc gitHub E.g: curationDetails or curationDetails.lastTroubledEvent.date.

Any property of a supported type. Currently supported types are:

String - property of String type, required value can be any String.

Number - any Number implementation. Required value must be a string parseable to the specific Number type.

Boolean - required value will be parsed to true only if the string matches 'true', ignoring case. Accepted operator keywords are:

'=' - equality

'!=' - non-equality

'<' - smaller than

'>' - larger than

'<=' - smaller or equal

'=>' - larger or equal

'like' - similar string, effectively means 'contains', translates to the sql 'LIKE' operator (given value will be surrounded by % signs) Multiple filters can be chained using AND and OR keywords. Leave space between the keywords and the previous/next word! E.g: ?filter=property1 < value1 AND property2 like value2 If chained filters are mixed conjunctions and disjunctions, the query must be in conjunctive normal form (CNF). Parentheses are not necessary

every AND keyword separates blocks of disjunctions. Example: ?filter=p1 = v1 OR p1 != v2 AND p2 <= v2 AND p3 > v3 OR p3 < v4 Above query will translate to: (p1 = v1 OR p1 != v2) AND (p2 <= v2) AND (p3 > v3 OR p3 < v4;) Breaking the CNF results in an error. Filter curationDetails.troubled will be ignored if user is not an administrator.

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async

TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised

Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

The data file for the given dataset. A 404 error if the given identifier does not map to any object.

Examples

dat <- getDatasetData("GSE2018")
str(dat)
#> Classes ‘data.table’ and 'data.frame':	22283 obs. of  38 variables:
#>  $ Probe       : chr  "1007_s_at" "1053_at" "117_at" "121_at" ...
#>  $ GeneSymbol  : chr  "DDR1" "RFC2" "HSPA7|HSPA6" "PAX8" ...
#>  $ GeneName    : chr  "discoidin domain receptor tyrosine kinase 1" "replication factor C subunit 2" "heat shock protein family A (Hsp70) member 7 (pseudogene)|heat shock protein family A (Hsp70) member 6" "paired box 8" ...
#>  $ NCBIid      : chr  "780" "5982" "3311|3310" "7849" ...
#>  $ BAL_13_A0B0 : num  7.66 6.82 7.7 8.87 4.02 ...
#>  $ BAL_44b_A0B0: num  7.56 6.82 7.74 8.73 4.02 ...
#>  $ BAL_41_A1B0 : num  7.9 6.43 8.3 9.43 4.27 ...
#>  $ BAL_31_A0B1 : num  7.48 6.87 7.71 9.11 4.13 ...
#>  $ BAL_12_A1B0 : num  7.61 7.08 7.5 9.13 4.13 ...
#>  $ BAL_15b_A0B0: num  7.68 6.77 8.43 9.06 3.99 ...
#>  $ BAL_40_A0B0 : num  7.57 6.75 7.31 9.13 4.06 ...
#>  $ BAL_47b_A0B1: num  7.61 6.79 7.8 9.02 4.18 ...
#>  $ BAL_47c_A0B1: num  7.63 6.91 7.47 9.17 3.96 ...
#>  $ BAL_45a_A1B0: num  7.45 6.8 7.62 9.29 4.03 ...
#>  $ BAL_47a_A0B0: num  7.79 6.44 7.84 9.04 3.91 ...
#>  $ BAL_44a_A0B0: num  7.73 6.94 7.6 9.25 4 ...
#>  $ BAL_39_A0B1 : num  7.27 6.81 7.33 8.88 4.02 ...
#>  $ BAL_33_A0B1 : num  7.43 7.14 7.57 8.95 4.03 ...
#>  $ BAL_29_A1B0 : num  7.91 6.43 7.79 9.59 4.13 ...
#>  $ BAL_25_A0B0 : num  7.57 6.89 7.44 8.92 3.98 ...
#>  $ BAL_27a_A0B0: num  7.91 6.98 7.78 9.25 4.03 ...
#>  $ BAL_1_A0B0  : num  7.52 7.16 7.84 9.06 4.17 ...
#>  $ BAL_15a_A0B1: num  7.63 6.91 7.69 8.76 4.03 ...
#>  $ BAL_43_A0B0 : num  7.33 6.97 7.46 9.09 4 ...
#>  $ BAL_42_A0B0 : num  7.89 6.65 8.58 8.93 4.11 ...
#>  $ BAL_4_A0B0  : num  7.47 6.95 7.74 9.12 4.04 ...
#>  $ BAL_3_A1B0  : num  7.54 6.9 7.54 8.87 3.86 ...
#>  $ BAL_18_A1B0 : num  7.41 6.78 7.93 9.15 4.05 ...
#>  $ BAL_17_A0B0 : num  7.23 6.96 7.97 8.81 4.02 ...
#>  $ BAL_21_A0B0 : num  7.87 6.62 7.42 9.55 4.14 ...
#>  $ BAL_19_A0B0 : num  7.82 6.71 7.83 9.57 3.99 ...
#>  $ BAL_45b_A2B0: num  7.78 6.59 7.93 9 3.94 ...
#>  $ BAL_26_A2B0 : num  7.26 6.94 7.97 8.66 4.01 ...
#>  $ BAL_20a_A1B1: num  NaN NaN NaN NaN NaN NaN NaN NaN NaN NaN ...
#>  $ BAL_20b_A1B1: num  7.68 6.65 7.74 9.24 4.03 ...
#>  $ BAL_37_A1B1 : num  7.47 6.86 7.69 9.06 4.06 ...
#>  $ BAL_46_A2B0 : num  7.93 6.54 7.81 10.08 3.86 ...
#>  $ BAL_27b_A2B1: num  7.7 6.85 7.56 9.14 4.12 ...
#>  - attr(*, ".internal.selfref")=<externalptr>