getDatasetResultSets

Lists the available resultSets for the queried dataset.

getDatasetResultSets(
  dataset,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset

Required, part of the URL path. Can either be the dataset ID or its short name (e.g. GSE1234). Retrieval by ID is more efficient. Only datasets that user has access to will be available

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forgetGemmaMemoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

A data table with the queried dataset's resultSet ID(s). Use getDatasetDE to get differential expression values (see examples).

Examples

resultSets <- getDatasetResultSets("GSE2018")
getDatasetDE(resultSet = resultSets$resultSet.id)
#> $`500184`
#>              Probe        NCBIid              GeneSymbol
#>     1:   205105_at          4124                  MAN2A1
#>     2:   217563_at                                      
#>     3:   206819_at                                      
#>     4:   219446_at         55188                   RIC8B
#>     5:   206630_at          7299                     TYR
#>    ---                                                  
#> 22267:   213691_at     101928378               PTOV1-AS2
#> 22268: 211485_s_at          8817                   FGF18
#> 22269: 220936_s_at         55766                    H2AJ
#> 22270: 205013_s_at 135|101730217 ADORA2A|SPECC1L-ADORA2A
#> 22271:   220392_at         64641                    EBF2
#>                                                                  GeneName
#>     1:                                mannosidase alpha class 2A member 1
#>     2:                                                                   
#>     3:                                                                   
#>     4:                          RIC8 guanine nucleotide exchange factor B
#>     5:                                                         tyrosinase
#>    ---                                                                   
#> 22267:                                              PTOV1 antisense RNA 2
#> 22268:                                        fibroblast growth factor 18
#> 22269:                                                      H2A.J histone
#> 22270: adenosine A2a receptor|SPECC1L-ADORA2A readthrough (NMD candidate)
#> 22271:                                         EBF transcription factor 2
#>          pvalue corrected_pvalue    rank
#>     1: 0.691200           0.8827 0.78310
#>     2: 0.023940           0.2607 0.09178
#>     3: 0.029350           0.2801 0.10470
#>     4: 0.680000           0.8786 0.77400
#>     5: 0.416300           0.7360 0.56560
#>    ---                                  
#> 22267: 0.160400           0.5186 0.30930
#> 22268: 0.154700           0.5131 0.30150
#> 22269: 0.375100           0.7092 0.52890
#> 22270: 0.373800           0.7085 0.52760
#> 22271: 0.001836           0.1074 0.01702
#>        contrast_Transplanted Organ Complication, Acute rejection_logFoldChange
#>     1:                                                                 0.07823
#>     2:                                                                 0.17390
#>     3:                                                                 0.14450
#>     4:                                                                -0.05527
#>     5:                                                                -0.06343
#>    ---                                                                        
#> 22267:                                                                -0.14890
#> 22268:                                                                -0.09667
#> 22269:                                                                 0.05851
#> 22270:                                                                 0.20720
#> 22271:                                                                -0.16590
#>        contrast_Transplanted Organ Complication, Acute rejection_tstat
#>     1:                                                          0.4009
#>     2:                                                          2.3750
#>     3:                                                          2.2840
#>     4:                                                         -0.4163
#>     5:                                                         -0.8240
#>    ---                                                                
#> 22267:                                                         -1.4380
#> 22268:                                                         -1.4590
#> 22269:                                                          0.8999
#> 22270:                                                          0.9023
#> 22271:                                                         -3.4070
#>        contrast_Transplanted Organ Complication, Acute rejection_pvalue
#>     1:                                                         0.691200
#>     2:                                                         0.023940
#>     3:                                                         0.029350
#>     4:                                                         0.680000
#>     5:                                                         0.416300
#>    ---                                                                 
#> 22267:                                                         0.160400
#> 22268:                                                         0.154700
#> 22269:                                                         0.375100
#> 22270:                                                         0.373800
#> 22271:                                                         0.001836
#>