Dataset samples

Retrieves samples for the given dataset

getDatasetSamples(
  dataset,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset

Required, part of the URL path. Can either be the dataset ID or its short name (e.g. GSE1234). Retrieval by ID is more efficient. Only datasets that user has access to will be available

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forgetGemmaMemoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

A data table with information about the samples of the queried dataset. A 404 error if the given identifier does not map to any object.

Examples

dat <- getDatasetSamples("GSE2018")
head(dat)
#>                   bioMaterial.Name  sample.Name sample.ID
#> 1:  GSE2018_bioMaterial_4|GSM36426   BAL_1_A0B0        31
#> 2: GSE2018_bioMaterial_15|GSM36437 BAL_27a_A0B0        20
#> 3: GSE2018_bioMaterial_26|GSM36448  BAL_43_A0B0         9
#> 4:  GSE2018_bioMaterial_6|GSM36428  BAL_13_A0B0        29
#> 5: GSE2018_bioMaterial_17|GSM36439  BAL_29_A1B0        18
#> 6: GSE2018_bioMaterial_28|GSM36450 BAL_44b_A0B0         7
#>                                                 sample.Correspondence
#> 1:                                Corresponding to Samples:  GSM36426
#> 2:                                Corresponding to Samples:  GSM36437
#> 3: Corresponding to samples:  ial corresponding to samples:  GSM36448
#> 4:                                Corresponding to Samples:  GSM36428
#> 5:                                Corresponding to Samples:  GSM36439
#> 6: Corresponding to samples:  ial corresponding to samples:  GSM36450
#>                                                                                                                                                                                                                                       sample.Description
#> 1:                                                             Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36426\nLast updated (according to GEO): May 27 2005
#> 2:  Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection). One of two samples from the same patient, different visits\nSource GEO sample is GSM36437\nLast updated (according to GEO): May 27 2005
#> 3:                                                             Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36448\nLast updated (according to GEO): May 27 2005
#> 4:                                                             Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36428\nLast updated (according to GEO): May 27 2005
#> 5:                                                             Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=1 and B=0 (Non-acute rejection).\nSource GEO sample is GSM36439\nLast updated (according to GEO): May 27 2005
#> 6:  Bronchoalveolar lavage cells from a patient with an ISHLT biopsy score of A=0 and B=0 (Non-acute rejection). One of two samples from the same patient, different visits\nSource GEO sample is GSM36450\nLast updated (according to GEO): May 27 2005
#>    sample.Outlier sample.Accession sample.Processed sample.Characteristics
#> 1:          FALSE         GSM36426     1.036695e+12      <data.table[2x4]>
#> 2:          FALSE         GSM36437     1.036693e+12      <data.table[2x4]>
#> 3:          FALSE         GSM36448     1.035573e+12      <data.table[2x4]>
#> 4:          FALSE         GSM36428     1.022102e+12      <data.table[2x4]>
#> 5:          FALSE         GSM36439     1.033757e+12      <data.table[2x4]>
#> 6:          FALSE         GSM36450     1.018980e+12      <data.table[2x4]>
#>    sample.FactorValues array.ShortName
#> 1:   <data.table[2x4]>           GPL96
#> 2:   <data.table[2x4]>           GPL96
#> 3:   <data.table[2x4]>           GPL96
#> 4:   <data.table[2x4]>           GPL96
#> 5:   <data.table[2x4]>           GPL96
#> 6:   <data.table[2x4]>           GPL96
#>                                                  array.Name array.Taxon
#> 1: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 2: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 3: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 4: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 5: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#> 6: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A          NA
#>    array.TaxonID array.Type array.Description array.Troubled
#> 1:            NA         NA                NA          FALSE
#> 2:            NA         NA                NA          FALSE
#> 3:            NA         NA                NA          FALSE
#> 4:            NA         NA                NA          FALSE
#> 5:            NA         NA                NA          FALSE
#> 6:            NA         NA                NA          FALSE