Datasets

Lists datasets filtered and organized by given parameters

getDatasetsInfo(
  datasets = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

datasets

Optional, defaults to empty. Limits the result to entities with given identifiers. A list of identifiers, separated by commas (e.g: GSE2871,GSE2869,GSE2868). Identifiers can either be the Dataset ID or its short name. Retrieval by ID is more efficient. Only datasets that user has access to will be available. Do not combine different identifiers in one query.

offset

Optional, defaults to 0. Skips the specified amount of objects when retrieving them from the database.

limit

Optional, defaults to 20. Limits the result to specified amount of objects.

sort

Optional, defaults to +id. Sets the ordering property and direction. Format is [+,-][property name]. E.g. -accession will translate to descending ordering by the 'Accession' property. Note that this does [not guarantee the order of the returned entities!].description-imp This is merely a signal to how the data should be pre-sorted before the limit and offset are applied. Nested properties are also supported (recursively). E.g: +curationDetails.lastTroubledEvent.date When using in scripts, remember to URL-encode the '+' plus character (see the compiled URL below).

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forgetGemmaMemoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

A data table with information about the queried dataset(s). Returns an empty list if no datasets matched. A successful response may contain 'Geeq' information, which aims to provide a unified metric to measure experiments by the quality of their data, and their suitability for use in Gemma. You can read more about the geeq properties here

Examples

getDatasetsInfo("GSE2018")
#>    ee.ShortName                     ee.Name ee.ID
#> 1:      GSE2018 Human Lung Transplant - BAL     1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         ee.Description
#> 1: Bronchoalveolar lavage samples collected from lung transplant recipients.  Numeric portion of sample name is an arbitrary patient ID and AxBx number indicates the perivascular (A) and bronchiolar (B) scores from biopsies collected on the same day as the BAL fluid was collected.  Several patients have more than one sample in this series and can be determined by patient number followed by a lower case letter.  Acute rejection state is determined by the combined A and B score - specifically, a combined AB score of 2 or greater is considered an acute rejection.
#>    ee.Public ee.Troubled ee.Accession ee.Database
#> 1:      TRUE       FALSE      GSE2018         GEO
#>                              ee.URL ee.Samples
#> 1: http://www.ncbi.nlm.nih.gov/geo/         34
#>                                                 ee.batchEffect
#> 1: This data set may have a batch artifact (PC 2), p=0.0079605
#>    geeq.batchCorrected geeq.qScore geeq.sScore taxon.Name taxon.ID
#> 1:               FALSE   0.5657485       0.875      human        1
#>    technology.Type
#> 1:              NA
getDatasetsInfo(c("GSE2018", "GSE2872"))
#>    ee.ShortName                     ee.Name ee.ID
#> 1:      GSE2018 Human Lung Transplant - BAL     1
#> 2:      GSE2872       d'mel-affy-rat-168311     2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              ee.Description
#> 1:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Bronchoalveolar lavage samples collected from lung transplant recipients.  Numeric portion of sample name is an arbitrary patient ID and AxBx number indicates the perivascular (A) and bronchiolar (B) scores from biopsies collected on the same day as the BAL fluid was collected.  Several patients have more than one sample in this series and can be determined by patient number followed by a lower case letter.  Acute rejection state is determined by the combined A and B score - specifically, a combined AB score of 2 or greater is considered an acute rejection.
#> 2: Neurological diseases disrupt the quality of the lives of patients and often lead to their death prematurely.  A common feature of most neurological diseases is the degeneration of neurons, which results from an inappropriate activation of apoptosis. Drugs that inhibit neuronal apoptosis could thus be candidates for therapeutic intervention in neurodegenerative disorders.  We have identified (and recently reported) a chemical called GW5074 that inhibits apoptosis in a variety of cell culture paradigms of neuronal apoptosis.  Additionally, we have found that GW5074 reduces neurodegeneration and improves behavioral outcome in a mouse model of Huntington's disease.  Although GW5074 is a c-Raf inhibitor, we know very little about the molecular mechanisms underlying its neuroprotective action.  Identifying genes that are regulated by GW5074 in neurons will shed insight into this issue. We believe that neuroprotection by GW5074 involves the regulation of several genes.  Some of these genes are likely to be induced whereas the expression of other genes might be inhibited.  The specific aim is to identify genes that are differentially expressed in neurons treated with GW5074. We believe that neuroprotection by GW5074 involves the regulation of several genes.  Some of these genes are likely to be induced whereas the expression of other genes might be inhibited. Cultures of cerebellar granule neurons undergo apoptosis when switched from medium containing levated levels of potassium (high K+ or HK) to medium containing low potassium (LK).  Although cell death begins at about 18 h, we have found that the by 6 h after LK treatment these neurons are irreversibly committed to cell death.  We will treat cerebellar granule neuron cultures with LK medium (which induces them to undergo apoptosis) or with GW5074 (1 uM).  We will extract RNA at two time-points after treatment: 3 h and 6 h. Analysis at the two different time-points will show us whether changes in expression of specific genes is transient or sustained or whether the changes occurs early or relatively late in the process.  Neuronal cultures will be prepared from 1 week old Wistar rats.  The cultures will be maintained in culture for 1 week before treatment.  Following treatment the cells will be lysed and total RNA isolated.  The RNA will be stored at -80oC and shipped to the microarray facility for analysis.  The experiment will be done in triplicate.  Thus, for each time-point (3 or 6h treatment) we will be hope to provide 3 sets of samples (each set coming from a different culture preparation and containing lysates from cells treated with LK or GW5074).  Having samples from 3 independent cultures will mitigate any expression differences resulting from subtle variations in culture quality or in the preparation or quality of RNA.\nDate GSE2872 Last Updated: Jul 06 2005\nContributors:  S R D'Mello\nIncludes GDS1837.\n Update date: Jun 09 2006.\n Dataset description GDS1837: Analysis of cultured cerebellar granule neurons in low potassium medium 3 and 6 hours following treatment with the c-Raf inhibitor GW5074. GW5074 blocks low potassium induced cell death. Results provide insight into the neuroprotective action of GW5074.
#>    ee.Public ee.Troubled ee.Accession ee.Database
#> 1:      TRUE       FALSE      GSE2018         GEO
#> 2:      TRUE       FALSE      GSE2872         GEO
#>                              ee.URL ee.Samples
#> 1: http://www.ncbi.nlm.nih.gov/geo/         34
#> 2: http://www.ncbi.nlm.nih.gov/geo/         12
#>                                                 ee.batchEffect
#> 1: This data set may have a batch artifact (PC 2), p=0.0079605
#> 2:                                        SINGLE_BATCH_SUCCESS
#>    geeq.batchCorrected geeq.qScore geeq.sScore taxon.Name taxon.ID
#> 1:               FALSE   0.5657485       0.875      human        1
#> 2:               FALSE   0.8552303       0.500        rat        3
#>    technology.Type
#> 1:              NA
#> 2:              NA