Genes at location

Finds genes overlapping a given region.

getGenesAtLocation(
  taxon = NA_character_,
  chromosome = NA_character_,
  strand = "+",
  start = NA_integer_,
  size = NA_integer_,
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

taxon

Not required, part of the URL path. can either be Taxon ID, Taxon NCBI ID, or one of its string identifiers: scientific name, common name It is recommended to use Taxon ID for efficiency. Please note, that not all taxa have all the possible identifiers available. Use the 'All Taxa' endpoint to retrieve the necessary information. For convenience, below is a list of officially supported taxa: ID Comm.name Scient.name NcbiID

1 human Homo sapiens 9606 2 mouse Mus musculus 10090 3 rat Rattus norvegicus 10116 11 yeast Saccharomyces cerevisiae 4932 12 zebrafish Danio rerio 7955 13 fly Drosophila melanogaster 7227 14 worm Caenorhabditis elegans 6239

chromosome

Required, defaults to empty. The chromosome of the query location. Eg: 3, 21, X

strand

Optional, defaults to +. Can either be + or -. This is a WIP parameter and does currently not do anything When using in scripts, remember to URL-encode the '+' plus character (see the compiled URL below).

start

Required, defaults to empty. Number of the start nucleotide of the desired region.

size

Required, defaults to empty. Amount of nucleotides in the desired region (i.e. the length of the region).

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async

TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised

Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

An array of genes (gene value objects) that overlap the given region. A 404 error if the given identifier does not map to any object. A 400 error if required parameters are missing.

Examples

getGenesAtLocation("human", chromosome = 21, strand = "+", start = 2, size = 20000)
#>     gene.Symbol    gene.Ensembl gene.NCBI
#> 1:    LOC728312            <NA>    728312
#> 2:        PRED4            <NA>    400858
#> 3:    LOC642398            <NA>    642398
#> 4: LOC100129305            <NA> 100129305
#> 5: LOC100134011            <NA> 100134011
#> 6: LOC100134305            <NA> 100134305
#> 7:      COL18A1 ENSG00000182871     80781
#> 8: LOC105379618            <NA> 105379618
#>                                      gene.Name gene.Aliases
#> 1:                         similar to tektin 4           NA
#> 2:                            similar to PRED4           NA
#> 3:                      hypothetical LOC642398           NA
#> 4: similar to Uncharacterized protein C21orf94           NA
#> 5:          similar to PR domain containing 15           NA
#> 6:          similar to PR domain containing 15           NA
#> 7:           collagen type XVIII alpha 1 chain           NA
#> 8:                uncharacterized LOC105379618           NA
#>                                         gene.Description gene.GO
#> 1: Imported from NCBI gene; Nomenclature status: UNKNOWN       0
#> 2: Imported from NCBI gene; Nomenclature status: UNKNOWN       0
#> 3: Imported from NCBI gene; Nomenclature status: UNKNOWN       0
#> 4: Imported from NCBI gene; Nomenclature status: UNKNOWN       0
#> 5: Imported from NCBI gene; Nomenclature status: UNKNOWN       0
#> 6: Imported from NCBI gene; Nomenclature status: UNKNOWN       0
#> 7: Imported from NCBI gene; Nomenclature status: INTERIM       0
#> 8: Imported from NCBI gene; Nomenclature status: UNKNOWN       0
#>    gene.Homologues gene.MFX.Rank taxon.Name taxon.ID taxon.Scientific
#> 1:              NA             0      human        1     Homo sapiens
#> 2:              NA             0      human        1     Homo sapiens
#> 3:              NA             0      human        1     Homo sapiens
#> 4:              NA             0      human        1     Homo sapiens
#> 5:              NA             0      human        1     Homo sapiens
#> 6:              NA             0      human        1     Homo sapiens
#> 7:              NA             0      human        1     Homo sapiens
#> 8:              NA             0      human        1     Homo sapiens
#>    phenotypes
#> 1:         NA
#> 2:         NA
#> 3:         NA
#> 4:         NA
#> 5:         NA
#> 6:         NA
#> 7:         NA
#> 8:         NA