Gemma platform annotations

Gets Gemma's platform annotation files that can be accessed from https://gemma.msl.ubc.ca/annots/

getPlatformAnnotation(
  platform,
  annotType = c("bioProcess", "noParents", "allParents"),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  unzip = FALSE
)

Arguments

platform

A platform identifier @seealso getPlatforms

annotType

Which GO terms should the output include

file

Where to save the annotation file to, or empty to just load into memory

overwrite

Whether or not to overwrite an existing file

memoised

Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

unzip

Whether or not to unzip the file (if @param file is not empty)

Value

A table of annotations

Examples

dat <- getPlatformAnnotation("GPL96")
str(dat)
#> Classes ‘data.table’ and 'data.frame':	22283 obs. of  6 variables:
#>  $ ProbeName  : chr  "211750_x_at" "216678_at" "216345_at" "207273_at" ...
#>  $ GeneSymbols: chr  "TUBA1A|TUBA1C" "" "ZSWIM8" "" ...
#>  $ GeneNames  : chr  "tubulin alpha 1a|tubulin alpha 1c" "" "zinc finger SWIM-type containing 8" "" ...
#>  $ GOTerms    : chr  "GO:0000226|GO:0007017|GO:0050807|GO:0051301|GO:0000278|GO:0030705" "" "GO:0043161|GO:0016567|GO:2000627" "" ...
#>  $ GemmaIDs   : chr  "172797|360802" "" "235733" "" ...
#>  $ NCBIids    : chr  "7846|84790" "" "23053" "" ...
#>  - attr(*, ".internal.selfref")=<externalptr>