Platform element genes

Retrieves the genes on the given platform element

getPlatformElementGenes(
  platform = NA_character_,
  element = NA_character_,
  offset = 0L,
  limit = 20L,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

platform

Required, part of the URL path. Can either be the platform ID or its short name (e.g: GPL1355) Retrieval by ID is more efficient. Only platforms that user has access to will be available.

element

Optional, defaults to empty. Limits the result to entities with given identifiers. A list of identifiers, separated by commas (e.g: AFFX_Rat_beta-actin_M_at, AFFX_Rat_Hexokinase_M_at). Can either be probes name or IDs. Do not combine different identifiers in one query. When using in scripts, remember to URL-encode any forward slashes in the probe name (see the compiled URL below).

offset

Optional, defaults to 0. Skips the specified amount of objects when retrieving them from the database.

limit

Optional, defaults to 20. Limits the result to specified amount of objects.

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

An array of genes (gene value objects) aligned with the the given platform element. All identifiers must be valid. A 404 error if the given identifier does not map to any object.

Examples

getPlatformElementGenes("GPL1355", "AFFX_Rat_beta-actin_M_at")
#>    gene.Symbol       gene.Ensembl gene.NCBI   gene.Name gene.Aliases
#> 1:        Actb ENSRNOG00000034254     81822 actin, beta           NA
#>                                         gene.Description gene.GO
#> 1: Imported from NCBI gene; Nomenclature status: INTERIM       0
#>    gene.Homologues gene.MFX.Rank taxon.Name taxon.ID  taxon.Scientific
#> 1:              NA             0        rat        3 Rattus norvegicus
#>    phenotypes
#> 1:         NA