getResultSets

Lists resultSets filtered and organized by given parameters.

getResultSets(
  resultSet = NA_character_,
  dataset = NA_character_,
  filter = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

resultSet
Optional, defaults to empty. A single resultSet identifier (ex. 423176)

dataset
Required, part of the URL path. Can either be the dataset ID or its short name (e.g. GSE1234). Retrieval by ID is more efficient. Only datasets that user has access to will be available

filter
Optional, defaults to empty. Filtering can be done on any* property or nested property that the appropriate object class defines or inherits (and that is mapped by hibernate). [These do not correspond to the properties of the objects returned by the API calls.].description-imp Class definitions:

Datasets: javaDoc gitHub

Platforms: javaDoc gitHub E.g: curationDetails or curationDetails.lastTroubledEvent.date.

Any property of a supported type. Currently supported types are:

String - property of String type, required value can be any String.

Number - any Number implementation. Required value must be a string parseable to the specific Number type.

Boolean - required value will be parsed to true only if the string matches 'true', ignoring case. Accepted operator keywords are:

'=' - equality

'!=' - non-equality

'<' - smaller than

'>' - larger than

'<=' - smaller or equal

'=>' - larger or equal

'like' - similar string, effectively means 'contains', translates to the sql 'LIKE' operator (given value will be surrounded by % signs) Multiple filters can be chained using AND and OR keywords. Leave space between the keywords and the previous/next word! E.g: ?filter=property1 < value1 AND property2 like value2 If chained filters are mixed conjunctions and disjunctions, the query must be in conjunctive normal form (CNF). Parentheses are not necessary

every AND keyword separates blocks of disjunctions. Example: ?filter=p1 = v1 OR p1 != v2 AND p2 <= v2 AND p3 > v3 OR p3 < v4 Above query will translate to: (p1 = v1 OR p1 != v2) AND (p2 <= v2) AND (p3 > v3 OR p3 < v4;) Breaking the CNF results in an error. Filter curationDetails.troubled will be ignored if user is not an administrator.

offset
Optional, defaults to 0. Skips the specified amount of objects when retrieving them from the database.

limit
Optional, defaults to 20. Limits the result to specified amount of objects. Use 0 for no limit.

sort
Optional, defaults to +id. Sets the ordering property and direction. Format is [+,-][property name]. E.g. -accession will translate to descending ordering by the 'Accession' property. Note that this does [not guarantee the order of the returned entities!].description-imp This is merely a signal to how the data should be pre-sorted before the limit and offset are applied. Nested properties are also supported (recursively). E.g: +curationDetails.lastTroubledEvent.date When using in scripts, remember to URL-encode the '+' plus character (see the compiled URL below).

raw
TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async
TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised
Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file
The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite
Whether or not to overwrite if a file exists at the specified filename.

Value

Varies

Examples

getResultSets(500184)
#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.
#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.
#>                   id probe_id  probe_name      gene_id               gene_name
#>     1: 1.231217e-314    17652   205105_at        88540                  MAN2A1
#>     2: 1.231217e-314     5356   217563_at                                     
#>     3: 1.231217e-314    15939   206819_at                                     
#>     4: 1.231217e-314     3474   219446_at       295003                   RIC8B
#>     5: 1.231217e-314    16128   206630_at       165557                     TYR
#>    ---                                                                        
#> 22267: 1.231224e-314     9213   213691_at      8797810               PTOV1-AS2
#> 22268: 1.231224e-314    11392 211485_s_at       185530                   FGF18
#> 22269: 1.231224e-314     1984 220936_s_at       303350                    H2AJ
#> 22270: 1.231224e-314    17744 205013_s_at 3694|8802617 ADORA2A|SPECC1L-ADORA2A
#> 22271: 1.231224e-314     2528   220392_at       327286                    EBF2
#>         gene_ncbi_id    gene_official_symbol   pvalue corrected_pvalue    rank
#>     1:          4124                  MAN2A1 0.691200           0.8827 0.78310
#>     2:                                       0.023940           0.2607 0.09178
#>     3:                                       0.029350           0.2801 0.10470
#>     4:         55188                   RIC8B 0.680000           0.8786 0.77400
#>     5:          7299                     TYR 0.416300           0.7360 0.56560
#>    ---                                                                        
#> 22267:     101928378               PTOV1-AS2 0.160400           0.5186 0.30930
#> 22268:          8817                   FGF18 0.154700           0.5131 0.30150
#> 22269:         55766                    H2AJ 0.375100           0.7092 0.52890
#> 22270: 135|101730217 ADORA2A|SPECC1L-ADORA2A 0.373800           0.7085 0.52760
#> 22271:         64641                    EBF2 0.001836           0.1074 0.01702
#>        contrast_1_log2fc contrast_1_tstat contrast_1_pvalue
#>     1:           0.07823           0.4009          0.691200
#>     2:           0.17390           2.3750          0.023940
#>     3:           0.14450           2.2840          0.029350
#>     4:          -0.05527          -0.4163          0.680000
#>     5:          -0.06343          -0.8240          0.416300
#>    ---                                                     
#> 22267:          -0.14890          -1.4380          0.160400
#> 22268:          -0.09667          -1.4590          0.154700
#> 22269:           0.05851           0.8999          0.375100
#> 22270:           0.20720           0.9023          0.373800
#> 22271:          -0.16590          -3.4070          0.001836