Retrieve the design of a dataset

Retrieve the design of a dataset

get_dataset_design(
  dataset,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)

Arguments

dataset

A numerical dataset identifier or a dataset short name

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

attributes

If TRUE additional information from the call will be added into the output object's attributes such as offset and available elements.

Value

A data table of the design matrix for the queried dataset. A 404 error if the given identifier does not map to any object

Examples

head(get_dataset_design("GSE2018"))
#>                                                     diseaseState
#> BAL_47c_A0B1                              reference_subject_role
#> BAL_47b_A0B1                              reference_subject_role
#> BAL_47a_A0B0                              reference_subject_role
#> BAL_46_A2B0  Transplanted_Organ_Complication_|_Acute_rejection_|
#> BAL_45b_A2B0 Transplanted_Organ_Complication_|_Acute_rejection_|
#> BAL_45a_A1B0                              reference_subject_role
#>                          batch
#> BAL_47c_A0B1  Batch_06_7/24/02
#> BAL_47b_A0B1  Batch_07_10/2/02
#> BAL_47a_A0B0 Batch_09_10/15/02
#> BAL_46_A2B0   Batch_01_1/31/02
#> BAL_45b_A2B0  Batch_03_5/22/02
#> BAL_45a_A1B0  Batch_06_7/24/02