Compile gene expression data and metadata

Return an annotated Bioconductor-compatible data structure or a long form tibble of the queried dataset, including expression data and the experimental design.

get_dataset_object(
  dataset,
  filter = FALSE,
  type = "se",
  memoised = getOption("gemma.memoised", FALSE)
)

Arguments

dataset

A dataset identifier.

filter

The filtered version corresponds to what is used in most Gemma analyses, removing some probes/elements. Unfiltered includes all elements.

type

"se"for a SummarizedExperiment or "eset" for Expression Set. We recommend using SummarizedExperiments which are more recent. See the Summarized experiment vignette or the ExpressionSet vignette for more details.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

Value

A SummarizedExperiment, ExpressionSet or tibble containing metadata and expression data for the queried dataset.

Examples

get_dataset_object("GSE2018")
#> class: SummarizedExperiment 
#> dim: 22283 34 
#> metadata(8): title abstract ... GemmaSuitabilityScore taxon
#> assays(1): counts
#> rownames(22283): 1007_s_at 1053_at ... AFFX-r2-P1-cre-3_at
#>   AFFX-r2-P1-cre-5_at
#> rowData names(4): Probe GeneSymbol GeneName NCBIid
#> colnames(34): BAL_47c_A0B1 BAL_47b_A0B1 ... BAL_15a_A0B1 BAL_37_A1B1
#> colData names(2): diseaseState batch