Retrieve the platforms of a dataset

Retrieve the platforms of a dataset

get_dataset_platforms(
  dataset,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset

A numerical dataset identifier or a dataset short name

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

A data table with information about the platform(s). A list if raw = TRUE. A 404 error if the given identifier does not map to any object

The fields of the output data.table are:

platform.ID: Internal identifier of the platform

platform.ShortName: Shortname of the platform.

platform.Name: Full name of the platform.

platform.Description: Free text description of the platform

platform.Troubled: Whether or not the platform was marked "troubled" by a Gemma process or a curator

platform.ExperimentCount: Number of experiments using the platform within Gemma

platform.Type: Technology type for the platform.

taxon.Name: Name of the species platform was made for

taxon.Scientific: Scientific name for the taxon

taxon.ID: Internal identifier given to the species by Gemma

taxon.NCBI: NCBI ID of the taxon

taxon.Database.Name: Underlying database used in Gemma for the taxon

taxon.Database.ID: ID of the underyling database used in Gemma for the taxon

Examples

get_dataset_platforms("GSE2018")
#>    platform.ID platform.ShortName
#> 1:           1              GPL96
#>                                               platform.Name
#> 1: Affymetrix GeneChip Human Genome U133 Array Set HG-U133A
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          platform.Description
#> 1:  The U133 set includes 2 arrays with a total of 44928 entries and was indexed 29-Jan-2002. The set includes over 1,000,000 unique oligonucleotide features covering more than 39,000 transcript variants, which in turn represent greater than 33,000 of the best characterized human genes. Sequences were selected from GenBank, dbEST, and RefSeq. Sequence clusters were created from Build 133 of UniGene (April 20, 2001) and refined by analysis and comparison with a number of other publicly available databases including the Washington University EST trace repository and the University of California, Santa Cruz golden-path human genome database (April 2001 release). In addition, ESTs were analyzed for untrimmed low-quality sequence information, correct orientation, false priming, false clustering, alternative splicing and alternative polyadenylation.  Keywords = high density oligonucleotide array\nFrom GPL96\nLast Updated: Mar 09 2006
#>    platform.Troubled platform.ExperimentCount platform.Type taxon.Name
#> 1:             FALSE                      404      ONECOLOR      human
#>    taxon.Scientific taxon.ID taxon.NCBI taxon.Database.Name taxon.Database.ID
#> 1:     Homo sapiens        1       9606                hg38                87