Retrieve processed expression data of a dataset

Retrieve processed expression data of a dataset

get_dataset_processed_expression(
  dataset,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset

A numerical dataset identifier or a dataset short name

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

If raw is FALSE (default), a data table of the expression matrix for the queried dataset. If raw is TRUE, returns the binary file in raw form.

Examples

get_dataset_processed_expression("GSE2018")
#>                         Probe          GeneSymbol
#>     1:              1007_s_at                DDR1
#>     2:                1053_at                RFC2
#>     3:                 117_at         HSPA7|HSPA6
#>     4:                 121_at                PAX8
#>     5:              1255_g_at LOC118142757|GUCA1A
#>    ---                                           
#> 22279: AFFX-r2-Hs28SrRNA-3_at                    
#> 22280: AFFX-r2-Hs28SrRNA-5_at                    
#> 22281: AFFX-r2-Hs28SrRNA-M_at                    
#> 22282:    AFFX-r2-P1-cre-3_at                    
#> 22283:    AFFX-r2-P1-cre-5_at                    
#>                                                                                                      GeneName
#>     1:                                                            discoidin domain receptor tyrosine kinase 1
#>     2:                                                                         replication factor C subunit 2
#>     3: heat shock protein family A (Hsp70) member 7 (pseudogene)|heat shock protein family A (Hsp70) member 6
#>     4:                                                                                           paired box 8
#>     5:                                             GUCA1ANB-GUCA1A readthrough|guanylate cyclase activator 1A
#>    ---                                                                                                       
#> 22279:                                                                                                       
#> 22280:                                                                                                       
#> 22281:                                                                                                       
#> 22282:                                                                                                       
#> 22283:                                                                                                       
#>                NCBIid BAL_21_A0B0 BAL_19_A0B0 BAL_12_A1B0 BAL_15b_A0B0
#>     1:            780    7.867851    7.821547    7.606950     7.675714
#>     2:           5982    6.615844    6.709687    7.081941     6.770301
#>     3:      3311|3310    7.421375    7.825478    7.497553     8.433980
#>     4:           7849    9.553362    9.574804    9.130426     9.062922
#>     5: 118142757|2978    4.137057    3.989322    4.130228     3.991846
#>    ---                                                                
#> 22279:                   7.957031    8.356067    6.242087     5.699829
#> 22280:                   5.204943    5.687296    4.349244     4.661939
#> 22281:                   7.226784    7.929119    5.663534     5.803358
#> 22282:                  12.589158   12.671734   12.376739    11.258021
#> 22283:                  12.909754   12.909754   12.476188    11.643198
#>        BAL_40_A0B0 BAL_13_A0B0 BAL_44b_A0B0 BAL_41_A1B0 BAL_31_A0B1 BAL_43_A0B0
#>     1:    7.568318    7.664396     7.555234    7.898749    7.482914    7.333711
#>     2:    6.754742    6.818718     6.817860    6.428031    6.873047    6.966837
#>     3:    7.307702    7.702311     7.736164    8.299871    7.714629    7.462743
#>     4:    9.126554    8.872670     8.728661    9.433338    9.106445    9.090870
#>     5:    4.057203    4.023544     4.022874    4.266776    4.127223    4.002775
#>    ---                                                                         
#> 22279:    5.119069    5.269825     4.883055    6.307157    6.406686    5.713751
#> 22280:    4.238217    4.032411     4.443605    4.487075    4.551892    4.326695
#> 22281:    5.372956    5.033840     5.321434    5.462065    5.793712    5.575244
#> 22282:   12.462382   12.634298    11.756798   11.946831   12.617424   12.728011
#> 22283:   12.634298   12.589158    11.830940   11.781727   12.318279   12.708045
#>        BAL_47a_A0B0 BAL_44a_A0B0 BAL_39_A0B1 BAL_33_A0B1 BAL_47b_A0B1
#>     1:     7.787959     7.734734    7.270490    7.428734     7.605180
#>     2:     6.440673     6.940845    6.808251    7.139178     6.785765
#>     3:     7.844755     7.603600    7.328971    7.568634     7.800827
#>     4:     9.040855     9.248298    8.881946    8.946895     9.024894
#>     5:     3.909490     4.004146    4.018348    4.025862     4.179423
#>    ---                                                               
#> 22279:     5.371702     5.183981    5.413069    5.689970     6.658047
#> 22280:     4.276494     4.627294    4.337369    4.311341     4.462518
#> 22281:     5.166088     6.134356    5.647816    5.397744     5.477958
#> 22282:    12.462382    12.257533   12.728011   11.887863    12.871473
#> 22283:    12.656571    12.342556   12.855262   11.810861    12.909754
#>        BAL_29_A1B0 BAL_25_A0B0 BAL_47c_A0B1 BAL_45a_A1B0 BAL_42_A0B0 BAL_4_A0B0
#>     1:    7.905805    7.565792     7.629114     7.446346    7.893288   7.474848
#>     2:    6.433760    6.887287     6.910238     6.803677    6.649465   6.947905
#>     3:    7.794066    7.444417     7.474227     7.623731    8.579609   7.741327
#>     4:    9.592714    8.923390     9.173378     9.293806    8.930581   9.119985
#>     5:    4.128827    3.979449     3.955497     4.034630    4.113324   4.037982
#>    ---                                                                         
#> 22279:    5.304832    4.986458     5.768701     5.107271    5.882000   5.745562
#> 22280:    4.163611    4.339306     4.355739     4.260343    4.201591   4.332832
#> 22281:    5.566520    5.449715     5.260586     5.040337    5.634897   5.390216
#> 22282:   12.742010   12.323225    12.033990    11.272546   12.758545  12.582693
#> 22283:   12.880875   12.294077    12.088908    11.459026   12.728011  12.742010
#>        BAL_3_A1B0 BAL_18_A1B0 BAL_17_A0B0 BAL_27a_A0B0 BAL_1_A0B0 BAL_15a_A0B1
#>     1:   7.543944    7.407762    7.229377     7.912942   7.520084     7.633544
#>     2:   6.895257    6.780074    6.963063     6.975750   7.158184     6.906474
#>     3:   7.542451    7.926313    7.972501     7.775969   7.843651     7.689818
#>     4:   8.873311    9.154426    8.806005     9.249823   9.057032     8.764082
#>     5:   3.861244    4.054710    4.024187     4.027976   4.168197     4.025454
#>    ---                                                                        
#> 22279:   6.067514    7.299089    6.076631     5.660599   5.754021     5.472515
#> 22280:   4.337072    4.395083    4.215431     4.354635   4.305040     4.360437
#> 22281:   5.559285    5.974870    5.704353     5.502226   5.673235     5.541409
#> 22282:  12.595888   12.841431   12.516955    12.363355  12.419389    12.462382
#> 22283:  12.750864   12.862965   12.449719    12.294077  12.401875    12.383058
#>        BAL_46_A2B0 BAL_27b_A2B1 BAL_26_A2B0 BAL_45b_A2B0 BAL_37_A1B1
#>     1:    7.932626     7.699664    7.257932     7.782975    7.466471
#>     2:    6.537636     6.845171    6.940845     6.593049    6.858354
#>     3:    7.809553     7.555798    7.969855     7.932626    7.689818
#>     4:   10.078527     9.141695    8.656306     8.999097    9.055730
#>     5:    3.859072     4.121438    4.013794     3.944738    4.062315
#>    ---                                                              
#> 22279:    8.848175     5.789951    4.948480     5.328197    5.377125
#> 22280:    4.782150     4.440793    4.171280     4.119304    4.253467
#> 22281:    7.281428     5.875893    5.167862     5.046239    5.369570
#> 22282:   13.047083    12.443970   12.088908    12.123510   11.975114
#> 22283:   13.239051    12.671734   12.211794    12.093898   12.088908
#>        BAL_20b_A1B1 BAL_20a_A1B1
#>     1:     7.680408          NaN
#>     2:     6.650542          NaN
#>     3:     7.738031          NaN
#>     4:     9.239082          NaN
#>     5:     4.030433          NaN
#>    ---                          
#> 22279:     6.952453          NaN
#> 22280:     4.250091          NaN
#> 22281:     5.876946          NaN
#> 22282:    12.624101          NaN
#> 22283:    12.449719          NaN