Retrieve raw expression data of a dataset

Retrieve raw expression data of a dataset

get_dataset_raw_expression(
  dataset,
  quantitationType,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset

A numerical dataset identifier or a dataset short name

quantitationType

Quantitation type id. These can be acquired using get_dataset_quantitation_types function. This endpoint can only return non-processed quantitation types.

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

If raw is FALSE (default), a data table of the expression matrix for the queried dataset. If raw is TRUE, returns the binary file in raw form.

Examples

q_types <- get_dataset_quantitation_types("GSE59918")
get_dataset_raw_expression("GSE59918", q_types$id[q_types$name == "Counts"])
#>            Probe GeneSymbol
#>            <int>     <char>
#>     1: 100034253      Gnl3l
#>     2: 100036765     Ccdc92
#>     3: 100049583      Trex1
#>     4: 100101439           
#>     5: 100124593      Cxcr6
#>    ---                     
#> 31212:     94338      Smpd3
#> 31213:     94339      Mmp23
#> 31214:     94340      Acsl5
#> 31215:     94341     Kcnj13
#> 31216:     94342       Bag6
#>                                                             GeneName    NCBIid
#>                                                               <char>     <int>
#>     1:                                    G protein nucleolar 3 like 100034253
#>     2:                              coiled-coil domain containing 92 100036765
#>     3:                              three prime repair exonuclease 1 100049583
#>     4:                                                                      NA
#>     5:                              C-X-C motif chemokine receptor 6 100124593
#>    ---                                                                        
#> 31212:                             sphingomyelin phosphodiesterase 3     94338
#> 31213:                                    matrix metallopeptidase 23     94339
#> 31214:                acyl-CoA synthetase long-chain family member 5     94340
#> 31215: potassium inwardly-rectifying channel, subfamily J, member 13     94341
#> 31216:                                             BAG cochaperone 6     94342
#>        rat hypothalamus Fru_rep1 rat hypothalamus Ctrl_rep4
#>                            <num>                      <num>
#>     1:                    1191.0                      729.0
#>     2:                    4681.6                     4629.4
#>     3:                     420.0                      373.0
#>     4:                       0.0                        0.0
#>     5:                       4.0                        6.0
#>    ---                                                     
#> 31212:                    1004.0                      703.0
#> 31213:                     133.0                      130.0
#> 31214:                     581.0                      390.0
#> 31215:                      85.0                      282.0
#> 31216:                    5406.0                     4690.0
#>        rat hypothalamus Fru_rep2 rat hypothalamus Fru_rep3
#>                            <num>                     <num>
#>     1:                    1224.0                    1187.0
#>     2:                    4380.2                    4158.3
#>     3:                     356.0                     416.0
#>     4:                       0.0                       0.0
#>     5:                       3.0                       2.0
#>    ---                                                    
#> 31212:                     926.0                     913.0
#> 31213:                      79.0                     141.0
#> 31214:                     524.0                     459.0
#> 31215:                       6.0                     142.0
#> 31216:                    5364.0                    5576.0
#>        rat hypothalamus Ctrl_rep1 rat hippocampus Fru_rep3
#>                             <num>                    <num>
#>     1:                     1059.0                   589.00
#>     2:                     4147.9                   782.76
#>     3:                      422.0                    60.00
#>     4:                        0.0                     0.00
#>     5:                        4.0                     1.00
#>    ---                                                    
#> 31212:                      880.0                  1442.00
#> 31213:                      135.0                    35.00
#> 31214:                      518.0                   828.00
#> 31215:                      237.0                   281.00
#> 31216:                     5472.0                  1594.00
#>        rat hippocampus Fru_rep1 rat hypothalamus Fru_rep4
#>                           <num>                     <num>
#>     1:                   834.00                     263.0
#>     2:                   868.61                    3229.1
#>     3:                    66.00                     449.0
#>     4:                     0.00                       0.0
#>     5:                     1.00                       0.0
#>    ---                                                   
#> 31212:                  1699.00                     406.0
#> 31213:                    17.00                      85.0
#> 31214:                  1215.00                     310.0
#> 31215:                  2100.00                      73.0
#> 31216:                  2016.00                    3896.0
#>        rat hypothalamus Ctrl_rep3 rat hippocampus Ctrl_rep3
#>                             <num>                     <num>
#>     1:                      932.0                    1126.0
#>     2:                     3527.7                    1273.7
#>     3:                      345.0                      72.0
#>     4:                        1.0                       0.0
#>     5:                        7.0                       0.0
#>    ---                                                     
#> 31212:                      883.0                    2456.0
#> 31213:                      146.0                      25.0
#> 31214:                      594.0                    1412.0
#> 31215:                      246.0                     269.0
#> 31216:                     4698.0                    2807.0
#>        rat hippocampus Ctrl_rep2 rat hippocampus Ctrl_rep4
#>                            <num>                     <num>
#>     1:                    770.00                    648.00
#>     2:                    871.83                    768.66
#>     3:                     55.00                     64.00
#>     4:                      0.00                      0.00
#>     5:                      0.00                      2.00
#>    ---                                                    
#> 31212:                   1620.00                   1541.00
#> 31213:                     16.00                     51.00
#> 31214:                   1177.00                   1431.00
#> 31215:                    641.00                   1199.00
#> 31216:                   1626.00                   1978.00
#>        rat hippocampus Ctrl_rep1 rat hippocampus Fru_rep4
#>                            <num>                    <num>
#>     1:                    539.00                  1090.00
#>     2:                    774.63                   684.74
#>     3:                     59.00                    57.00
#>     4:                      0.00                     0.00
#>     5:                      2.00                     1.00
#>    ---                                                   
#> 31212:                   1176.00                  2361.00
#> 31213:                     33.00                    21.00
#> 31214:                    845.00                  1863.00
#> 31215:                   1165.00                  2345.00
#> 31216:                   1661.00                  2064.00
#>        rat hippocampus Fru_rep2 rat hypothalamus Ctrl_rep2
#>                           <num>                      <num>
#>     1:                   769.00                      953.0
#>     2:                   862.79                     4084.7
#>     3:                    96.00                      388.0
#>     4:                     0.00                        0.0
#>     5:                     9.00                        0.0
#>    ---                                                    
#> 31212:                  1736.00                      809.0
#> 31213:                    83.00                      127.0
#> 31214:                  1254.00                      484.0
#> 31215:                   711.00                      120.0
#> 31216:                  1881.00                     4465.0