Retrieve datasets by their identifiers

Retrieve datasets by their identifiers

get_datasets_by_ids(
  datasets = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)

Arguments

datasets

Numerical dataset identifiers or dataset short names

offset

The offset of the first retrieved result.

limit

Optional, defaults to 20. Limits the result to specified amount of objects. Has a maximum value of 100. Use together with offset and the totalElements attribute in the output to compile all data if needed.

sort

Order results by the given property and direction. The '+' sign indicate ascending order whereas the '-' indicate descending.

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

attributes

If TRUE additional information from the call will be added into the output object's attributes such as offset and available elements.

Value

A data table with information about the queried dataset(s). A list if raw = TRUE. Returns an empty list if no datasets matched. A successful response may contain 'Geeq' information, which aims to provide a unified metric to measure experiments by the quality of their data, and their suitability for use in Gemma. You can read more about the geeq properties here.

The fields of the output data.table are:

experiment.ShortName: Shortname given to the dataset within Gemma. Often corresponds to accession ID

experiment.Name: Full title of the dataset

experiment.ID: Internal ID of the dataset.

experiment.Description: Description of the dataset

experiment.Public: Is the dataset publicly available. Only useful for logged in users with access to non-public data

experiment.Troubled: Did an automatic process within gemma or a curator mark the dataset as "troubled"

experiment.Accession: Accession ID of the dataset in the external database it was taken from

experiment.Database: The name of the database where the dataset was taken from

experiment.URI: URI of the original database

experiment.SampleCount: Number of samples in the dataset

experiment.batchEffect: A text field describing whether the dataset has batch effects

experiment.batchCorrected: Whether batch correction has been performed on the dataset. This is a text field with possible details about the batch correction. Use geeq.batchEffect if a more structured field is needed.

geeq.batchConfound: 0 if batch info isn't available, -1 if batch counfoud is detected, 1 if batch information is available and no batch confound found

geeq.batchEffect: -1 if batch p value < 0.0001, 1 if batch p value > 0.1, 0 if otherwise and when there is no batch information is available or when the data is confounded with batches.

geeq.rawData: -1 if no raw data available, 1 if raw data was available. When available, Gemma reprocesses raw data to get expression values and batches

geeq.qScore: Data quality score given to the dataset by Gemma.

geeq.sScore: Suitability score given to the dataset by Gemma. Refers to factors like batches, platforms and other aspects of experimental design

taxon.Name: The taxa of the study. In Gemma each study will include a single species. If the original source has samples from multiple species, they will be split into different studies within Gemma

taxon.ID: Internal ID given to the taxon by Gemma

Examples

get_datasets_by_ids("GSE2018")
#>    experiment.ShortName             experiment.Name experiment.ID
#> 1:              GSE2018 Human Lung Transplant - BAL             1
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 experiment.Description
#> 1: Bronchoalveolar lavage samples collected from lung transplant recipients.  Numeric portion of sample name is an arbitrary patient ID and AxBx number indicates the perivascular (A) and bronchiolar (B) scores from biopsies collected on the same day as the BAL fluid was collected.  Several patients have more than one sample in this series and can be determined by patient number followed by a lower case letter.  Acute rejection state is determined by the combined A and B score - specifically, a combined AB score of 2 or greater is considered an acute rejection.
#>    experiment.Public experiment.Troubled experiment.Accession
#> 1:              TRUE               FALSE              GSE2018
#>    experiment.Database                   experiment.URI experiment.SampleCount
#> 1:                 GEO http://www.ncbi.nlm.nih.gov/geo/                     34
#>    experiment.LastUpdated
#> 1:    2022-08-19 16:13:48
#>                                         experiment.batchEffect
#> 1: This data set may have a batch artifact (PC 2), p=0.0079605
#>    geeq.batchCorrected geeq.batchConfound geeq.batchEffect geeq.rawData
#> 1:               FALSE                  1                0            1
#>    geeq.qScore geeq.sScore taxon.Name taxon.ID
#> 1:   0.5657485       0.875      human        1
get_datasets_by_ids(c("GSE2018", "GSE2872"))
#>    experiment.ShortName             experiment.Name experiment.ID
#> 1:              GSE2018 Human Lung Transplant - BAL             1
#> 2:              GSE2872       d'mel-affy-rat-168311             2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      experiment.Description
#> 1:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Bronchoalveolar lavage samples collected from lung transplant recipients.  Numeric portion of sample name is an arbitrary patient ID and AxBx number indicates the perivascular (A) and bronchiolar (B) scores from biopsies collected on the same day as the BAL fluid was collected.  Several patients have more than one sample in this series and can be determined by patient number followed by a lower case letter.  Acute rejection state is determined by the combined A and B score - specifically, a combined AB score of 2 or greater is considered an acute rejection.
#> 2: Neurological diseases disrupt the quality of the lives of patients and often lead to their death prematurely.  A common feature of most neurological diseases is the degeneration of neurons, which results from an inappropriate activation of apoptosis. Drugs that inhibit neuronal apoptosis could thus be candidates for therapeutic intervention in neurodegenerative disorders.  We have identified (and recently reported) a chemical called GW5074 that inhibits apoptosis in a variety of cell culture paradigms of neuronal apoptosis.  Additionally, we have found that GW5074 reduces neurodegeneration and improves behavioral outcome in a mouse model of Huntington's disease.  Although GW5074 is a c-Raf inhibitor, we know very little about the molecular mechanisms underlying its neuroprotective action.  Identifying genes that are regulated by GW5074 in neurons will shed insight into this issue. We believe that neuroprotection by GW5074 involves the regulation of several genes.  Some of these genes are likely to be induced whereas the expression of other genes might be inhibited.  The specific aim is to identify genes that are differentially expressed in neurons treated with GW5074. We believe that neuroprotection by GW5074 involves the regulation of several genes.  Some of these genes are likely to be induced whereas the expression of other genes might be inhibited. Cultures of cerebellar granule neurons undergo apoptosis when switched from medium containing levated levels of potassium (high K+ or HK) to medium containing low potassium (LK).  Although cell death begins at about 18 h, we have found that the by 6 h after LK treatment these neurons are irreversibly committed to cell death.  We will treat cerebellar granule neuron cultures with LK medium (which induces them to undergo apoptosis) or with GW5074 (1 uM).  We will extract RNA at two time-points after treatment: 3 h and 6 h. Analysis at the two different time-points will show us whether changes in expression of specific genes is transient or sustained or whether the changes occurs early or relatively late in the process.  Neuronal cultures will be prepared from 1 week old Wistar rats.  The cultures will be maintained in culture for 1 week before treatment.  Following treatment the cells will be lysed and total RNA isolated.  The RNA will be stored at -80oC and shipped to the microarray facility for analysis.  The experiment will be done in triplicate.  Thus, for each time-point (3 or 6h treatment) we will be hope to provide 3 sets of samples (each set coming from a different culture preparation and containing lysates from cells treated with LK or GW5074).  Having samples from 3 independent cultures will mitigate any expression differences resulting from subtle variations in culture quality or in the preparation or quality of RNA.\nDate GSE2872 Last Updated: Jul 06 2005\nContributors:  S R D'Mello\nIncludes GDS1837.\n Update date: Jun 09 2006.\n Dataset description GDS1837: Analysis of cultured cerebellar granule neurons in low potassium medium 3 and 6 hours following treatment with the c-Raf inhibitor GW5074. GW5074 blocks low potassium induced cell death. Results provide insight into the neuroprotective action of GW5074.
#>    experiment.Public experiment.Troubled experiment.Accession
#> 1:              TRUE               FALSE              GSE2018
#> 2:              TRUE               FALSE              GSE2872
#>    experiment.Database                   experiment.URI experiment.SampleCount
#> 1:                 GEO http://www.ncbi.nlm.nih.gov/geo/                     34
#> 2:                 GEO http://www.ncbi.nlm.nih.gov/geo/                     12
#>    experiment.LastUpdated
#> 1:    2022-08-19 16:13:48
#> 2:    2022-09-04 13:49:21
#>                                         experiment.batchEffect
#> 1: This data set may have a batch artifact (PC 2), p=0.0079605
#> 2:                                        SINGLE_BATCH_SUCCESS
#>    geeq.batchCorrected geeq.batchConfound geeq.batchEffect geeq.rawData
#> 1:               FALSE                  1                0            1
#> 2:               FALSE                  1                1            1
#>    geeq.qScore geeq.sScore taxon.Name taxon.ID
#> 1:   0.5657485       0.875      human        1
#> 2:   0.8552303       0.500        rat        3