Retrieve Platform Annotations by Gemma

Gets Gemma's platform annotations including mappings of microarray probes to genes.

get_platform_annotations(
  platform,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

platform

A platform numerical identifier or a platform short name

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

A table of annotations

ElementName: Probeset names provided by the platform. Gene symbols for generic annotations typically used for RNA-seq experiments.

GeneSymbols: Genes that were found to be aligned to the probe sequence. Note that it is possible for probes to be non-specific. Alignment to multiple genes are indicated with gene symbols separated by "|"s

GeneNames: Name of the gene

GOTerms: GO Terms associated with the genes.

GemmaIDs and NCBIids: respective IDs for the genes.

Examples

head(get_platform_annotations("GPL96"))
#>    ElementName GeneSymbols
#>         <char>      <char>
#> 1: 202619_s_at       PLOD2
#> 2: 219952_s_at      MCOLN1
#> 3:   205192_at     MAP3K14
#> 4:   212724_at        RND3
#> 5:   207373_at      HOXD10
#> 6: 204991_s_at         NF2
#>                                                   GeneNames
#>                                                      <char>
#> 1:        procollagen-lysine,2-oxoglutarate 5-dioxygenase 2
#> 2:                           mucolipin TRP cation channel 1
#> 3:        mitogen-activated protein kinase kinase kinase 14
#> 4:                                      Rho family GTPase 3
#> 5:                                             homeobox D10
#> 6: NF2, moesin-ezrin-radixin like (MERLIN) tumor suppressor
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            GOTerms
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             <char>
#> 1:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             GO:0050896|GO:0003824|GO:0031418|GO:0070062|GO:0110165|GO:0043169|GO:0043168|GO:0005488|GO:0008475|GO:0006493|GO:0046872|GO:0043167|GO:1901363|GO:0140096|GO:0046914|GO:0045229|GO:0031090|GO:0019842|GO:0030246|GO:0031012|GO:0005737|GO:0031406|GO:0003674|GO:0033823|GO:0043170|GO:0005575|GO:0048029|GO:0009059|GO:0016706|GO:0009058|GO:0043177|GO:0016705|GO:0019538|GO:0046527|GO:0036094|GO:0030312|GO:0036293|GO:0030199|GO:0016740|GO:0030198|GO:0016020|GO:0005506|GO:0043062|GO:0001666|GO:0030867|GO:0070482|GO:0005789|GO:0032964|GO:0006950|GO:0044238|GO:1901137|GO:0065010|GO:0043226|GO:0005783|GO:1901135|GO:0009101|GO:0009100|GO:0016757|GO:0016758|GO:1903561|GO:0043227|GO:0043229|GO:0009987|GO:0032963|GO:0009628|GO:0098588|GO:0051213|GO:0016491|GO:0005794|GO:0043231|GO:0043230|GO:0008152|GO:0008150|GO:0031982|GO:0071840|GO:0016043
#> 2:                                                                                                                                                                                                                  GO:0050896|GO:0006816|GO:0045184|GO:0006813|GO:0006814|GO:0006811|GO:0006812|GO:0104004|GO:0006810|GO:0042592|GO:0010638|GO:0019725|GO:0030003|GO:0019722|GO:0030001|GO:0015085|GO:0006826|GO:0005216|GO:0005217|GO:0003674|GO:0006820|GO:0000323|GO:0005215|GO:0050789|GO:0010008|GO:0005575|GO:0008324|GO:0001891|GO:0099094|GO:0000041|GO:0042802|GO:0051592|GO:0022857|GO:0071467|GO:0051234|GO:0098655|GO:0051128|GO:0098771|GO:0048878|GO:0006996|GO:0097682|GO:0005267|GO:0006873|GO:0043226|GO:0005261|GO:0007165|GO:0005262|GO:0043227|GO:0048518|GO:0043229|GO:0009987|GO:0009628|GO:0022607|GO:0050794|GO:0072345|GO:0098656|GO:0033572|GO:0051259|GO:0015081|GO:0090382|GO:0043235|GO:0006882|GO:1905671|GO:0043231|GO:1905673|GO:0005272|GO:0048522|GO:0065003|GO:0033043|GO:0065007|GO:0008150|GO:0071805|GO:0051130|GO:0022834|GO:0022836|GO:0015079|GO:0071248|GO:0032991|GO:0016043|GO:0015075|GO:0005768|GO:0005764|GO:0110165|GO:0005765|GO:0005886|GO:0005488|GO:0046873|GO:0046915|GO:0097708|GO:0031090|GO:0010038|GO:0051262|GO:0070588|GO:0051260|GO:0031410|GO:0099604|GO:0033036|GO:0071277|GO:0035725|GO:0005773|GO:0012506|GO:0005774|GO:0051716|GO:0005253|GO:0005770|GO:0005381|GO:0034755|GO:0043933|GO:0030670|GO:0016020|GO:0098852|GO:0030666|GO:1901660|GO:0055085|GO:0009268|GO:0055080|GO:0141124|GO:0035556|GO:0055082|GO:0070887|GO:0022803|GO:0015267|GO:0071214|GO:0051289|GO:0034220|GO:0098588|GO:0008509|GO:0005515|GO:0030659|GO:0031902|GO:0050801|GO:0008104|GO:0042221|GO:0044088|GO:0015276|GO:0044090|GO:0031982|GO:0015278|GO:0071840|GO:0051179|GO:0015031
#> 3:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             GO:0050896|GO:0003824|GO:0005524|GO:0002376|GO:0104004|GO:0110165|GO:0004672|GO:0051707|GO:0043168|GO:0005488|GO:0004674|GO:0043167|GO:1901363|GO:0140096|GO:0044419|GO:0043207|GO:0016772|GO:0016773|GO:0009605|GO:0009607|GO:0005737|GO:0106310|GO:0000166|GO:0000165|GO:0071260|GO:0003674|GO:0050789|GO:0005575|GO:0051716|GO:0009612|GO:0036094|GO:0030554|GO:0009615|GO:0016740|GO:0016301|GO:0017076|GO:0071496|GO:0097367|GO:0006955|GO:0006952|GO:0006950|GO:0035639|GO:0051607|GO:0007165|GO:0035556|GO:0141124|GO:0009987|GO:0009628|GO:0005829|GO:0071214|GO:0050794|GO:0005515|GO:0032559|GO:0038061|GO:1901265|GO:0065007|GO:0008150|GO:0004709|GO:0032555|GO:0032553
#> 4:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  GO:0005925|GO:0003824|GO:0005525|GO:0110165|GO:0005886|GO:0043168|GO:0005488|GO:0016817|GO:0016818|GO:0043167|GO:1901363|GO:0031090|GO:0019001|GO:0000166|GO:0003674|GO:0050789|GO:0005575|GO:0007155|GO:0016787|GO:0019899|GO:0030036|GO:0036094|GO:0017076|GO:0032970|GO:0097435|GO:0016020|GO:0017111|GO:0016462|GO:0051128|GO:0097367|GO:0000139|GO:0030029|GO:0098772|GO:0070161|GO:0006996|GO:0035639|GO:0007165|GO:0019901|GO:0009987|GO:0019900|GO:0003925|GO:0005829|GO:0003924|GO:0050794|GO:0032561|GO:0098588|GO:0005515|GO:0032956|GO:0007015|GO:1901265|GO:0007010|GO:0033043|GO:0065007|GO:0008150|GO:0051493|GO:0032555|GO:0032553|GO:0030054|GO:0071840|GO:0030055|GO:0016043
#> 5:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         GO:0003700|GO:0110165|GO:0043565|GO:0005488|GO:0045786|GO:0010557|GO:0009891|GO:0010556|GO:0045944|GO:0009889|GO:2001141|GO:0001216|GO:0003674|GO:0003677|GO:0003676|GO:0050789|GO:0005575|GO:0010604|GO:0060255|GO:0009893|GO:0019219|GO:1990837|GO:0080090|GO:0001228|GO:0001067|GO:0051726|GO:0003682|GO:0006357|GO:0043226|GO:0006355|GO:0140110|GO:0043227|GO:0048519|GO:0043229|GO:0048518|GO:0019222|GO:0000976|GO:0000977|GO:0050794|GO:0000978|GO:0005515|GO:0005634|GO:0000981|GO:0000785|GO:0043231|GO:0048523|GO:0048522|GO:0003690|GO:0045893|GO:1902680|GO:0065007|GO:0008150|GO:0010468|GO:0051252|GO:0000987|GO:0051254
#> 6: GO:0042995|GO:0060341|GO:0010638|GO:0008285|GO:0008283|GO:0070306|GO:0060627|GO:0032388|GO:0032386|GO:0005737|GO:0048585|GO:0005856|GO:0048583|GO:0003674|GO:0051239|GO:0048869|GO:0050789|GO:0005575|GO:0005730|GO:0010648|GO:0005178|GO:0010646|GO:0045595|GO:1902531|GO:0042127|GO:0051128|GO:0048471|GO:0070161|GO:0030863|GO:0006996|GO:0030864|GO:0110020|GO:0050839|GO:0043226|GO:0007165|GO:0043067|GO:0043228|GO:0043227|GO:0048519|GO:0048518|GO:0043229|GO:0009987|GO:0022603|GO:0035591|GO:0022604|GO:0005829|GO:0014013|GO:0050793|GO:0050794|GO:0032880|GO:0014010|GO:0003779|GO:0032879|GO:0030855|GO:0030336|GO:0043231|GO:0048523|GO:0048522|GO:0043232|GO:0033043|GO:1905666|GO:0065007|GO:0008150|GO:0065008|GO:1905668|GO:2000641|GO:2000643|GO:0051130|GO:0051493|GO:0030334|GO:0051492|GO:0030175|GO:0030054|GO:0032233|GO:0051495|GO:0032231|GO:0016043|GO:0051496|GO:0005768|GO:0005769|GO:0110165|GO:0005886|GO:0032502|GO:1902965|GO:1902966|GO:0005488|GO:0040012|GO:0040013|GO:0043005|GO:0098590|GO:0044297|GO:0008360|GO:0097708|GO:0009966|GO:0032587|GO:0009968|GO:0031256|GO:0031410|GO:0031253|GO:0001726|GO:0035330|GO:0031527|GO:1903829|GO:0060090|GO:0005654|GO:0035329|GO:0110053|GO:0007162|GO:2000145|GO:0014009|GO:2000146|GO:0008092|GO:2000026|GO:0030036|GO:0030674|GO:0030155|GO:0032970|GO:0016020|GO:0030154|GO:0023057|GO:1902905|GO:0051049|GO:1902903|GO:0030029|GO:0051726|GO:0030027|GO:0010812|GO:0005102|GO:0023051|GO:0010810|GO:0060284|GO:0120025|GO:1904893|GO:1904892|GO:0035556|GO:0044877|GO:0022407|GO:0098858|GO:0022408|GO:0051960|GO:0001952|GO:0044087|GO:0005515|GO:0001953|GO:0005911|GO:0005912|GO:0005634|GO:0032956|GO:0050767|GO:0042981|GO:0046426|GO:0046425|GO:1902115|GO:0045177|GO:0007010|GO:0044089|GO:0051050|GO:0031982|GO:0071840|GO:0032154
#>    GemmaIDs NCBIids      EnsemblIds
#>      <char>  <char>          <char>
#> 1:   127276    5352 ENSG00000152952
#> 2:   313978   57192 ENSG00000090674
#> 3:   188876    9020 ENSG00000006062
#> 4:     9353     390 ENSG00000115963
#> 5:    72301    3236 ENSG00000128710
#> 6:   116071    4771 ENSG00000186575
head(get_platform_annotations("Generic_human_ncbiIds"))
#>    ElementName  GeneSymbols                    GeneNames GOTerms GemmaIDs
#>          <int>       <char>                       <char>  <char>    <int>
#> 1:   105373804 LOC105373804 uncharacterized LOC105373804          9236458
#> 2:      285540   SEPSECS-DT SEPSECS divergent transcript           421333
#> 3:   105373805    STAT4-AS1        STAT4 antisense RNA 1          9236463
#> 4:   107985219 LOC107985219 uncharacterized LOC107985219          9667149
#> 5:   105373802                                                         NA
#> 6:   107985216 LOC107985216 uncharacterized LOC107985216          9667148
#>      NCBIids      EnsemblIds
#>        <int>          <char>
#> 1: 105373804                
#> 2:    285540                
#> 3: 105373805 ENSG00000231858
#> 4: 107985219                
#> 5:        NA                
#> 6: 107985216