Retrieve all experiments using a given platform

Retrieve all experiments using a given platform

get_platform_datasets(
  platform,
  offset = 0L,
  limit = 20L,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

platform

A platform numerical identifier or a platform short name

offset

The offset of the first retrieved result.

limit

Defaults to 20. Limits the result to specified amount of objects. Has a maximum value of 100. Use together with offset and the totalElements attribute in the output to compile all data if needed.

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

A data table with information about the queried dataset(s). A list if raw = TRUE. Returns an empty list if no datasets matched.

The fields of the output data.table are:

experiment.shortName: Shortname given to the dataset within Gemma. Often corresponds to accession ID

experiment.name: Full title of the dataset

experiment.ID: Internal ID of the dataset.

experiment.description: Description of the dataset

experiment.troubled: Did an automatic process within gemma or a curator mark the dataset as "troubled"

experiment.accession: Accession ID of the dataset in the external database it was taken from

experiment.database: The name of the database where the dataset was taken from

experiment.URI: URI of the original database

experiment.sampleCount: Number of samples in the dataset

experiment.batchEffectText: A text field describing whether the dataset has batch effects

experiment.batchCorrected: Whether batch correction has been performed on the dataset.

experiment.batchConfound: 0 if batch info isn't available, -1 if batch counfoud is detected, 1 if batch information is available and no batch confound found

experiment.batchEffect: -1 if batch p value < 0.0001, 1 if batch p value > 0.1, 0 if otherwise and when there is no batch information is available or when the data is confounded with batches.

experiment.rawData: -1 if no raw data available, 1 if raw data was available. When available, Gemma reprocesses raw data to get expression values and batches

geeq.qScore: Data quality score given to the dataset by Gemma.

geeq.sScore: Suitability score given to the dataset by Gemma. Refers to factors like batches, platforms and other aspects of experimental design

taxon.name: Name of the species

taxon.scientific: Scientific name for the taxon

taxon.ID: Internal identifier given to the species by Gemma

taxon.NCBI: NCBI ID of the taxon

taxon.database.name: Underlying database used in Gemma for the taxon

taxon.database.ID: ID of the underyling database used in Gemma for the taxon

Examples

head(get_platform_datasets("GPL1355"))
#>    experiment.shortName
#>                  <char>
#> 1:             GSE40348
#> 2:             GSE14903
#> 3:             GSE40421
#> 4:              GSE2872
#> 5:             GSE74676
#> 6:             GSE19366
#>                                                                                                                      experiment.name
#>                                                                                                                               <char>
#> 1:                                                                                                                    Hepatotoxicity
#> 2:                                    The relationship between brain mRNA levels and behaviour among inbred Lewis rats: Experiment 2
#> 3:                                                                 Generation of oligodendroglial cells by direct lineage conversion
#> 4:                                                                                                             d'mel-affy-rat-168311
#> 5: Toxicogenomic Characterization of Molecular Mechanisms Contributing to Chlorpyrifos Neurotoxicity in Adult Male Rats [microarray]
#> 6:                    Genomic-Derived Markers for Early Detection of Calcineurin Inhibitor Immunosuppressant–Mediated Nephrotoxicity
#>    experiment.ID
#>            <int>
#> 1:          5633
#> 2:         12033
#> 3:         12289
#> 4:             2
#> 5:         13570
#> 6:          5124
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            experiment.description
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            <char>
#> 1: Summary from subseries GSE40336:  This study provides an evaluation of changes in gene expression associated with acetominophen treatment of rat hepatocytes in vitro. Primary rat hepatocytes were treated for 24 and 48 hours with two doses (500 uM and 5 mM) of acetominophen and 1% DMSO vehicle control.  Five replicates of each treatment were performed.  Cells were then extracted and RNA processed for microarray analysis.\nSummary from subseries GSE40338:  This study provides an evaluation of changes in gene expression associated with sodium valproate treatment of rat hepatocytes in vitro. Primary rat hepatocytes were treated for 24 and 48 hours with two doses (500 uM and 10 mM) of sodium valproate and water vehicle control.  Five replicates of each treatment were performed.  Cells were then extracted and RNA processed for microarray analysis.\nSummary from subseries GSE40337:  This study provides an evaluation of changes in gene expression associated with dioctyl phthalate treatment of rat hepatocytes in vitro. Primary rat hepatocytes were treated for 24 and 48 hours with two doses (250 uM and 1 mM) of dioctyl phthalate and 1% DMSO vehicle control.  Five replicates of each treatment were performed.  Cells were then extracted and RNA processed for microarray analysis.\nSummary from subseries GSE40339:  This study provides an evaluation of changes in gene expression associated with phenobarbital treatment of rat hepatocytes in vitro. Primary rat hepatocytes were treated for 24 and 48 hours with two doses (300 uM and 3 mM) of phenobarbital and water vehicle control.  Five replicates of each treatment were performed.  Cells were then extracted and RNA processed for microarray analysis.\nSummary from subseries GSE40347:  This study provides an evaluation of changes in gene expression associated with methapyrilene treatment of rat hepatocytes in vitro. Primary rat hepatocytes were treated for 24 and 48 hours with two doses (3 uM and 100 uM) of methaphyriline and 1% DMSO vehicle control.  Five replicates of each treatment were performed.  Cells were then extracted and RNA processed for microarray analysis.\nSummary from subseries GSE40346:  This study provides an evaluation of changes in gene expression associated with WY-14643 treatment of rat hepatocytes in vitro. Primary rat hepatocytes were treated for 24 and 48 hours with two doses 8 uM and 200 uM) of WY-14643 and 1% DMSO vehicle control.  Five replicates of each treatment were performed.  Cells were then extracted and RNA processed for microarray analysis.\nSummary from subseries GSE40340:  This study provides an evaluation of changes in gene expression associated with beta-naphthaflavone treatment of rat hepatocytes in vitro. Primary rat hepatocytes were treated for 24 and 48 hours with two doses 1 uM and 100 mM) of beta-naphthaflavone and 1% DMSO vehicle control.  Five replicates of each treatment were performed.  Cells were then extracted and RNA processed for microarray analysis.\nSummary from subseries GSE40341:  This study provides an evaluation of changes in gene expression associated with chlorpromazine HCl treatment of rat hepatocytes in vitro. Primary rat hepatocytes were treated for 24 and 48 hours with two doses (0.8 uM and 20 uM) of chlorpromazine HCl and 1% DMSO vehicle control.  Five replicates of each treatment were performed.  Cells were then extracted and RNA processed for microarray analysis.\nSummary from subseries GSE40342:  This study provides an evaluation of changes in gene expression associated with diisononyl phthalate treatment of rat hepatocytes in vitro. Primary rat hepatocytes were treated for 24 and 48 hours with two doses (25 mM and 100 mM) of diisononyl phthalate and 1% DMSO vehicle control.  Five replicates of each treatment were performed.  Cells were then extracted and RNA processed for microarray analysis.\nSummary from subseries GSE40344:  This study provides an evaluation of changes in gene expression associated with clofibrate treatment of rat hepatocytes in vitro. Primary rat hepatocytes were treated for 24 and 48 hours with two doses (60 uM and 1 mM) of clobibrate and 1% DMSO vehicle control.  Five replicates of each treatment were performed.  Cells were then extracted and RNA processed for microarray analysis.\nLast Updated (by provider): Aug 28 2012\nContributors:  Jay P Tiesman
#> 2:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             Behaviour is derived from complex molecular interactions within the brain, in response to specific environmental stimuli. In some rare cases, the direct causes of behaviour have been attributed to the interactions of a single or small group of gene transcripts and proteins. We conducted two experiments with the hope of defining some of the molecular interactors for four separate behaviours: sugar feeding, locomotor activity in a novel field, and acoustic startle reflex, and prepulse inhibition of the acoustic startle reflex, which have been linked to prefrontal cortex dopaminergic function or as predictors of sensorimotor gating in diseases such as schizophrenia. Rats with high and low response phenotypes were selected to determine the differences between these two extremes of behaviour. From our analyses, transcripts of Homer1, a neuronal scaffolding protein which interacts with group1 metabotropic glutamate receptors, were found to be significantly correlated with array data in both experiments, and with behaviour data across three separate tests in the second experiment, indicating that this gene's transcripts and probably downstream protein interactions have a significant correlation with behaviour phenotype in the inbred Lewis rat. Future areas of pursuit for this data should involve modification of the expression of Homer1 in an isolated fashion to determine a pharmacological threshold for differences in behaviour. \nLast Updated (by provider): Jul 31 2017\nContributors:  Susan Rotzinger Arturas Petronis Albert H Wong Franco Vaccarino Laura A Feldcamp Mawahib O Semeralul Paul C Boutros
#> 3:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       We report the generation of induced oligodendrocyte precursor cells (iOPCs) by direct lineage conversion. Forced expression of the three transcription factors Sox10, Olig2 and Zfp536 was sufficient to convert mouse and rat fibroblasts into iOPCs with morphologies and gene expression signatures that resemble OPCs.\nLast Updated (by provider): Jul 31 2017\nContributors:  Samuele Marro Nan Yang Marius Wernig
#> 4:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       Neurological diseases disrupt the quality of the lives of patients and often lead to their death prematurely.  A common feature of most neurological diseases is the degeneration of neurons, which results from an inappropriate activation of apoptosis. Drugs that inhibit neuronal apoptosis could thus be candidates for therapeutic intervention in neurodegenerative disorders.  We have identified (and recently reported) a chemical called GW5074 that inhibits apoptosis in a variety of cell culture paradigms of neuronal apoptosis.  Additionally, we have found that GW5074 reduces neurodegeneration and improves behavioral outcome in a mouse model of Huntington's disease.  Although GW5074 is a c-Raf inhibitor, we know very little about the molecular mechanisms underlying its neuroprotective action.  Identifying genes that are regulated by GW5074 in neurons will shed insight into this issue. We believe that neuroprotection by GW5074 involves the regulation of several genes.  Some of these genes are likely to be induced whereas the expression of other genes might be inhibited.  The specific aim is to identify genes that are differentially expressed in neurons treated with GW5074. We believe that neuroprotection by GW5074 involves the regulation of several genes.  Some of these genes are likely to be induced whereas the expression of other genes might be inhibited. Cultures of cerebellar granule neurons undergo apoptosis when switched from medium containing levated levels of potassium (high K+ or HK) to medium containing low potassium (LK).  Although cell death begins at about 18 h, we have found that the by 6 h after LK treatment these neurons are irreversibly committed to cell death.  We will treat cerebellar granule neuron cultures with LK medium (which induces them to undergo apoptosis) or with GW5074 (1 uM).  We will extract RNA at two time-points after treatment: 3 h and 6 h. Analysis at the two different time-points will show us whether changes in expression of specific genes is transient or sustained or whether the changes occurs early or relatively late in the process.  Neuronal cultures will be prepared from 1 week old Wistar rats.  The cultures will be maintained in culture for 1 week before treatment.  Following treatment the cells will be lysed and total RNA isolated.  The RNA will be stored at -80oC and shipped to the microarray facility for analysis.  The experiment will be done in triplicate.  Thus, for each time-point (3 or 6h treatment) we will be hope to provide 3 sets of samples (each set coming from a different culture preparation and containing lysates from cells treated with LK or GW5074).  Having samples from 3 independent cultures will mitigate any expression differences resulting from subtle variations in culture quality or in the preparation or quality of RNA.\nDate GSE2872 Last Updated: Jul 06 2005\nContributors:  S R D'Mello\nIncludes GDS1837.\n Update date: Jun 09 2006.\n Dataset description GDS1837: Analysis of cultured cerebellar granule neurons in low potassium medium 3 and 6 hours following treatment with the c-Raf inhibitor GW5074. GW5074 blocks low potassium induced cell death. Results provide insight into the neuroprotective action of GW5074.
#> 5:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        Chlorpyrifos (CPF) is an organophosphorus pesticide (OP), and one of the most widely used pesticides in the world. Metabolites of CPF and other OPs continue to be identified in the majority of human samples, even in countries such as the United States where OP use is declining (Arcury et al., 2010). The effects of repeated occupational and environmental exposures to OPs are poorly understood, although human and animal studies consistently identify neurotoxicity as the primary endpoint of concern.  Thus, occupational exposures to sublethal doses of CPF are consistently associated with problems in cognitive abilities, such as learning and memory but the biological mechanism(s) underlying this association remain speculative.  To identify potential mechanisms of CPF neurotoxicity, we employed a rat model that simulated documented CPF exposures in Egyptian agricultural workers. We quantified mRNA expression profiles in the CA1 region of the hippocampus of adult male Long Evans (LE) rats administered CPF at 3 or 10 mg/kg/d (s.c.)  for 21 days. Despite significant inhibition of cholinesterase activity by the end of the 21 d exposure period, the CPF-exposed rats displayed minimal signs of cholinergic toxicity. Distinct hippocampal mRNA and miRNA signatures were associated with CPF exposure. Toxicogenomics-based evidence identified increased expression of neuropeptide genes in the hippocampi of CPF-exposed rats, which have been shown to activate receptor-mediated signaling pathways involved in cell survival. The analysis of small non-coding RNA profiles suggested the possibility that miR132/212-mediated homeostatic regulatory pathways may also be activated by repeated exposures to CPF. These findings identify potential molecular effects that may contribute to neurobehavioral deficits.\nLast Updated (by provider): Jul 31 2017\nContributors:  John A Lewis Jonathan D Stallings Naissan Hussainzada Pamela J Lein Young S Lee David A Jackson
#> 6:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            The use of calcineurin inhibitor (CI) immunosuppressants has significantly improved the early allograft survival rate in organ transplantation. However, CI therapy has been associated with chronic nephrotoxicity, which limits their long-term utility. In order to understand the mechanisms of the toxicity, we analyzed the gene expression changes that underlie the development of CI immunosuppressant-mediated nephrotoxicity, in male Sprague-Dawley (SD) rats dosed daily with cyclosporine (CsA), FK506 or rapamycin (Rapa) for 1 to 28 days. We identified a group of genes, whose expression in rat kidney is quantitatively correlated with CI-induced kidney injury as observed in changes in blood urea nitrogen (BUN) levels and kidney histopathology. These genes include both up-regulated genes, such as Ren1 and Klks3, and down-regulated genes, such as Calb1, Egf, NCC, and kidney specific Wnk1 (KS-Wnk1). Using the down-regulated genes alone we successfully predicted CI immunosuppressant-mediated kidney injury in rats following 7 days of treatment. Among these genes are two mechanism-related genes, NCC and KS-Wnk1, both of which are involved in the sodium transport in the distal nephrons. The down-regulation of both genes at the mRNA and protein level in rat kidney following CI treatment was confirmed by quantitative RT-PCR and immunohistochemical staining, respectively. We hypothesize that decreased expression of NCC may cause reduced sodium chloride reabsorption in the distal tubules, and contribute to the prolonged activation of the Renin-Angiotensin-System (RAS), a demonstrated contributor to the development of CI-induced nephrotoxicity in both animal models and clinical settings. Therefore, NCC and KS-Wnk1 could potentially be used as biomarkers for early detection and prevention of CI-related nephrotoxicity in clinical practice.\nLast Updated (by provider): Mar 21 2012\nContributors:  James T Auman Richard S Paules Jeff Chou Qihong Huang Supriya Jayadev Yuxia Cui
#>    experiment.troubled experiment.accession experiment.database
#>                 <lgcl>               <char>              <char>
#> 1:               FALSE             GSE40348                 GEO
#> 2:               FALSE             GSE14903                 GEO
#> 3:               FALSE             GSE40421                 GEO
#> 4:               FALSE              GSE2872                 GEO
#> 5:               FALSE             GSE74676                 GEO
#> 6:               FALSE             GSE19366                 GEO
#>                                                 experiment.URI
#>                                                         <char>
#> 1: https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE40348
#> 2: https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE14903
#> 3: https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE40421
#> 4:  https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE2872
#> 5: https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE74676
#> 6: https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE19366
#>    experiment.sampleCount experiment.lastUpdated experiment.batchEffectText
#>                     <int>                 <POSc>                     <char>
#> 1:                    300    2023-12-19 13:26:42       BATCH_EFFECT_FAILURE
#> 2:                     10    2023-12-17 12:18:16       SINGLE_BATCH_SUCCESS
#> 3:                      9    2024-03-18 16:21:27       BATCH_EFFECT_FAILURE
#> 4:                     12    2024-01-02 18:57:36       SINGLE_BATCH_SUCCESS
#> 5:                     60    2023-12-21 00:32:14    BATCH_CORRECTED_SUCCESS
#> 6:                    120    2023-12-17 18:53:19       BATCH_EFFECT_FAILURE
#>    experiment.batchCorrected experiment.batchConfound experiment.batchEffect
#>                       <lgcl>                    <num>                  <num>
#> 1:                     FALSE                       -1                      0
#> 2:                     FALSE                        1                      1
#> 3:                     FALSE                        1                      0
#> 4:                     FALSE                        1                      1
#> 5:                      TRUE                        1                      1
#> 6:                     FALSE                       -1                      0
#>    experiment.rawData geeq.qScore geeq.sScore taxon.name  taxon.scientific
#>                 <num>       <num>       <num>     <char>            <char>
#> 1:                  1   0.2845997      0.6250        rat Rattus norvegicus
#> 2:                  1   0.8563497      0.5000        rat Rattus norvegicus
#> 3:                  1   0.5511185      0.7125        rat Rattus norvegicus
#> 4:                  1   0.8552303      0.5000        rat Rattus norvegicus
#> 5:                  1   0.9987732      0.6250        rat Rattus norvegicus
#> 6:                  1   0.4268694      0.8750        rat Rattus norvegicus
#>    taxon.ID taxon.NCBI taxon.database.name taxon.database.ID
#>       <int>      <int>              <char>             <int>
#> 1:        3      10116                 rn6                86
#> 2:        3      10116                 rn6                86
#> 3:        3      10116                 rn6                86
#> 4:        3      10116                 rn6                86
#> 5:        3      10116                 rn6                86
#> 6:        3      10116                 rn6                86