Processes JSON as a vector of datasets

Processes JSON as a vector of datasets

processDatasets(d)

Arguments

d

The JSON to process

Value

A data table with information about the queried dataset(s). A list if raw = TRUE. Returns an empty list if no datasets matched. A successful response may contain 'Geeq' information, which aims to provide a unified metric to measure experiments by the quality of their data, and their suitability for use in Gemma. You can read more about the geeq properties here.

The fields of the output data.table are:

experiment.ShortName: Shortname given to the dataset within Gemma. Often corresponds to accession ID

experiment.Name: Full title of the dataset

experiment.ID: Internal ID of the dataset.

experiment.Description: Description of the dataset

experiment.Public: Is the dataset publicly available. Only useful for logged in users with access to non-public data

experiment.Troubled: Did an automatic process within gemma or a curator mark the dataset as "troubled"

experiment.Accession: Accession ID of the dataset in the external database it was taken from

experiment.Database: The name of the database where the dataset was taken from

experiment.URI: URI of the original database

experiment.SampleCount: Number of samples in the dataset

experiment.batchEffect: A text field describing whether the dataset has batch effects

experiment.batchCorrected: Whether batch correction has been performed on the dataset. This is a text field with possible details about the batch correction. Use geeq.batchEffect if a more structured field is needed.

geeq.batchConfound: 0 if batch info isn't available, -1 if batch counfoud is detected, 1 if batch information is available and no batch confound found

geeq.batchEffect: -1 if batch p value < 0.0001, 1 if batch p value > 0.1, 0 if otherwise and when there is no batch information is available or when the data is confounded with batches.

geeq.rawData: -1 if no raw data available, 1 if raw data was available. When available, Gemma reprocesses raw data to get expression values and batches

geeq.qScore: Data quality score given to the dataset by Gemma.

geeq.sScore: Suitability score given to the dataset by Gemma. Refers to factors like batches, platforms and other aspects of experimental design

taxon.Name: The taxa of the study. In Gemma each study will include a single species. If the original source has samples from multiple species, they will be split into different studies within Gemma

taxon.ID: Internal ID given to the taxon by Gemma