Annotation search

Does a search for annotations based on the given string

searchAnnotations(
  query,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

query

Required, defaults to empty. The search query. Either plain text ('traumatic'), or an ontology term URI ('http://purl.obolibrary.org/obo/UBERON_0002048'). Datasets that contain the given string in their short of full name will also be matched ('GSE201', 'Bronchoalveolar lavage samples'. Can be multiple identifiers separated by commas. When using in scripts, remember to URL-encode any forward slashes in the phenotype value URIs (see the compiled URL below).

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

An array of annotations (annotation search result value objects) matching the given identifiers. A 400 error if required parameters are missing.

Examples

# \donttest{
searchAnnotations("traumatic")
#>     category.Name                         category.URL
#>  1:     treatment http://www.ebi.ac.uk/efo/EFO_0000727
#>  2:     treatment http://www.ebi.ac.uk/efo/EFO_0000727
#>  3:          <NA>                                 <NA>
#>  4:      genotype http://www.ebi.ac.uk/efo/EFO_0000513
#>  5:       disease http://www.ebi.ac.uk/efo/EFO_0000408
#>  6:     BioSource http://www.ebi.ac.uk/efo/EFO_0000635
#>  7:     BioSource http://www.ebi.ac.uk/efo/EFO_0000635
#>  8:          <NA>                                 <NA>
#>  9:     BioSource http://www.ebi.ac.uk/efo/EFO_0000635
#> 10:          <NA>                                 <NA>
#> 11:          <NA>                                 <NA>
#> 12:          <NA>                                 <NA>
#> 13:     BioSource http://www.ebi.ac.uk/efo/EFO_0000635
#> 14:     BioSource http://www.ebi.ac.uk/efo/EFO_0000635
#>                                                                                            value.Name
#>  1:                                                                            Traumatic brain injury
#>  2:                                                                            traumatic brain injury
#>  3:                                                                  traumatic optic neuropathy (TON)
#>  4:                                                                  traumatic optic neuropathy (TON)
#>  5:                                                                    post-traumatic stress disorder
#>  6:                 Hippocampus LCM Neurons, traumatic brain injury, Saline, 2wk, pool from 3 animals
#>  7: Hippocampus LCM Neurons, traumatic brain injury, Scramble virus control, 2wk, pool from 3 animals
#>  8:                                                                 treatment: traumatic brain injury
#>  9:        Hippocampus LCM Neurons, traumatic brain injury, nNos3 knockdown, 2wk, pool from 3 animals
#> 10:                                                           treatment: traumatic brain injury (TBI)
#> 11:                                                                  received: traumatic brain injury
#> 12:                                                treatment: contralateral to traumatic brain injury
#> 13:         Hippocampus LCM Neurons, traumatic brain injury, GPX2 knockdown, 2wk, pool from 3 animals
#> 14:                                                            lateral cortex, traumatic brain injury
#>                                                                      value.URL
#>  1: http://ontology.neuinfo.org/NIF/Dysfunction/NIF-Dysfunction.owl#nlx_143898
#>  2:                                                                       <NA>
#>  3:                                                                       <NA>
#>  4:                                                                           
#>  5:                                   http://purl.obolibrary.org/obo/DOID_2055
#>  6:                                                                       <NA>
#>  7:                                                                       <NA>
#>  8:                                                                       <NA>
#>  9:                                                                       <NA>
#> 10:                                                                       <NA>
#> 11:                                                                       <NA>
#> 12:                                                                       <NA>
#> 13:                                                                       <NA>
#> 14:                                                                       <NA>
# }