Search for annotation tags

Search for annotation tags

search_annotations(
  query,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

query

The search query

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

A data table with annotations (annotation search result value objects) matching the given identifiers. A list if raw = TRUE. A 400 error if required parameters are missing.

The fields of the output data.table are:

category.Name: Category that the annotation belongs to

category.URI: URI for the category.Name

value.Name: Annotation term

value.URI: URI for the value.Name

Examples

search_annotations("traumatic")
#>      category.name                               category.URI
#>             <char>                                     <char>
#>   1:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>   2:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>   3:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>   4:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>   5:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>   6:      protocol http://purl.obolibrary.org/obo/OBI_0000272
#>   7:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>   8:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>   9:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  10:      protocol http://purl.obolibrary.org/obo/OBI_0000272
#>  11:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  12:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  13:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  14:      protocol http://purl.obolibrary.org/obo/OBI_0000272
#>  15:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  16:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  17:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  18:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  19:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  20:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  21:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  22:      received                                       <NA>
#>  23:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  24:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  25:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  26:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  27:      received                                       <NA>
#>  28:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  29:      received                                       <NA>
#>  30:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  31:      received                                       <NA>
#>  32:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  33:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  34:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  35:      received                                       <NA>
#>  36:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  37:      received                                       <NA>
#>  38:         group                                       <NA>
#>  39:      received                                       <NA>
#>  40:      protocol http://purl.obolibrary.org/obo/OBI_0000272
#>  41:      received                                       <NA>
#>  42:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  43:      protocol http://purl.obolibrary.org/obo/OBI_0000272
#>  44:         group                                       <NA>
#>  45:      received                                       <NA>
#>  46:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  47:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  48:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  49:      received                                       <NA>
#>  50:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  51:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  52:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  53:      received                                       <NA>
#>  54:      genotype       http://www.ebi.ac.uk/efo/EFO_0000513
#>  55:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  56:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  57:         group                                       <NA>
#>  58:      received                                       <NA>
#>  59:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  60:      protocol http://purl.obolibrary.org/obo/OBI_0000272
#>  61:      protocol http://purl.obolibrary.org/obo/OBI_0000272
#>  62:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  63:      genotype       http://www.ebi.ac.uk/efo/EFO_0000513
#>  64:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  65:      protocol http://purl.obolibrary.org/obo/OBI_0000272
#>  66:      protocol http://purl.obolibrary.org/obo/OBI_0000272
#>  67:      protocol http://purl.obolibrary.org/obo/OBI_0000272
#>  68:  case/control                                       <NA>
#>  69:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  70:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  71:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  72:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  73:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  74:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  75:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  76:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  77:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  78:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  79:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  80:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  81:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  82:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  83:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  84:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  85:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  86:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  87:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  88:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  89:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  90:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  91:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  92:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  93:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  94:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  95:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#>  96:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  97:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>  98:     BioSource                                       <NA>
#>  99:     treatment       http://www.ebi.ac.uk/efo/EFO_0000727
#> 100:       disease       http://www.ebi.ac.uk/efo/EFO_0000408
#>      category.name                               category.URI
#>                                                  value.name
#>                                                      <char>
#>   1:                               traumatic encephalopathy
#>   2:                                 traumatic brain injury
#>   3:                                 traumatic brain injury
#>   4:                                 traumatic brain injury
#>   5:                                 traumatic brain injury
#>   6:                                 traumatic brain injury
#>   7:                                 traumatic brain injury
#>   8:                                 traumatic brain injury
#>   9:                                 traumatic brain injury
#>  10:                                 traumatic brain injury
#>  11:                                 traumatic brain injury
#>  12:                                 traumatic brain injury
#>  13:                           traumatic brain injury (TBI)
#>  14:                                 traumatic brain injury
#>  15:                                 traumatic brain injury
#>  16:                                 traumatic brain injury
#>  17:                           traumatic brain injury (TBI)
#>  18:                                 traumatic brain injury
#>  19:                           traumatic brain injury (TBI)
#>  20:                                 traumatic brain injury
#>  21:                                 traumatic brain injury
#>  22:                                 traumatic brain injury
#>  23:                           traumatic brain injury (TBI)
#>  24:                                 traumatic brain injury
#>  25:                                 traumatic brain injury
#>  26:                                 traumatic brain injury
#>  27:                                 traumatic brain injury
#>  28:                           traumatic brain injury (TBI)
#>  29:                                 traumatic brain injury
#>  30:                           traumatic brain injury (TBI)
#>  31:                                 traumatic brain injury
#>  32:                                 traumatic brain injury
#>  33:                           traumatic brain injury (TBI)
#>  34:                                 traumatic brain injury
#>  35:                                 traumatic brain injury
#>  36:                           traumatic brain injury (TBI)
#>  37:                                 traumatic brain injury
#>  38:                       traumatic optic neuropathy (TON)
#>  39:                                 traumatic brain injury
#>  40:                                 traumatic brain injury
#>  41:                                 traumatic brain injury
#>  42:                                 traumatic brain injury
#>  43:                                 traumatic brain injury
#>  44:                       traumatic optic neuropathy (TON)
#>  45:                                 traumatic brain injury
#>  46:                           traumatic brain injury (TBI)
#>  47:                                 traumatic brain injury
#>  48:                                 traumatic brain injury
#>  49:                                 traumatic brain injury
#>  50:                                 traumatic brain injury
#>  51:                                 traumatic brain injury
#>  52:                                 traumatic brain injury
#>  53:                                 traumatic brain injury
#>  54:                       traumatic optic neuropathy (TON)
#>  55:                           traumatic brain injury (TBI)
#>  56:                           traumatic brain injury (TBI)
#>  57:                       traumatic optic neuropathy (TON)
#>  58:                                 traumatic brain injury
#>  59:                                 traumatic brain injury
#>  60:                                 traumatic brain injury
#>  61:                                 traumatic brain injury
#>  62:                                 traumatic brain injury
#>  63:                       traumatic optic neuropathy (TON)
#>  64:                                 traumatic brain injury
#>  65:                                 traumatic brain injury
#>  66:                                 traumatic brain injury
#>  67:                                 traumatic brain injury
#>  68:                         post-traumatic stress disorder
#>  69:                                post-traumatic epilepsy
#>  70:                 Chronic Traumatic Encelopathy - severe
#>  71:                 Chronic Traumatic Encelopathy - severe
#>  72:                 Chronic Traumatic Encelopathy - severe
#>  73:                 Chronic Traumatic Encelopathy - severe
#>  74:                 Chronic Traumatic Encelopathy - severe
#>  75:                 Chronic Traumatic Encelopathy - severe
#>  76:                 Chronic Traumatic Encelopathy - severe
#>  77:                 Chronic Traumatic Encelopathy - severe
#>  78:                 Chronic Traumatic Encelopathy - severe
#>  79:                 Chronic Traumatic Encelopathy - severe
#>  80: Chronic Traumatic Encelopathy - repetitive head injury
#>  81:                   Chronic Traumatic Encelopathy - mild
#>  82: Chronic Traumatic Encelopathy - repetitive head injury
#>  83:                 Chronic Traumatic Encelopathy - severe
#>  84:                 Chronic Traumatic Encelopathy - severe
#>  85:                 Chronic Traumatic Encelopathy - severe
#>  86: Chronic Traumatic Encelopathy - repetitive head injury
#>  87:                 Chronic Traumatic Encelopathy - severe
#>  88:                 Chronic Traumatic Encelopathy - severe
#>  89:                            Mild traumatic brain injury
#>  90:                 Chronic Traumatic Encelopathy - severe
#>  91: Chronic Traumatic Encelopathy - repetitive head injury
#>  92:                 Chronic Traumatic Encelopathy - severe
#>  93:                            Mild traumatic brain injury
#>  94:                 Chronic Traumatic Encelopathy - severe
#>  95:                            Mild traumatic brain injury
#>  96:                 Chronic Traumatic Encelopathy - severe
#>  97: Chronic Traumatic Encelopathy - repetitive head injury
#>  98:                 lateral cortex, traumatic brain injury
#>  99:                contralateral to traumatic brain injury
#> 100:                 Chronic Traumatic Encelopathy - severe
#>                                                  value.name
#>                                         value.URI
#>                                            <char>
#>   1: http://purl.obolibrary.org/obo/MONDO_0043512
#>   2: http://purl.obolibrary.org/obo/MONDO_0858950
#>   3:                                         <NA>
#>   4:                                         <NA>
#>   5:                                         <NA>
#>   6:                                         <NA>
#>   7:                                         <NA>
#>   8:                                         <NA>
#>   9:                                         <NA>
#>  10:                                         <NA>
#>  11:                                         <NA>
#>  12:                                         <NA>
#>  13:                                         <NA>
#>  14:                                         <NA>
#>  15:                                         <NA>
#>  16:                                         <NA>
#>  17:                                         <NA>
#>  18:                                         <NA>
#>  19:                                         <NA>
#>  20:                                         <NA>
#>  21:                                         <NA>
#>  22:                                         <NA>
#>  23:                                         <NA>
#>  24:                                         <NA>
#>  25:                                         <NA>
#>  26:                                         <NA>
#>  27:                                         <NA>
#>  28:                                         <NA>
#>  29:                                         <NA>
#>  30:                                         <NA>
#>  31:                                         <NA>
#>  32:                                         <NA>
#>  33:                                         <NA>
#>  34:                                         <NA>
#>  35:                                         <NA>
#>  36:                                         <NA>
#>  37:                                         <NA>
#>  38:                                         <NA>
#>  39:                                         <NA>
#>  40:                                         <NA>
#>  41:                                         <NA>
#>  42:                                         <NA>
#>  43:                                         <NA>
#>  44:                                         <NA>
#>  45:                                         <NA>
#>  46:                                         <NA>
#>  47:                                         <NA>
#>  48:                                         <NA>
#>  49:                                         <NA>
#>  50:                                         <NA>
#>  51:                                         <NA>
#>  52:                                         <NA>
#>  53:                                         <NA>
#>  54:                                         <NA>
#>  55:                                         <NA>
#>  56:                                         <NA>
#>  57:                                         <NA>
#>  58:                                         <NA>
#>  59:                                         <NA>
#>  60:                                         <NA>
#>  61:                                         <NA>
#>  62:                                         <NA>
#>  63:                                         <NA>
#>  64:                                         <NA>
#>  65:                                         <NA>
#>  66:                                         <NA>
#>  67:                                         <NA>
#>  68: http://purl.obolibrary.org/obo/MONDO_0005146
#>  69: http://purl.obolibrary.org/obo/MONDO_0043264
#>  70:                                         <NA>
#>  71:                                         <NA>
#>  72:                                         <NA>
#>  73:                                         <NA>
#>  74:                                         <NA>
#>  75:                                         <NA>
#>  76:                                         <NA>
#>  77:                                         <NA>
#>  78:                                         <NA>
#>  79:                                         <NA>
#>  80:                                         <NA>
#>  81:                                         <NA>
#>  82:                                         <NA>
#>  83:                                         <NA>
#>  84:                                         <NA>
#>  85:                                         <NA>
#>  86:                                         <NA>
#>  87:                                         <NA>
#>  88:                                         <NA>
#>  89:                                         <NA>
#>  90:                                         <NA>
#>  91:                                         <NA>
#>  92:                                         <NA>
#>  93:                                         <NA>
#>  94:                                         <NA>
#>  95:                                         <NA>
#>  96:                                         <NA>
#>  97:                                         <NA>
#>  98:                                         <NA>
#>  99:                                         <NA>
#> 100:                                         <NA>
#>                                         value.URI