Search everything in Gemma

Search everything in Gemma

search_gemma(
  query,
  taxon = NA_character_,
  platform = NA_character_,
  limit = 20,
  resultType = "experiment",
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

query

The search query. Either plain text ('traumatic'), or an ontology term URI ('http://purl.obolibrary.org/obo/UBERON_0002048'). Datasets that contain the given string in their short of full name will also be matched. Can be multiple identifiers separated by commas.

taxon

A numerical taxon identifier or an ncbi taxon identifier or a taxon identifier that matches either its scientific or common name

platform

A platform numerical identifier or a platform short name

limit

Optional, defaults to 20. Limits the result to specified amount of objects. Has a maximum value of 100. Use together with offset and the totalElements attribute in the output to compile all data if needed.

resultType

The kind of results that should be included in the output. Can be experiment, gene, platform or a long object type name, documented in the API documentation.

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

If raw = FALSE and resultType is experiment, gene or platform, a data.table containing the search results. If it is any other type, a list of results. A list with additional details about the search if raw = TRUE

Examples

search_gemma("bipolar")
#>     experiment.ShortName
#>  1:             GSE67645
#>  2:             GSE35077
#>  3:             GSE33085
#>  4:             GSE29318
#>  5:             GSE98838
#>  6:            GSE126942
#>  7:            GSE127771
#>  8:             GSE10843
#>  9:       stanley_dobrin
#> 10:             GSE66196
#> 11:         stanley_chen
#> 12:             GSE23848
#> 13:     stanley_feinberg
#> 14:         stanley_kato
#> 15:            GSE210064
#> 16:       stanley_sklarB
#> 17:              byne-cc
#> 18:   McLean Hippocampus
#> 19:             GSE46449
#> 20:        stanley_young
#>                                                                                                                                                                                       experiment.Name
#>  1: Transcriptome dynamics of developing photoreceptors in 3-D retina cultures recapitulates temporal sequence of human cone and rod differentiation revealing cell surface markers and gene networks
#>  2:                                                                                                                                            A gene expression database for retinal neuron subtypes
#>  3:                                                                                                                                                Transcriptome analysis of adult retina cell types.
#>  4:                                                                                                                                             Expression profile of FAC-sorted murine retinal cells
#>  5:                                                                                                        Combining RNA-Seq and Somatic CRISPR Mutagenesis to Study Mouse Neural Development in vivo
#>  6:                                              The LIM-homeodomain transcription factor LHX4 is required for the differentiation of retinal rod bipolar cells and rod-connecting cone bipolar cells
#>  7:                                         The LIM-homeodomain transcription factor LHX4 is required for the differentiation of retinal rod bipolar cells and rod-connecting cone bipolar cells [P7]
#>  8:                                                                                                                                                                    mRNA Cancer Cell Line Profiles
#>  9:                                                                                                                                                           Stanley array collection DLPFC - Dobrin
#> 10:                                                                                                                Bipolar disorder and lithium-induced gene expression in two peripheral cell models
#> 11:                                                                                                                                                        Stanley consortium collection DLPFC - Chen
#> 12:                                                                                                 Peripheral blood gene-expression in depressed subjects with bipolar disorder vs healthy controls.
#> 13:                                                                                                                                               Stanley consortium collection Cerebellum - Feinberg
#> 14:                                                                                                                                                             Stanley array collection DLPFC - Kato
#> 15:                                                               Gene expression alterations in the postmortem hippocampus from older patients with bipolar disorder – a hypothesis generating study
#> 16:                                                                                                                                                 Stanley consortium collection Cerebellum - SklarB
#> 17:                                                                                                                                              Corpus Callosum data from Stanley collection samples
#> 18:                                                                                                                                                                                McLean Hippocampus
#> 19:                                                                                                             Expression data from Patients with Bipolar (BP) Disorder and Matched Control Subjects
#> 20:                                                                                                                                                            Stanley array collection DLPFC - Young
#>     experiment.ID
#>  1:         11634
#>  2:          7608
#>  3:          8887
#>  4:         12310
#>  5:         13214
#>  6:         19454
#>  7:         19455
#>  8:          7196
#>  9:           843
#> 10:         24428
#> 11:           842
#> 12:          1958
#> 13:           831
#> 14:           833
#> 15:         28041
#> 16:           835
#> 17:          4354
#> 18:           670
#> 19:          6583
#> 20:           837
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    experiment.Description
#>  1:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           To define molecular mechanisms underlying rod and cone differentiation, we generated H9 human embryonic stem cell line carrying a GFP reporter that is controlled by the promoter of cone-rod homeobox (CRX) gene, the first known marker of post-mitotic photoreceptor precursors. CRXp-GFP reporter in H9 line replicates endogenous CRX expression when induced to form self-organizing 3-D retina-like tissue. We define temporal transcriptome dynamics of developing photoreceptors during the establishment of cone and rod cell fate. Our studies provide an essential framework for delineating molecules and cellular pathways that guide human photoreceptor development and should assist in chemical screening and cell-based therapies of retinal degeneration.\nLast Updated (by provider): Nov 14 2017\nContributors:  Kohei Homma Rossukon Kaewkhaw Anand Swaroop Koray D Kaya Jizhong Zou Mahendra Rao Matthew Brooks Vijender Chaitankar
#>  2:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   The goal of this experiment was to define gene expression patterns of thirteen mouse retinal neuron subsets, labeled by expression of fluorescent proteins in transgenic mice.\nLast Updated (by provider): Sep 23 2013\nContributors:  Jeremy N Kay Joshua R Sanes
#>  3: Brain circuits are assembled from a large variety of morphologically and functionally diverse cell types. It is not known how the intermingled cell types of individual brain regions differ in their expressed genomes. Here we describe an atlas of cell type transcriptomes of the adult retina. We found that each adult cell type expresses a specific set of genes, including a unique set of transcription factors, forming a “barcode” for cell identity. Cell type transcriptomes carry enough information to categorize cells into corresponding morphological classes and types. Surprisingly, several barcode genes are eye disease-associated genes that we demonstrate to be specifically expressed not only in photoreceptors but also in particular retinal circuit elements such as inhibitory neurons as well as in retinal microglia. Our data suggest that distinct cell types of individual brain regions are characterized by marked differences in their expressed genomes. We assembled a library of 22 transgenic mouse lines in which each line had a group of retinal cells marked with fluorescent proteins. We built up the library with the purpose of having some mouse lines in which single retinal cell types and others in which combinations of types from a single class are labeled. The library had mouse lines with labeled cells representing each of the six retinal cell classes. Retinal cells were characterized by physiological recording and immunohistochemical staining. We isolated 200 fluorescent protein-labeled retinal cells (“cell groups”) from at least three different mice of each mouse line by fluorescence-activated cell sorting. The transcripts of each cell group of these biological triplicates were independently amplified in batches. Each batch contained an internal control cell group from the Arc-line.\nLast Updated (by provider): Jun 10 2014\nContributors:  Botond Roska Erik Cabuy Sandra Siegert
#>  4:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Microarray experiments were performed using FAC-sorted young photoreceptors to analyze their transcriptome in comparison to remaining retinal cells at same developmental stage and retinal progenitors.\nLast Updated (by provider): Apr 02 2018\nContributors:  Marius Ader Kai Postel
#>  5:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  We used RNA sequencing to identify candidate regulators of interactions between photoreceptor axons and bipolar cell (BCs) dendrites in developing mouse retina. We chose three time points: P7, just after the OPL forms and synaptogenesis with BCs begins; P13, as synaptogenesis nears completion and sublamination begins; and P30, when the OPL is mature. We purified cone and rod photoreceptors separately by fluorescence activated cell sorting (FACS) using transgene markers: Rhoicre;Ai9 for rods and HRGPcre;Ai9 for cones. We purified ON BCs, which include ON cone bipolars plus rod bipolars using Grm6:GFP. As appropriate transgenic lines to separate RBCs from CBCs were not available, we performed RNAseq on cells from Grm6:GFP mice that were fixed and immunostained prior to FACS, allowing us to purify RBCs (GFP+PKC+) and CBCs (GFP+PKC-) from the same retinas. As PKC is not highly expressed at P7, profiling of developing rod bipolars separate from developing ON cone bipolars was restricted to P13.\nLast Updated (by provider): Jun 25 2018\nContributors:  Elizabeth Z Sanchez Lawrence S Zipursky Yerbol Z Kurmangaliyev
#>  6:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Study the role of the LIM-homeodomain transcription factor LHX4 in the development of retinal bipolar cell subtypes\nAt time of import, last updated (by provider) on: Sep 30 2020\n\nContributors: ; [Lin Gan, Xuhui Dong]
#>  7:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Study the role of the LIM-homeodomain transcription factor LHX4 in the development of retinal bipolar cell subtypes\nAt time of import, last updated (by provider) on: Sep 30 2020\n\nContributors: ; [Lin Gan, Xuhui Dong]
#>  8:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               mRNA cancer cell line profiles\nLast Updated (by provider): Oct 25 2013
#>  9:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                78 samples of individuals from three different diagnostic groups: bipolar, schizophrenia and controls. Samples taken from the DLPFC Broadmann area 46.
#> 10:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             Bipolar disorder is a severe and heritable psychiatric disorder and affects up to 1% of the population worldwide. Lithium is recommended as first-line treatment for the maintenance treatment of bipolar-affective disorder in current guidelines, its molecular modes of action are however poorly understood. Cell models derived from bipolar patients could prove useful to gain more insight in the molecular mechanisms of bipolar disorder and the common pharmacological treatments. As primary neuronal cell lines cannot be easily derived from patients, peripheral cell models should be evaluated in their usefulness to study pathomechanisms and the mode of action of medication as well as in regard to develop biomarkers for diagnosis and treatment response.\nAt time of import, last updated (by provider) on: Mar 25 2019\n\nContributors: ; [Sarah Kittel-Schneider, Max Hilscher, Andreas Reif, Claus J Scholz]
#> 11:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  27 samples of individuals from two different diagnostic groups: bipolar, and controls. Samples taken from the DLPFC Brodmann area 6.
#> 12:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Analysis of gene-expression changes in depressed subjects with bipolar disorder compared to healthy controls. Results provide information on pathways that may be involved in the pathogenesis of bipolar depression.\nLast Updated (by provider): Aug 27 2010\nContributors:  Robert D Beech
#> 13:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     50 samples of individuals from 4 different diagnostic groups: bipolar, schizophrenia, depression and controls. Samples taken from the Cerebellum.
#> 14:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  102 samples of individuals from 3 different diagnostic groups: bipolar, schizophrenia, and controls. Samples taken from the DLPFC Broadmann area 46.
#> 15:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             Gene expression of samples from the postmortem hippocampus of older bipolar disorder subjects and controls. Gene expression data was generated using the SurePrint G3 Human Gene Expression v3 microarray. Rank feature selection was performed to identify a subset of features that can optimally differentiate BD and controls.\nAt time of import, last updated (by provider) on: Feb 19 2023\n\nContributors: ; [Carlos A Pasqualucci, Claudia K Suemoto, Ricardo Nitrini, Fernanda B Bertonha, Paula V Nunes, Katia C De Oliveira, Carlos M Filho, Helena K Kim, Helena Brentani, Lea T Grinberg, Beny Lafer, André Barbosa, Camila Nascimento, Renata P Leite, Wilson Jacob-Filho]
#> 16:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     46 samples of individuals from 4 different diagnostic groups: bipolar, schizophrenia, depression and controls. Samples taken from the Cerebellum.
#> 17:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> 18:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Hippocampus of schizophrenic, bipolar, and control subjects. Analyzed from CEL files.
#> 19:                                                     There are currently no biological tests that differentiate patients with bipolar disorder (BPD) from healthy controls. While there is evidence that peripheral gene expression differences between patients and controls can be utilized as biomarkers for psychiatric illness, it is unclear whether current use or residual effects of antipsychotic and mood stabilizer medication drives much of the differential transcription. We therefore tested whether expression changes in first-episode, never-medicated bipolar patients, can contribute to a biological classifier that is less influenced by medication and could potentially form a practicable biomarker assay for BPD. We employed microarray technology to measure global leukocyte gene expression in first-episode (n=3) and currently medicated BPD patients (n=26), and matched healthy controls (n=25). Following an initial feature selection of the microarray data, we developed a cross-validated 10-gene model that was able to correctly predict the diagnostic group of the training sample (26 medicated patients and 12 controls), with 89% sensitivity and 75% specificity (p<0.001). The 10-gene predictor was further explored via testing on an independent test cohort consisting of three pairs of monozygotic twins discordant for BPD, plus the original enrichment sample cohort (the three never-medicated BPD patients and 13 matched control subjects), and a sample of experimental replicates (n=34). 83% of the independent test sample was correctly predicted, with a sensitivity of 67% and specificity of 100% (although this result did not reach statistical significance). Additionally, 88% of sample diagnostic classes were classified correctly for both the enrichment (p=0.015) and the replicate samples (p<0.001).\nLast Updated (by provider): Jun 25 2013\nContributors:  James D Clelland Catherine L Clelland
#> 20:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  105 samples of individuals from 3 different diagnostic groups: bipolar, schizophrenia, and controls. Samples taken from the DLPFC Broadmann area 46.
#>     experiment.Troubled experiment.Accession experiment.Database experiment.URI
#>  1:               FALSE                 <NA>                <NA>           <NA>
#>  2:               FALSE                 <NA>                <NA>           <NA>
#>  3:               FALSE                 <NA>                <NA>           <NA>
#>  4:               FALSE                 <NA>                <NA>           <NA>
#>  5:               FALSE                 <NA>                <NA>           <NA>
#>  6:               FALSE                 <NA>                <NA>           <NA>
#>  7:               FALSE                 <NA>                <NA>           <NA>
#>  8:               FALSE                 <NA>                <NA>           <NA>
#>  9:               FALSE                 <NA>                <NA>           <NA>
#> 10:               FALSE                 <NA>                <NA>           <NA>
#> 11:               FALSE                 <NA>                <NA>           <NA>
#> 12:               FALSE                 <NA>                <NA>           <NA>
#> 13:               FALSE                 <NA>                <NA>           <NA>
#> 14:               FALSE                 <NA>                <NA>           <NA>
#> 15:               FALSE                 <NA>                <NA>           <NA>
#> 16:               FALSE                 <NA>                <NA>           <NA>
#> 17:               FALSE                 <NA>                <NA>           <NA>
#> 18:               FALSE                 <NA>                <NA>           <NA>
#> 19:               FALSE                 <NA>                <NA>           <NA>
#> 20:               FALSE                 <NA>                <NA>           <NA>
#>     experiment.SampleCount experiment.LastUpdated
#>  1:                     23    2022-08-22 07:31:30
#>  2:                     26    2022-09-09 07:33:10
#>  3:                     78    2022-09-01 07:31:34
#>  4:                      9    2022-08-17 07:30:12
#>  5:                     22    2022-09-06 07:32:36
#>  6:                      6    2022-08-13 17:05:54
#>  7:                      6    2022-08-10 01:09:22
#>  8:                    206    2020-03-28 19:12:15
#>  9:                     78    2023-09-07 23:20:22
#> 10:                     12    2022-05-19 22:58:39
#> 11:                     27    2023-09-07 23:19:35
#> 12:                     35    2023-04-24 13:58:56
#> 13:                     50    2023-09-07 23:11:52
#> 14:                    102    2023-09-07 23:14:08
#> 15:                     22    2023-06-30 07:36:09
#> 16:                     46    2023-09-07 23:15:42
#> 17:                     93    2022-08-30 23:36:54
#> 18:                     23    2023-09-07 22:56:32
#> 19:                     88    2022-08-20 07:32:52
#> 20:                    105    2023-06-09 07:33:11
#>                                           experiment.batchEffect
#>  1:                                      NO_BATCH_EFFECT_SUCCESS
#>  2:  This data set may have a batch artifact (PC 3), p=0.0022570
#>  3: This data set may have a batch artifact (PC 1), p=8.2617e-08
#>  4:                                      BATCH_CORRECTED_SUCCESS
#>  5:                                      NO_BATCH_EFFECT_SUCCESS
#>  6:                                UNINFORMATIVE_HEADERS_FAILURE
#>  7:                                UNINFORMATIVE_HEADERS_FAILURE
#>  8: This data set may have a batch artifact (PC 1), p=4.7909e-27
#>  9:                                                NO_BATCH_INFO
#> 10:                                      NO_BATCH_EFFECT_SUCCESS
#> 11:                                                NO_BATCH_INFO
#> 12:                                                NO_BATCH_INFO
#> 13:                                                NO_BATCH_INFO
#> 14:                                                NO_BATCH_INFO
#> 15:                                         SINGLE_BATCH_SUCCESS
#> 16:                                                NO_BATCH_INFO
#> 17:  This data set may have a batch artifact (PC 3), p=0.0065166
#> 18:                                                         <NA>
#> 19: This data set may have a batch artifact (PC 5), p=2.0241e-08
#> 20:                                                NO_BATCH_INFO
#>     geeq.batchCorrected geeq.batchConfound geeq.batchEffect geeq.rawData
#>  1:                  NA                 NA               NA           NA
#>  2:                  NA                 NA               NA           NA
#>  3:                  NA                 NA               NA           NA
#>  4:                  NA                 NA               NA           NA
#>  5:                  NA                 NA               NA           NA
#>  6:                  NA                 NA               NA           NA
#>  7:                  NA                 NA               NA           NA
#>  8:                  NA                 NA               NA           NA
#>  9:                  NA                 NA               NA           NA
#> 10:                  NA                 NA               NA           NA
#> 11:                  NA                 NA               NA           NA
#> 12:                  NA                 NA               NA           NA
#> 13:                  NA                 NA               NA           NA
#> 14:                  NA                 NA               NA           NA
#> 15:                  NA                 NA               NA           NA
#> 16:                  NA                 NA               NA           NA
#> 17:                  NA                 NA               NA           NA
#> 18:                  NA                 NA               NA           NA
#> 19:                  NA                 NA               NA           NA
#> 20:                  NA                 NA               NA           NA
#>     geeq.qScore geeq.sScore taxon.Name taxon.Scientific taxon.ID taxon.NCBI
#>  1:          NA          NA      human     Homo sapiens        1       9606
#>  2:          NA          NA      mouse     Mus musculus        2      10090
#>  3:          NA          NA      mouse     Mus musculus        2      10090
#>  4:          NA          NA      mouse     Mus musculus        2      10090
#>  5:          NA          NA      mouse     Mus musculus        2      10090
#>  6:          NA          NA      mouse     Mus musculus        2      10090
#>  7:          NA          NA      mouse     Mus musculus        2      10090
#>  8:          NA          NA      human     Homo sapiens        1       9606
#>  9:          NA          NA      human     Homo sapiens        1       9606
#> 10:          NA          NA      human     Homo sapiens        1       9606
#> 11:          NA          NA      human     Homo sapiens        1       9606
#> 12:          NA          NA      human     Homo sapiens        1       9606
#> 13:          NA          NA      human     Homo sapiens        1       9606
#> 14:          NA          NA      human     Homo sapiens        1       9606
#> 15:          NA          NA      human     Homo sapiens        1       9606
#> 16:          NA          NA      human     Homo sapiens        1       9606
#> 17:          NA          NA      human     Homo sapiens        1       9606
#> 18:          NA          NA      human     Homo sapiens        1       9606
#> 19:          NA          NA      human     Homo sapiens        1       9606
#> 20:          NA          NA      human     Homo sapiens        1       9606
#>     taxon.Database.Name taxon.Database.ID
#>  1:                hg38                87
#>  2:                mm10                81
#>  3:                mm10                81
#>  4:                mm10                81
#>  5:                mm10                81
#>  6:                mm10                81
#>  7:                mm10                81
#>  8:                hg38                87
#>  9:                hg38                87
#> 10:                hg38                87
#> 11:                hg38                87
#> 12:                hg38                87
#> 13:                hg38                87
#> 14:                hg38                87
#> 15:                hg38                87
#> 16:                hg38                87
#> 17:                hg38                87
#> 18:                hg38                87
#> 19:                hg38                87
#> 20:                hg38                87