Dataset differential expression

Retrieves the differential expression resultSet(s) associated with the dataset. If there is more than one resultSet, use getDatasetResultSets() to see the options and get the ID you want. Alternatively, you can query the resultSet directly if you know its ID beforehand.

getDatasetDE(
  dataset = NA_character_,
  resultSet = NA_integer_,
  all = FALSE,
  memoised = getOption("gemma.memoise", FALSE)
)

Arguments

dataset

A dataset identifier.

resultSet

A resultSet identifier.

all

If TRUE, will download all differential expression resultSets for the dataset.

memoised

Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

Value

A data table with differential expression values. If there are multiple resultSets and all = TRUE, a list of data tables with differential expression values.

Examples

getDatasetDE("GSE2018")
#> Warning: Some columns are type 'integer64' but package bit64 is not installed. Those columns will print as strange looking floating point data. There is no need to reload the data. Simply install.packages('bit64') to obtain the integer64 print method and print the data again.
#>              Probe        NCBIid              GeneSymbol
#>     1:   205105_at          4124                  MAN2A1
#>     2:   217563_at                                      
#>     3:   206819_at                                      
#>     4:   219446_at         55188                   RIC8B
#>     5:   206630_at          7299                     TYR
#>    ---                                                  
#> 22267:   213691_at     101928378               PTOV1-AS2
#> 22268: 211485_s_at          8817                   FGF18
#> 22269: 220936_s_at         55766                    H2AJ
#> 22270: 205013_s_at 135|101730217 ADORA2A|SPECC1L-ADORA2A
#> 22271:   220392_at         64641                    EBF2
#>                                                                  GeneName
#>     1:                                mannosidase alpha class 2A member 1
#>     2:                                                                   
#>     3:                                                                   
#>     4:                          RIC8 guanine nucleotide exchange factor B
#>     5:                                                         tyrosinase
#>    ---                                                                   
#> 22267:                                              PTOV1 antisense RNA 2
#> 22268:                                        fibroblast growth factor 18
#> 22269:                                                      H2A.J histone
#> 22270: adenosine A2a receptor|SPECC1L-ADORA2A readthrough (NMD candidate)
#> 22271:                                         EBF transcription factor 2
#>          pvalue corrected_pvalue    rank
#>     1: 0.691200           0.8827 0.78310
#>     2: 0.023940           0.2607 0.09178
#>     3: 0.029350           0.2801 0.10470
#>     4: 0.680000           0.8786 0.77400
#>     5: 0.416300           0.7360 0.56560
#>    ---                                  
#> 22267: 0.160400           0.5186 0.30930
#> 22268: 0.154700           0.5131 0.30150
#> 22269: 0.375100           0.7092 0.52890
#> 22270: 0.373800           0.7085 0.52760
#> 22271: 0.001836           0.1074 0.01702
#>        contrast_Transplanted Organ Complication, Acute rejection_logFoldChange
#>     1:                                                                 0.07823
#>     2:                                                                 0.17390
#>     3:                                                                 0.14450
#>     4:                                                                -0.05527
#>     5:                                                                -0.06343
#>    ---                                                                        
#> 22267:                                                                -0.14890
#> 22268:                                                                -0.09667
#> 22269:                                                                 0.05851
#> 22270:                                                                 0.20720
#> 22271:                                                                -0.16590
#>        contrast_Transplanted Organ Complication, Acute rejection_tstat
#>     1:                                                          0.4009
#>     2:                                                          2.3750
#>     3:                                                          2.2840
#>     4:                                                         -0.4163
#>     5:                                                         -0.8240
#>    ---                                                                
#> 22267:                                                         -1.4380
#> 22268:                                                         -1.4590
#> 22269:                                                          0.8999
#> 22270:                                                          0.9023
#> 22271:                                                         -3.4070
#>        contrast_Transplanted Organ Complication, Acute rejection_pvalue
#>     1:                                                         0.691200
#>     2:                                                         0.023940
#>     3:                                                         0.029350
#>     4:                                                         0.680000
#>     5:                                                         0.416300
#>    ---                                                                 
#> 22267:                                                         0.160400
#> 22268:                                                         0.154700
#> 22269:                                                         0.375100
#> 22270:                                                         0.373800
#> 22271:                                                         0.001836