Retrieve annotations and surface level stats for a dataset's differential analyses

Retrieve annotations and surface level stats for a dataset's differential analyses

get_dataset_differential_expression_analyses(
  dataset,
  offset = 0L,
  limit = 20L,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset

A numerical dataset identifier or a dataset short name

offset

The offset of the first retrieved result.

limit

Optional, defaults to 20. Limits the result to specified amount of objects. Has a maximum value of 100. Use together with offset and the totalElements attribute in the output to compile all data if needed.

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

A data table with information about the differential expression analysis of the queried dataset. Note that this funciton does not return differential expression values themselves. Use get_differential_expression_values

to get differential expression values (see examples).

The fields of the output data.table are:

result.ID: Result set ID of the differential expression analysis. May represent multiple factors in a single model.

contrast.ID: Id of the specific contrast factor. Together with the result.ID they uniquely represent a given contrast.

experiment.ID: Id of the source experiment

baseline.category: Category for the contrast

baseline.categoryURI: URI for the baseline category

baseline.factors: Characteristics of the baseline. This field is a data.table

experimental.factors: Characteristics of the experimental group. This field is a data.table

subsetFactor.subset: TRUE if the result set belong to a subset, FALSE if not. Subsets are created when performing differential expression to avoid unhelpful comparisons.

subsetFactor.category: Category of the subset

subsetFactor: Characteristics of the subset. This field is a data.table

probes.Analyzed: Number of probesets represented in the contrast

genes.Analyzed: Number of genes represented in the contrast

Examples

result <- get_dataset_differential_expression_analyses("GSE2872")
get_differential_expression_values(resultSet = result$result.ID[1])
#> $`566186`
#>             Probe NCBIid GeneSymbol
#>     1: 1378140_at 364396      Arl11
#>     2: 1381500_at                  
#>     3: 1372405_at 308398    Ppp1r37
#>     4: 1377459_at 406866     Ly6g6e
#>     5: 1374753_at                  
#>    ---                             
#> 31069: 1375597_at                  
#> 31070: 1390122_at                  
#> 31071: 1371969_at                  
#> 31072: 1377916_at                  
#> 31073: 1378558_at                  
#>                                            GeneName pvalue corrected_pvalue
#>     1:       ADP-ribosylation factor like GTPase 11 0.7767           0.9999
#>     2:                                              0.8259           0.9999
#>     3: protein phosphatase 1, regulatory subunit 37 0.2884           0.9999
#>     4:       lymphocyte antigen 6 family member G6E 0.5870           0.9999
#>     5:                                              0.5591           0.9999
#>    ---                                                                     
#> 31069:                                              0.2151           0.9999
#> 31070:                                              0.8334           0.9999
#> 31071:                                              0.0816           0.9999
#> 31072:                                              0.4157           0.9999
#> 31073:                                              0.9913           0.9999
#>          rank contrast_134441_log2fc contrast_134441_tstat
#>     1: 0.7321               -0.05524               -0.2141
#>     2: 0.7919               -0.04937               -0.5725
#>     3: 0.2085                0.11750                0.8622
#>     4: 0.5140               -0.09026               -0.5499
#>     5: 0.4842               -0.03761               -0.1320
#>    ---                                                    
#> 31069: 0.1403               -0.50850               -1.8780
#> 31070: 0.8004                0.07260                0.4588
#> 31071: 0.0429               -0.05175               -0.3984
#> 31072: 0.3298                0.09150                0.5815
#> 31073: 0.9891                0.09560                0.6987
#>        contrast_134441_pvalue
#>     1:                 0.8344
#>     2:                 0.5783
#>     3:                 0.4067
#>     4:                 0.5932
#>     5:                 0.8974
#>    ---                       
#> 31069:                 0.0866
#> 31070:                 0.6552
#> 31071:                 0.6978
#> 31072:                 0.5724
#> 31073:                 0.4990
#>