Retrieve the differential analyses of a dataset

Retrieve the differential analyses of a dataset

get_dataset_differential_expression_analyses(
  dataset,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)

Arguments

dataset

A numerical dataset identifier or a dataset short name

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

attributes

If TRUE additional information from the call will be added into the output object's attributes such as offset and available elements.

Value

A data table with information about the differential expression analysis of the queried dataset. Note that this funciton does not return differential expression values themselves. Use get_differential_expression_values

to get differential expression values (see examples).

The fields of the output data.table are:

result.ID: Result set ID of the differential expression analysis. May represent multiple factors in a single model.

contrast.ID: Id of the specific contrast factor. Together with the result.ID they uniquely represent a given contrast.

experiment.ID: Id of the source experiment

baseline.category: Category for the contrast

baseline.categoryURI: URI for the baseline category

baseline.factorValue: Factor value assigned as the baseline in the contrast. Typically represent control samples

baseline.factorValueURI: URI for the baseline.factorValue

experimental.factorValue: Factor value assigned to the experimental group.

experimental.factorValueURI: URI for the experimental.factorValue

subsetFactor.subset: TRUE if the result set belong to a subset, FALSE if not. Subsets are created when performing differential expression to avoid unhelpful comparisons.

subsetFactor.category: Category of the subset

subsetFactor.categoryURI: URI of the subset category

subsetFactor.factorValue: Factor Value of the subset

subsetFactor.factorValueURI: URI of the subset factor value

probes.Analyzed: Number of probesets represented in the contrast

genes.Analyzed: Number of genes represented in the contrast

platform.ID: Platform id for the contrast

Examples

result <- get_dataset_differential_expression_analyses("GSE2018")
get_differential_expression_values(resultSet = result$result.ID)
#> $`525879`
#>              Probe NCBIid GeneSymbol
#>     1:   205680_at   4319      MMP10
#>     2:   202890_at   9053       MAP7
#>     3:   207753_at  57343     ZNF304
#>     4: 218612_s_at  10078      TSSC4
#>     5: 211883_x_at    634    CEACAM1
#>    ---                              
#> 22246: 210744_s_at   3568      IL5RA
#> 22247:   204693_at  11135   CDC42EP1
#> 22248: 205324_s_at  24140      FTSJ1
#> 22249: 200843_s_at   2058      EPRS1
#> 22250:    36936_at   7264       GFUS
#>                                             GeneName  pvalue corrected_pvalue
#>     1:                    matrix metallopeptidase 10 0.84060           0.9423
#>     2:              microtubule associated protein 7 0.64770           0.8647
#>     3:                       zinc finger protein 304 0.85960           0.9508
#>     4: tumor suppressing subtransferable candidate 4 0.12460           0.4740
#>     5:                  CEA cell adhesion molecule 1 0.80960           0.9306
#>    ---                                                                       
#> 22246:          interleukin 5 receptor subunit alpha 0.50000           0.7889
#> 22247:                      CDC42 effector protein 1 0.62570           0.8557
#> 22248:             FtsJ RNA 2'-O-methyltransferase 1 0.39720           0.7232
#> 22249:             glutamyl-prolyl-tRNA synthetase 1 0.51230           0.7959
#> 22250:                         GDP-L-fucose synthase 0.01474           0.2203
#>           rank contrast_1_log2fc contrast_1_tstat contrast_1_pvalue
#>     1: 0.89210          -0.01328          -0.2026           0.84060
#>     2: 0.74910           0.04313           0.4610           0.64770
#>     3: 0.90410           0.02816           0.1782           0.85960
#>     4: 0.26290          -0.12190          -1.5740           0.12460
#>     5: 0.86990           0.02672           0.2429           0.80960
#>    ---                                                             
#> 22246: 0.63360           0.06279           0.6818           0.50000
#> 22247: 0.73110           0.04890           0.4922           0.62570
#> 22248: 0.54920          -0.12670          -0.8574           0.39720
#> 22249: 0.64370          -0.05590          -0.6622           0.51230
#> 22250: 0.06688          -0.28510          -2.5690           0.01474
#>