Retrieve annotations and surface level stats for a dataset's differential analyses

Retrieve annotations and surface level stats for a dataset's differential analyses

get_dataset_differential_expression_analyses(
  dataset,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset

A numerical dataset identifier or a dataset short name

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

A data table with information about the differential expression analysis of the queried dataset. Note that this funciton does not return differential expression values themselves. Use get_differential_expression_values

to get differential expression values (see examples).

The fields of the output data.table are:

result.ID: Result set ID of the differential expression analysis. May represent multiple factors in a single model.

contrast.ID: Id of the specific contrast factor. Together with the result.ID they uniquely represent a given contrast.

experiment.ID: Id of the source experiment

factor.category: Category for the contrast

factor.category.URI: URI for the contrast category

factor.ID: ID of the factor

baseline.factors: Characteristics of the baseline. This field is a data.table

experimental.factors: Characteristics of the experimental group. This field is a data.table

isSubset: TRUE if the result set belong to a subset, FALSE if not. Subsets are created when performing differential expression to avoid unhelpful comparisons.

subsetFactor: Characteristics of the subset. This field is a data.table

probes.analyzed: Number of probesets represented in the contrast

genes.analyzed: Number of genes represented in the contrast

Examples

result <- get_dataset_differential_expression_analyses("GSE2872")
get_differential_expression_values(resultSet = result$result.ID[1])
#> $`570867`
#>               Probe NCBIid GeneSymbol
#>              <char> <char>     <char>
#>     1:   1368058_at  64196       Safb
#>     2:   1386270_at                  
#>     3:   1368851_at  24356       Ets1
#>     4:   1395939_at                  
#>     5:   1381999_at 296580    Camsap1
#>    ---                               
#> 31069: 1393379_a_at                  
#> 31070:   1377975_at                  
#> 31071:   1382534_at                  
#> 31072:   1389727_at 314848     Lrrc10
#> 31073:   1392428_at 360604    Spata20
#>                                                  GeneName pvalue
#>                                                    <char>  <num>
#>     1:                       scaffold attachment factor B 0.9716
#>     2:                                                    0.1520
#>     3:         ETS proto-oncogene 1, transcription factor 0.6364
#>     4:                                                    0.6581
#>     5: calmodulin regulated spectrin-associated protein 1 0.5268
#>    ---                                                          
#> 31069:                                                    0.4792
#> 31070:                                                    0.3199
#> 31071:                                                    0.5406
#> 31072:                  leucine-rich repeat-containing 10 0.5045
#> 31073:                      spermatogenesis associated 20 0.6297
#>        corrected_pvalue   rank contrast_134441_log2fc contrast_134441_tstat
#>                   <num>  <num>                  <num>                 <num>
#>     1:           0.9999 0.9655                 0.0434                0.3027
#>     2:           0.9999 0.0911                -0.0748               -0.9062
#>     3:           0.9999 0.5707                -0.0850               -0.5670
#>     4:           0.9999 0.5957                 0.0411                0.4143
#>     5:           0.9999 0.4488                 0.0158                0.0797
#>    ---                                                                     
#> 31069:           0.9999 0.3961                -0.0235               -0.1884
#> 31070:           0.9999 0.2388                 0.0924                0.6052
#> 31071:           0.9999 0.4637                -0.0054               -0.0561
#> 31072:           0.9999 0.4249                -0.0593               -0.5642
#> 31073:           0.9999 0.5629                 0.0828                0.5923
#>        contrast_134441_pvalue
#>                         <num>
#>     1:                 0.7676
#>     2:                 0.3839
#>     3:                 0.5819
#>     4:                 0.6865
#>     5:                 0.9379
#>    ---                       
#> 31069:                 0.8539
#> 31070:                 0.5572
#> 31071:                 0.9562
#> 31072:                 0.5838
#> 31073:                 0.5654
#>