Retrieve the expression data of a dataset

Retrieve the expression data of a dataset

get_dataset_expression(
  dataset,
  filter = FALSE,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)

Arguments

dataset

A numerical dataset identifier or a dataset short name

filter

The filtered version (filter = TRUE) corresponds to what is used in most Gemma analyses, removing some probes/elements. Unfiltered includes all elements.

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

attributes

If TRUE additional information from the call will be added into the output object's attributes such as offset and available elements.

Value

If raw is FALSE (default), a data table of the expression matrix for the queried dataset. If raw is TRUE, returns the binary file in raw form.

Examples

get_dataset_expression("GSE2018")
#>                         Probe          GeneSymbol
#>     1:              1007_s_at                DDR1
#>     2:                1053_at                RFC2
#>     3:                 117_at         HSPA7|HSPA6
#>     4:                 121_at                PAX8
#>     5:              1255_g_at LOC118142757|GUCA1A
#>    ---                                           
#> 22279: AFFX-r2-Hs28SrRNA-3_at                    
#> 22280: AFFX-r2-Hs28SrRNA-5_at                    
#> 22281: AFFX-r2-Hs28SrRNA-M_at                    
#> 22282:    AFFX-r2-P1-cre-3_at                    
#> 22283:    AFFX-r2-P1-cre-5_at                    
#>                                                                                                      GeneName
#>     1:                                                            discoidin domain receptor tyrosine kinase 1
#>     2:                                                                         replication factor C subunit 2
#>     3: heat shock protein family A (Hsp70) member 7 (pseudogene)|heat shock protein family A (Hsp70) member 6
#>     4:                                                                                           paired box 8
#>     5:                                             GUCA1ANB-GUCA1A readthrough|guanylate cyclase activator 1A
#>    ---                                                                                                       
#> 22279:                                                                                                       
#> 22280:                                                                                                       
#> 22281:                                                                                                       
#> 22282:                                                                                                       
#> 22283:                                                                                                       
#>                NCBIid BAL_13_A0B0 BAL_44b_A0B0 BAL_41_A1B0 BAL_31_A0B1
#>     1:            780    7.664396     7.555234    7.898749    7.482914
#>     2:           5982    6.818718     6.817860    6.428031    6.873047
#>     3:      3311|3310    7.702311     7.736164    8.299871    7.714629
#>     4:           7849    8.872670     8.728661    9.433338    9.106445
#>     5: 118142757|2978    4.023544     4.022874    4.266776    4.127223
#>    ---                                                                
#> 22279:                   5.269825     4.883055    6.307157    6.406686
#> 22280:                   4.032411     4.443605    4.487075    4.551892
#> 22281:                   5.033840     5.321434    5.462065    5.793712
#> 22282:                  12.634298    11.756798   11.946831   12.617424
#> 22283:                  12.589158    11.830940   11.781727   12.318279
#>        BAL_12_A1B0 BAL_15b_A0B0 BAL_40_A0B0 BAL_47b_A0B1 BAL_47c_A0B1
#>     1:    7.606950     7.675714    7.568318     7.605180     7.629114
#>     2:    7.081941     6.770301    6.754742     6.785765     6.910238
#>     3:    7.497553     8.433980    7.307702     7.800827     7.474227
#>     4:    9.130426     9.062922    9.126554     9.024894     9.173378
#>     5:    4.130228     3.991846    4.057203     4.179423     3.955497
#>    ---                                                               
#> 22279:    6.242087     5.699829    5.119069     6.658047     5.768701
#> 22280:    4.349244     4.661939    4.238217     4.462518     4.355739
#> 22281:    5.663534     5.803358    5.372956     5.477958     5.260586
#> 22282:   12.376739    11.258021   12.462382    12.871473    12.033990
#> 22283:   12.476188    11.643198   12.634298    12.909754    12.088908
#>        BAL_45a_A1B0 BAL_47a_A0B0 BAL_44a_A0B0 BAL_39_A0B1 BAL_33_A0B1
#>     1:     7.446346     7.787959     7.734734    7.270490    7.428734
#>     2:     6.803677     6.440673     6.940845    6.808251    7.139178
#>     3:     7.623731     7.844755     7.603600    7.328971    7.568634
#>     4:     9.293806     9.040855     9.248298    8.881946    8.946895
#>     5:     4.034630     3.909490     4.004146    4.018348    4.025862
#>    ---                                                               
#> 22279:     5.107271     5.371702     5.183981    5.413069    5.689970
#> 22280:     4.260343     4.276494     4.627294    4.337369    4.311341
#> 22281:     5.040337     5.166088     6.134356    5.647816    5.397744
#> 22282:    11.272546    12.462382    12.257533   12.728011   11.887863
#> 22283:    11.459026    12.656571    12.342556   12.855262   11.810861
#>        BAL_29_A1B0 BAL_25_A0B0 BAL_27a_A0B0 BAL_1_A0B0 BAL_15a_A0B1 BAL_43_A0B0
#>     1:    7.905805    7.565792     7.912942   7.520084     7.633544    7.333711
#>     2:    6.433760    6.887287     6.975750   7.158184     6.906474    6.966837
#>     3:    7.794066    7.444417     7.775969   7.843651     7.689818    7.462743
#>     4:    9.592714    8.923390     9.249823   9.057032     8.764082    9.090870
#>     5:    4.128827    3.979449     4.027976   4.168197     4.025454    4.002775
#>    ---                                                                         
#> 22279:    5.304832    4.986458     5.660599   5.754021     5.472515    5.713751
#> 22280:    4.163611    4.339306     4.354635   4.305040     4.360437    4.326695
#> 22281:    5.566520    5.449715     5.502226   5.673235     5.541409    5.575244
#> 22282:   12.742010   12.323225    12.363355  12.419389    12.462382   12.728011
#> 22283:   12.880875   12.294077    12.294077  12.401875    12.383058   12.708045
#>        BAL_42_A0B0 BAL_4_A0B0 BAL_3_A1B0 BAL_18_A1B0 BAL_17_A0B0 BAL_21_A0B0
#>     1:    7.893288   7.474848   7.543944    7.407762    7.229377    7.867851
#>     2:    6.649465   6.947905   6.895257    6.780074    6.963063    6.615844
#>     3:    8.579609   7.741327   7.542451    7.926313    7.972501    7.421375
#>     4:    8.930581   9.119985   8.873311    9.154426    8.806005    9.553362
#>     5:    4.113324   4.037982   3.861244    4.054710    4.024187    4.137057
#>    ---                                                                      
#> 22279:    5.882000   5.745562   6.067514    7.299089    6.076631    7.957031
#> 22280:    4.201591   4.332832   4.337072    4.395083    4.215431    5.204943
#> 22281:    5.634897   5.390216   5.559285    5.974870    5.704353    7.226784
#> 22282:   12.758545  12.582693  12.595888   12.841431   12.516955   12.589158
#> 22283:   12.728011  12.742010  12.750864   12.862965   12.449719   12.909754
#>        BAL_19_A0B0 BAL_45b_A2B0 BAL_26_A2B0 BAL_20a_A1B1 BAL_20b_A1B1
#>     1:    7.821547     7.782975    7.257932          NaN     7.680408
#>     2:    6.709687     6.593049    6.940845          NaN     6.650542
#>     3:    7.825478     7.932626    7.969855          NaN     7.738031
#>     4:    9.574804     8.999097    8.656306          NaN     9.239082
#>     5:    3.989322     3.944738    4.013794          NaN     4.030433
#>    ---                                                               
#> 22279:    8.356067     5.328197    4.948480          NaN     6.952453
#> 22280:    5.687296     4.119304    4.171280          NaN     4.250091
#> 22281:    7.929119     5.046239    5.167862          NaN     5.876946
#> 22282:   12.671734    12.123510   12.088908          NaN    12.624101
#> 22283:   12.909754    12.093898   12.211794          NaN    12.449719
#>        BAL_37_A1B1 BAL_46_A2B0 BAL_27b_A2B1
#>     1:    7.466471    7.932626     7.699664
#>     2:    6.858354    6.537636     6.845171
#>     3:    7.689818    7.809553     7.555798
#>     4:    9.055730   10.078527     9.141695
#>     5:    4.062315    3.859072     4.121438
#>    ---                                     
#> 22279:    5.377125    8.848175     5.789951
#> 22280:    4.253467    4.782150     4.440793
#> 22281:    5.369570    7.281428     5.875893
#> 22282:   11.975114   13.047083    12.443970
#> 22283:   12.088908   13.239051    12.671734