Retrieve processed expression data of a dataset

Retrieve processed expression data of a dataset

get_dataset_processed_expression(
  dataset,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

dataset

A numerical dataset identifier or a dataset short name

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

If raw is FALSE (default), a data table of the expression matrix for the queried dataset. If raw is TRUE, returns the binary file in raw form.

Examples

get_dataset_processed_expression("GSE2018")
#>                         Probe             GeneSymbol
#>                        <char>                 <char>
#>     1:              1007_s_at                   DDR1
#>     2:                1053_at                   RFC2
#>     3:                 117_at            HSPA7|HSPA6
#>     4:                 121_at                   PAX8
#>     5:              1255_g_at GUCA1ANB-GUCA1A|GUCA1A
#>    ---                                              
#> 22279: AFFX-r2-Hs28SrRNA-3_at                       
#> 22280: AFFX-r2-Hs28SrRNA-5_at                       
#> 22281: AFFX-r2-Hs28SrRNA-M_at                       
#> 22282:    AFFX-r2-P1-cre-3_at                       
#> 22283:    AFFX-r2-P1-cre-5_at                       
#>                                                                                                      GeneName
#>                                                                                                        <char>
#>     1:                                                            discoidin domain receptor tyrosine kinase 1
#>     2:                                                                         replication factor C subunit 2
#>     3: heat shock protein family A (Hsp70) member 7 (pseudogene)|heat shock protein family A (Hsp70) member 6
#>     4:                                                                                           paired box 8
#>     5:                                             GUCA1ANB-GUCA1A readthrough|guanylate cyclase activator 1A
#>    ---                                                                                                       
#> 22279:                                                                                                       
#> 22280:                                                                                                       
#> 22281:                                                                                                       
#> 22282:                                                                                                       
#> 22283:                                                                                                       
#>                NCBIid BAL_1_A0B0 BAL_27a_A0B0 BAL_43_A0B0 BAL_13_A0B0
#>                <char>      <num>        <num>       <num>       <num>
#>     1:            780     7.4623       7.8763      7.4546      7.5897
#>     2:           5982     7.0674       6.8416      6.9020      6.8889
#>     3:      3311|3310     7.8392       7.7334      7.6140      7.7068
#>     4:           7849     9.1243       9.3388      9.1342      9.0008
#>     5: 118142757|2978     4.1629       4.0015      3.9937      4.0622
#>    ---                                                               
#> 22279:                    6.0442       5.8953      6.0414      5.7437
#> 22280:                    4.3702       4.4491      4.3248      4.2274
#> 22281:                    5.8420       5.5735      5.6337      5.4054
#> 22282:                   12.4001      12.3097     12.7080     12.6076
#> 22283:                   12.4956      12.3234     12.7085     12.6159
#>        BAL_29_A1B0 BAL_44b_A0B0 BAL_18_A1B0 BAL_31_A0B1 BAL_45b_A2B0
#>              <num>        <num>       <num>       <num>        <num>
#>     1:      7.8408       7.5389      7.5057      7.4757       7.7217
#>     2:      6.5479       6.8723      6.7586      6.9175       6.6493
#>     3:      7.8061       7.6090      7.7618      7.5906       7.9382
#>     4:      9.4695       8.8174      9.2959      9.1201       9.1434
#>     5:      4.0921       3.9806      4.0721      4.0626       3.9763
#>    ---                                                              
#> 22279:      5.5061       5.0290      7.1344      6.3791       5.8032
#> 22280:      4.2469       4.4103      4.5497      4.5338       4.2476
#> 22281:      5.7107       5.4534      6.0937      5.9405       5.4252
#> 22282:     12.5784      12.0046     12.6783     12.7013      12.0769
#> 22283:     12.7121      12.1745     12.7071     12.5966      12.0967
#>        BAL_20a_A1B1 BAL_39_A0B1 BAL_47a_A0B0 BAL_37_A1B1 BAL_21_A0B0
#>              <lgcl>       <num>        <num>       <num>       <num>
#>     1:           NA      7.3643       7.7570      7.6151      7.6638
#>     2:           NA      6.8158       6.5584      6.7764      6.7385
#>     3:           NA      7.4394       7.8905      7.8420      7.5918
#>     4:           NA      8.9355       9.0757      9.0833      9.1553
#>     5:           NA      4.0417       3.9452      4.0728      4.1046
#>    ---                                                              
#> 22279:           NA      5.6036       5.5698      5.5831      6.2889
#> 22280:           NA      4.3762       4.3247      4.2902      4.5706
#> 22281:           NA      5.6782       5.3193      5.4597      5.9314
#> 22282:           NA     12.6380      12.4291     11.9554     12.2941
#> 22283:           NA     12.7395      12.5967     12.0389     12.4245
#>        BAL_40_A0B0 BAL_47c_A0B1 BAL_15b_A0B0 BAL_26_A2B0 BAL_42_A0B0
#>              <num>        <num>        <num>       <num>       <num>
#>     1:      7.6959       7.6721       7.6584      7.3875      7.8968
#>     2:      6.7033       6.8924       6.7241      6.9261      6.6162
#>     3:      7.4063       7.5564       8.0925      8.0984      8.2502
#>     4:      9.2632       9.1023       9.0725      8.8144      9.0405
#>     5:      4.0604       3.9816       3.9857      4.0082      4.1257
#>    ---                                                              
#> 22279:      5.7738       6.0573       5.6573      5.5300      5.6857
#> 22280:      4.2906       4.4464       4.6204      4.2650      4.3131
#> 22281:      5.5689       5.5811       5.8045      5.3966      5.7005
#> 22282:     12.5281      12.3181      11.8423     12.0778     12.5757
#> 22283:     12.6254      12.3710      12.0150     12.1340     12.5472
#>        BAL_12_A1B0 BAL_27b_A2B1 BAL_44a_A0B0 BAL_17_A0B0 BAL_3_A1B0
#>              <num>        <num>        <num>       <num>      <num>
#>     1:      7.5837       7.5376       7.7166      7.3620     7.6154
#>     2:      7.0007       6.8936       6.9086      6.9581     6.8842
#>     3:      7.5155       7.7496       7.6796      7.7964     7.4748
#>     4:      9.1463       8.9047       9.2588      8.8983     8.9751
#>     5:      4.1134       4.0868       4.0291      4.0442     3.8952
#>    ---                                                             
#> 22279:      6.2286       5.0878       5.4162      5.8847     5.8753
#> 22280:      4.3363       4.0392       4.6214      4.3300     4.4788
#> 22281:      5.6526       5.1224       6.0406      5.7808     5.6130
#> 22282:     12.4795      12.1495      12.2680     12.2767    12.3744
#> 22283:     12.5716      12.1941      12.3711     12.2172    12.5742
#>        BAL_45a_A1B0 BAL_19_A0B0 BAL_33_A0B1 BAL_46_A2B0 BAL_20b_A1B1 BAL_4_A0B0
#>               <num>       <num>       <num>       <num>        <num>      <num>
#>     1:       7.4853      7.6282      7.4844      7.7165       7.6784     7.5597
#>     2:       6.7775      6.8080      7.0474      6.6655       6.6880     6.9416
#>     3:       7.7160      7.8906      7.6490      7.9372       7.8253     7.6235
#>     4:       9.2323      9.1685      8.9928      9.4814       9.2515     9.2566
#>     5:       4.0628      4.0098      4.0483      3.9185       4.0499     4.0568
#>    ---                                                                         
#> 22279:       5.3838      6.5142      5.8302      6.8142       6.8727     5.5462
#> 22280:       4.3461      4.8138      4.3541      4.2114       4.2569     4.4736
#> 22281:       5.3784      6.2741      5.4919      5.8082       5.7742     5.4175
#> 22282:      11.8487     12.3542     11.9772     12.5889      12.5257    12.3581
#> 22283:      11.9828     12.4245     11.9890     12.6001      12.3887    12.5638
#>        BAL_47b_A0B1 BAL_15a_A0B1 BAL_25_A0B0 BAL_41_A1B0
#>               <num>        <num>       <num>       <num>
#>     1:       7.5602       7.5819      7.5234      7.8389
#>     2:       6.8371       6.7558      6.9205      6.5528
#>     3:       7.8130       7.5988      7.4471      8.0923
#>     4:       9.0188       8.7984      8.9383      9.3820
#>     5:       4.1360       3.9986      3.9627      4.1723
#>    ---                                                  
#> 22279:       6.6788       5.5957      5.2302      6.2910
#> 22280:       4.5392       4.4584      4.4187      4.4599
#> 22281:       5.6101       5.6350      5.5780      5.5984
#> 22282:      12.7032      12.4696     12.1750     12.1585
#> 22283:      12.7371      12.4656     12.2041     12.1318