Retrieve the GO terms associated to a gene

Retrieve the GO terms associated to a gene

get_gene_go_terms(
  gene,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

gene

An ensembl gene identifier which typically starts with ensg or an ncbi gene identifier or an official gene symbol approved by hgnc

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be the raw endpoint from the API, likely a JSON or a gzip file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

A data table with information about the GO terms assigned to the queried gene. A list if raw = TRUE. A 404 error if the given identifier does not map to any object.

The fields of the output data.table are:

term.Name: Name of the term

term.ID: ID of the term

term.URI: URI of the term

Examples

get_gene_go_terms("DYRK1A")
#>                                                                         term.name
#>                                                                            <char>
#>   1:                                                                      nucleus
#>   2:                                             regulation of cell communication
#>   3:                                regulation of cellular component organization
#>   4:                                         regulation of chromatin organization
#>   5:                                                                anion binding
#>  ---                                                                             
#> 220:                                             peptidyl-amino acid modification
#> 221:                                                                neurofilament
#> 222: regulation of DNA damage response, signal transduction by p53 class mediator
#> 223:                                                    primary metabolic process
#> 224:                                                    ribonucleoprotein complex
#>         term.ID                                  term.URI
#>          <char>                                    <char>
#>   1: GO:0005634 http://purl.obolibrary.org/obo/GO_0005634
#>   2: GO:0010646 http://purl.obolibrary.org/obo/GO_0010646
#>   3: GO:0051128 http://purl.obolibrary.org/obo/GO_0051128
#>   4: GO:1902275 http://purl.obolibrary.org/obo/GO_1902275
#>   5: GO:0043168 http://purl.obolibrary.org/obo/GO_0043168
#>  ---                                                     
#> 220: GO:0018193 http://purl.obolibrary.org/obo/GO_0018193
#> 221: GO:0005883 http://purl.obolibrary.org/obo/GO_0005883
#> 222: GO:0043516 http://purl.obolibrary.org/obo/GO_0043516
#> 223: GO:0044238 http://purl.obolibrary.org/obo/GO_0044238
#> 224: GO:1990904 http://purl.obolibrary.org/obo/GO_1990904