Retrieve Platform Annotations by Gemma

Gets Gemma's platform annotations including mappings of microarray probes to genes.

get_platform_annotations(
  platform,
  annotType = c("noParents", "allParents", "bioProcess"),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  unzip = FALSE
)

Arguments

platform

A platform numerical identifiers or platform short name.

annotType

Which GO terms should the output include

file

Where to save the annotation file to, or empty to just load into memory

overwrite

Whether or not to overwrite an existing file

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

unzip

Whether or not to unzip the file (if @param file is not empty)

Value

A table of annotations

ElementName: Probeset names provided by the platform. Gene symbols for generic annotations typicall used for RNA-seq experiments.

GeneSymbols: Genes that were found to be aligned to the probe sequence. Note that it is possible for probes to be non-specific. Alignment to multiple genes are indicated with gene symbols separated by "|"s

GeneNames: Name of the gene

GOTerms: GO Terms associated with the genes. annotType argument can be used to choose which terms should be included.

GemmaIDs and NCBIids: respective IDs for the genes.

Examples

head(get_platform_annotations("GPL96"))
#>    ElementName GeneSymbols
#>         <char>      <char>
#> 1: 202619_s_at       PLOD2
#> 2: 219952_s_at      MCOLN1
#> 3:   205192_at     MAP3K14
#> 4:   212724_at        RND3
#> 5:   207373_at      HOXD10
#> 6: 204991_s_at         NF2
#>                                                   GeneNames
#>                                                      <char>
#> 1:        procollagen-lysine,2-oxoglutarate 5-dioxygenase 2
#> 2:                           mucolipin TRP cation channel 1
#> 3:        mitogen-activated protein kinase kinase kinase 14
#> 4:                                      Rho family GTPase 3
#> 5:                                             homeobox D10
#> 6: NF2, moesin-ezrin-radixin like (MERLIN) tumor suppressor
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         GOTerms
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          <char>
#> 1:                                                                                                                                                                                                                                                                                                                                                                                                                                                                               GO:0005506|GO:0001666|GO:0030867|GO:0031418|GO:0005789|GO:0070062|GO:0033823|GO:0005783|GO:0008475|GO:0006493|GO:0046947|GO:0017185|GO:0036211
#> 2:                                                                                                                                                                                            GO:0006812|GO:0051209|GO:0005764|GO:0005765|GO:0005886|GO:0097682|GO:0042995|GO:0005261|GO:0005262|GO:0002250|GO:1901660|GO:0019722|GO:0070588|GO:0099604|GO:0051289|GO:0072345|GO:0033572|GO:0071277|GO:0005515|GO:0031902|GO:0090382|GO:0005654|GO:0097352|GO:0010008|GO:0043235|GO:0001891|GO:0005794|GO:0043231|GO:0008289|GO:0005770|GO:1905673|GO:0042802|GO:0005381|GO:0034755|GO:0030670|GO:0071467|GO:0016020|GO:0051179
#> 3:                                                                                                                                                                                                                                                                                                                                                                                                                                   GO:0106310|GO:0005515|GO:0000165|GO:0001650|GO:0005524|GO:0006955|GO:0071260|GO:0005654|GO:0007249|GO:0038061|GO:0004672|GO:0004674|GO:0051607|GO:0043231|GO:0004709|GO:0016310|GO:0005829
#> 4:                                                                                                                                                                                                                                                                                                                                                                                                                                                         GO:0005925|GO:0005515|GO:0032956|GO:0000139|GO:0005525|GO:0005886|GO:0007015|GO:0007155|GO:0007165|GO:0007264|GO:0019901|GO:0030036|GO:0016477|GO:0005829|GO:0003924
#> 5:                                                                                                                                                                                                                                                                                                                     GO:0001228|GO:0030326|GO:0048935|GO:0003682|GO:0006357|GO:0045786|GO:0035136|GO:0035137|GO:0045944|GO:0005829|GO:0000978|GO:0036464|GO:0007519|GO:0005515|GO:0005634|GO:0021520|GO:0005654|GO:0007338|GO:0000981|GO:0000785|GO:0048704|GO:0050905|GO:0008344|GO:0009952|GO:0009954|GO:1990837|GO:0017053
#> 6: GO:0005769|GO:0005886|GO:1902966|GO:0042475|GO:0043005|GO:0044297|GO:0008285|GO:0008360|GO:0043409|GO:2000177|GO:0070306|GO:0032587|GO:0035330|GO:0005737|GO:0005856|GO:0031527|GO:0000165|GO:0031647|GO:0005730|GO:0005178|GO:0007398|GO:0045597|GO:0030036|GO:0021766|GO:0016020|GO:0048471|GO:0051726|GO:0030027|GO:0030864|GO:0042532|GO:0001707|GO:0005829|GO:0014013|GO:0022408|GO:0014010|GO:0005515|GO:0001953|GO:0005912|GO:0003779|GO:0005634|GO:0030336|GO:0045216|GO:0010626|GO:0042981|GO:0046426|GO:0072091|GO:1902115|GO:0045177|GO:1900180|GO:2000643|GO:0033689|GO:0033687|GO:0030175|GO:0032154|GO:0051496
#>    GemmaIDs NCBIids      EnsemblIds
#>      <char>  <char>          <char>
#> 1:   127276    5352 ENSG00000152952
#> 2:   313978   57192 ENSG00000090674
#> 3:   188876    9020 ENSG00000006062
#> 4:     9353     390 ENSG00000115963
#> 5:    72301    3236 ENSG00000128710
#> 6:   116071    4771 ENSG00000186575
head(get_platform_annotations('Generic_human_ncbiIds'))
#>    ElementName  GeneSymbols                              GeneNames GOTerms
#>          <int>       <char>                                 <char>  <char>
#> 1:   105373804 LOC105373804           uncharacterized LOC105373804        
#> 2:      285540  SEPSECS-AS1 SEPSECS antisense RNA 1 (head to head)        
#> 3:   105373805    STAT4-AS1                  STAT4 antisense RNA 1        
#> 4:   107985219 LOC107985219           uncharacterized LOC107985219        
#> 5:   105373802                                                            
#> 6:   107985216 LOC107985216           uncharacterized LOC107985216        
#>    GemmaIDs   NCBIids      EnsemblIds
#>       <int>     <int>          <char>
#> 1:  9236458 105373804                
#> 2:   421333    285540                
#> 3:  9236463 105373805 ENSG00000231858
#> 4:  9667149 107985219                
#> 5:       NA        NA                
#> 6:  9667148 107985216