Retrieve Platform Annotations by Gemma

Gets Gemma's platform annotations including mappings of microarray probes to genes.

get_platform_annotations(
  platform,
  annotType = c("noParents", "allParents", "bioProcess"),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  unzip = FALSE
)

Arguments

platform

A platform identifier @seealso getPlatforms

annotType

Which GO terms should the output include

file

Where to save the annotation file to, or empty to just load into memory

overwrite

Whether or not to overwrite an existing file

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

unzip

Whether or not to unzip the file (if @param file is not empty)

Value

A table of annotations

ProbeName: Probeset names provided by the platform. Gene symbols for generic annotations

GeneSymbols: Genes that were found to be aligned to the probe sequence. Note that it is possible for probes to be non-specific. Alignment to multiple genes are indicated with gene symbols separated by "|"s

GeneNames: Name of the gene

GOTerms: GO Terms associated with the genes. annotType argument can be used to choose which terms should be included.

GemmaIDs and NCBIids: respective IDs for the genes.

Examples

head(get_platform_annotations("GPL96"))
#>      ProbeName   GeneSymbols                                     GeneNames
#> 1: 211750_x_at TUBA1C|TUBA1A             tubulin alpha 1c|tubulin alpha 1a
#> 2:   216678_at                                                            
#> 3:   216345_at        ZSWIM8            zinc finger SWIM-type containing 8
#> 4:   207273_at                                                            
#> 5: 216025_x_at        CYP2C9 cytochrome P450 family 2 subfamily C member 9
#> 6: 218191_s_at        LMBRD1                     LMBR1 domain containing 1
#>                                                                                                                                                                                                                                                                                                                                                                       GOTerms
#> 1:                                                                                                                          GO:0005737|GO:0005525|GO:0051301|GO:0005515|GO:0000226|GO:0005634|GO:0007017|GO:0000278|GO:0005874|GO:0005200|GO:0005198|GO:0005881|GO:0030705|GO:0015630|GO:0031982|GO:0070062|GO:0050807|GO:0036464|GO:0042802|GO:0055037|GO:0005829|GO:0031594
#> 2:                                                                                                                                                                                                                                                                                                                                                                           
#> 3:                                                                                                                                                                                                                                                                                    GO:0008270|GO:0043161|GO:0016567|GO:0031463|GO:0005829|GO:0031462|GO:1990756|GO:2000627
#> 4:                                                                                                                                                                                                                                                                                                                                                                           
#> 5: GO:0005737|GO:0052741|GO:0008405|GO:0008404|GO:0020037|GO:0008203|GO:0008202|GO:0043231|GO:0008392|GO:0034875|GO:0070330|GO:0019373|GO:0006805|GO:0005506|GO:0005789|GO:0097267|GO:0005886|GO:0032787|GO:0042178|GO:0004497|GO:0046456|GO:0008210|GO:0008395|GO:0016712|GO:0019627|GO:0042759|GO:0006082|GO:0070989|GO:0043603|GO:0018676|GO:0018675|GO:0016491|GO:0016098
#> 6:                                                                                                                                                GO:0031419|GO:0072583|GO:0005789|GO:0005515|GO:0005886|GO:0005765|GO:0030136|GO:0005774|GO:0005158|GO:0007369|GO:0045334|GO:0061462|GO:0140318|GO:0043231|GO:0038016|GO:0032050|GO:0035612|GO:0072665|GO:0016020|GO:0016021
#>         GemmaIDs    NCBIids
#> 1: 360802|172797 84790|7846
#> 2:                         
#> 3:        235733      23053
#> 4:                         
#> 5:         32964       1559
#> 6:        303717      55788
head(get_platform_annotations('Generic_human'))
#>       ProbeName  GeneSymbols
#> 1:        LCN10        LCN10
#> 2:     STAG3L5P     STAG3L5P
#> 3: LOC101059976 LOC101059976
#> 4:         GAB3         GAB3
#> 5: LOC100287155 LOC100287155
#> 6:       RASSF2       RASSF2
#>                                                                  GeneNames
#> 1:                                                            lipocalin 10
#> 2:                                     stromal antigen 3-like 5 pseudogene
#> 3: arf-GAP with GTPase, ANK repeat and PH domain-containing protein 2-like
#> 4:                                       GRB2 associated binding protein 3
#> 5:                                       hypothetical protein LOC100287155
#> 6:                                  Ras association domain family member 2
#>                                                                                                                                                                                                                                                                                                     GOTerms
#> 1:                                                                                                                                                                                                                                                                                    GO:0005576|GO:0036094
#> 2:                                                                                                                                                                                                                                                                                                         
#> 3:                                                                                                                                                                                                                                                                                                         
#> 4:                                                                                                                                                                                                                                                              GO:0005737|GO:0030225|GO:0035591|GO:0007165
#> 5:                                                                                                                                                                                                                                                                                                         
#> 6: GO:0045670|GO:0005737|GO:0005515|GO:0005634|GO:0050821|GO:0005654|GO:0005794|GO:1901222|GO:1901223|GO:0046849|GO:0032991|GO:0048872|GO:0046330|GO:0001501|GO:0000776|GO:0005886|GO:0031954|GO:0007049|GO:0045667|GO:0004672|GO:0043065|GO:0045860|GO:0007165|GO:0033137|GO:0001503|GO:0005829|GO:0038168
#>    GemmaIDs   NCBIids
#> 1:   441399    414332
#> 2:  8799043 101735302
#> 3:  8779607 101059976
#> 4:   389635    139716
#> 5:  8090381 100287155
#> 6:   201914      9770