Retrieve all experiments within a given platform

Retrieve all experiments within a given platform

get_platform_datasets(
  platform,
  offset = 0L,
  limit = 20L,
  raw = getOption("gemma.raw", FALSE),
  memoised = getOption("gemma.memoised", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE),
  attributes = getOption("gemma.attributes", TRUE)
)

Arguments

platform

A platform numerical identifier or a platform short name

offset

The offset of the first retrieved result.

limit

Optional, defaults to 20. Limits the result to specified amount of objects. Has a maximum value of 100. Use together with offset and the totalElements attribute in the output to compile all data if needed.

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing. Raw results usually contain additional fields and flags that are omitted in the parsed results.

memoised

Whether or not to save to cache for future calls with the same inputs and use the result saved in cache if a result is already saved. Doing options(gemma.memoised = TRUE) will ensure that the cache is always used. Use forget_gemma_memoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

attributes

If TRUE additional information from the call will be added into the output object's attributes such as offset and available elements.

Value

A data table with information about the queried dataset(s). A list if raw = TRUE. Returns an empty list if no datasets matched. A successful response may contain 'Geeq' information, which aims to provide a unified metric to measure experiments by the quality of their data, and their suitability for use in Gemma. You can read more about the geeq properties here.

The fields of the output data.table are:

experiment.ShortName: Shortname given to the dataset within Gemma. Often corresponds to accession ID

experiment.Name: Full title of the dataset

experiment.ID: Internal ID of the dataset.

experiment.Description: Description of the dataset

experiment.Troubled: Did an automatic process within gemma or a curator mark the dataset as "troubled"

experiment.Accession: Accession ID of the dataset in the external database it was taken from

experiment.Database: The name of the database where the dataset was taken from

experiment.URI: URI of the original database

experiment.SampleCount: Number of samples in the dataset

experiment.batchEffect: A text field describing whether the dataset has batch effects

experiment.batchCorrected: Whether batch correction has been performed on the dataset. This is a text field with possible details about the batch correction. Use geeq.batchEffect if a more structured field is needed.

geeq.batchConfound: 0 if batch info isn't available, -1 if batch counfoud is detected, 1 if batch information is available and no batch confound found

geeq.batchEffect: -1 if batch p value < 0.0001, 1 if batch p value > 0.1, 0 if otherwise and when there is no batch information is available or when the data is confounded with batches.

geeq.rawData: -1 if no raw data available, 1 if raw data was available. When available, Gemma reprocesses raw data to get expression values and batches

geeq.qScore: Data quality score given to the dataset by Gemma.

geeq.sScore: Suitability score given to the dataset by Gemma. Refers to factors like batches, platforms and other aspects of experimental design

taxon.Name: Name of the species

taxon.Scientific: Scientific name for the taxon

taxon.ID: Internal identifier given to the species by Gemma

taxon.NCBI: NCBI ID of the taxon

taxon.Database.Name: Underlying database used in Gemma for the taxon

taxon.Database.ID: ID of the underyling database used in Gemma for the taxon

Examples

head(get_platform_datasets("GPL1355"))
#>    experiment.ShortName
#> 1:              GSE2872
#> 2:              GSE2870
#> 3:              GSE4037
#> 4:              GSE3110
#> 5:              GSE3125
#> 6:              GSE4753
#>                                                                                                           experiment.Name
#> 1:                                                                                                  d'mel-affy-rat-168311
#> 2:                                                                                                   Ogle-5P01NS037520-05
#> 3:                                                                                                  prosc-affy-rat-185420
#> 4: Comprehensive description of the transcriptome of hypothalamo-neurohypophyseal system in euhydrated and dehydrated rat
#> 5:              Comprehensive description of the transcriptome of neurointermediate lobe in euhydrated and dehydrated rat
#> 6:                                                                                                  wilso-affy-rat-132990
#>    experiment.ID
#> 1:             2
#> 2:           199
#> 3:           289
#> 4:           294
#> 5:           295
#> 6:           342
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                experiment.Description
#> 1:                                                                                                                                                                                                                                                                                           Neurological diseases disrupt the quality of the lives of patients and often lead to their death prematurely.  A common feature of most neurological diseases is the degeneration of neurons, which results from an inappropriate activation of apoptosis. Drugs that inhibit neuronal apoptosis could thus be candidates for therapeutic intervention in neurodegenerative disorders.  We have identified (and recently reported) a chemical called GW5074 that inhibits apoptosis in a variety of cell culture paradigms of neuronal apoptosis.  Additionally, we have found that GW5074 reduces neurodegeneration and improves behavioral outcome in a mouse model of Huntington's disease.  Although GW5074 is a c-Raf inhibitor, we know very little about the molecular mechanisms underlying its neuroprotective action.  Identifying genes that are regulated by GW5074 in neurons will shed insight into this issue. We believe that neuroprotection by GW5074 involves the regulation of several genes.  Some of these genes are likely to be induced whereas the expression of other genes might be inhibited.  The specific aim is to identify genes that are differentially expressed in neurons treated with GW5074. We believe that neuroprotection by GW5074 involves the regulation of several genes.  Some of these genes are likely to be induced whereas the expression of other genes might be inhibited. Cultures of cerebellar granule neurons undergo apoptosis when switched from medium containing levated levels of potassium (high K+ or HK) to medium containing low potassium (LK).  Although cell death begins at about 18 h, we have found that the by 6 h after LK treatment these neurons are irreversibly committed to cell death.  We will treat cerebellar granule neuron cultures with LK medium (which induces them to undergo apoptosis) or with GW5074 (1 uM).  We will extract RNA at two time-points after treatment: 3 h and 6 h. Analysis at the two different time-points will show us whether changes in expression of specific genes is transient or sustained or whether the changes occurs early or relatively late in the process.  Neuronal cultures will be prepared from 1 week old Wistar rats.  The cultures will be maintained in culture for 1 week before treatment.  Following treatment the cells will be lysed and total RNA isolated.  The RNA will be stored at -80oC and shipped to the microarray facility for analysis.  The experiment will be done in triplicate.  Thus, for each time-point (3 or 6h treatment) we will be hope to provide 3 sets of samples (each set coming from a different culture preparation and containing lysates from cells treated with LK or GW5074).  Having samples from 3 independent cultures will mitigate any expression differences resulting from subtle variations in culture quality or in the preparation or quality of RNA.\nDate GSE2872 Last Updated: Jul 06 2005\nContributors:  S R D'Mello\nIncludes GDS1837.\n Update date: Jun 09 2006.\n Dataset description GDS1837: Analysis of cultured cerebellar granule neurons in low potassium medium 3 and 6 hours following treatment with the c-Raf inhibitor GW5074. GW5074 blocks low potassium induced cell death. Results provide insight into the neuroprotective action of GW5074.
#> 2:                                                                                                                                              Prolonged stress has adverse consequences for neurons in the hippocampus (HIPP). The amygdala (AMY) is a brain region that is similar to HIPP across many measures, suggesting that chronic stress might modulate AMY and HIPP function in similar ways. However, studies addressing this issue have produced surprising results. For example, a regimen of chronic stress shown to produce atrophy in HIPP neurons caused dendritic branching in AMY neurons. These and other data have revealed that excessive stress induces fundamentally opposing processes in the AMY and HIPP, such that AMY function is facilitated by levels of stress that produce deleterious effects in HIPP. The mechanisms that contribute to the opposing responses of these brain regions to chronic stress are unknown. Such knowledge may suggest novel directions for pharmacological interventions to protect the HIPP from stress-mediated damage. We will determine gene expression patterns in the hippocampus and amygdala under basal and stressful conditions, with a particular interest in those genes that are differentially regulated across the two regions. We hypothesize that genes that are oppositely regulated in the amygdala and hippocampus after stress mediate the opposing functional consequences of chronic stress in these two brain regions.  Moreover, genes which are differentially expressed in these regions under basal conditions may underlie the distinct responses of these regions to stress. Rats will receive either 30sec of handling (Control group; n=15) or 3hr of immobilization stress (Stress group; n=15) each day for 14 consecutive days.  This stress paradigm has previously been shown to produce opposing effects on dendritic morphology and region-dependent behaviors for  the hippocampus versus amygdala.  Twenty-four hours after the last stress or handling session, the rats will be overanaesthetized, and the brain will be removed and placed in ice-cold buffer.  Tissue will be rapidly dissected from the basolateral complex of the amygdala and the CA3 region of the hippocampus of both hemispheres, flash frozen in liquid nitrogen, and stored at -80C until further processing.  The tissue from each brain region will be pooled across groups, yielding four samples (Control-Hippocampus, Control-Amygdala, Stress-Hippocampus, and Stress-Amygdala).  Total RNA will be extracted using standard methods, and sent to the centers.  Each sample will be run in triplicate, utilizing a total of 12 Affymetrix Rat Genome U34A gene chips.  A triplicate array assay of pooled samples has previously been shown to both substantially increase the sensitivity of detecting genes of interest and reduce the variability associated with individual microarrays.  In addition, the large number of rats used in each group will reduce the variability in gene expression associated with individual responses to chronic stress, as well as variability due to slight anatomical differences in the tissue extracted from each rat.\nLast Updated (by provider): Jul 06 2005\nContributors:  W Ogle\nIncludes GDS1794.\n Update date: Jun 08 2006.\n Dataset description GDS1794: Analysis of the amygdalae (Amy) and hippocampi (Hipp) of animals subjected to chronic stress induced by daily handling sessions for 14 days. Studies indicate that chronic stress produce a response in Hipp that is opposite of that in Amy.
#> 3: The goal of the research proposed herein is to identify (i) gene products that allow identification of specific precursor populations in vivo and (ii) genetic programs that regulate the generation of lineaege-restricted precursor (LRP) cells from neuroepithelial stem cells (NSC).  As these cells play a major role in development and tissue repair, this effort will enhance both basic and clinically relevant research. Note: Funding for this micro-array analysis was granted through NINDS as supplement 3R01NS042820-02S1. Our ability to generate essentially unlimited numbers of purified rodent NSCs, thier restricted decendants leading the generation of macroglia - and also to control the generation of each of these cell types from its immediate ancestor - provides us with a unique system for studying the multi-step nature of differentiation. Using defined conditions, samples will be isolated from each type of purified cell population. Expression profiles will be compared to reveal cell type specific markers and regulatory factors. The generation of differentiated cell types from neuroepithelial stem cells appears to require progression through a sequence of lineaege-restricted intermediate cell stages.  In the case of at least one of the pathways involved in oliigodendrocyte generation, it now appears that all of the stages of this sequence have been defined and all of the intermediate cells have been isolated (as detailed in the Noble, M. et al. (2004)  Dev. Biol. 265, 33-52).  Studies in vitro, and consequent to cell transplantation, have enabled us to gain an increasingly detailed understanding of the differentiation potential of these cells. The microarray analysis proposed herein will provide the specific markers that will enable the LRP cells to be unambiguously identified in vivo, thereby setting the stage to develop a detailed mapping of the role of these cells during development. RNA will be isolated from purified cell populations.  All populations will initially be derived from spinal cord.  Preparation of purified populations of each of these cell types has been described in multiple of our previous publications (Mayer, M. et al. (1994) Development 120, 142-153; Mayer-Pröschel, M. et al. (1997) Neuron 19, 773-785; Rao, M. and Mayer-Pröschel, M. (1997) Dev. Biology 188, 48-63). Total RNA from purified cell cultures will be extracted in a two step process. RNA will first be extracted by dounce homogenization in phenol/guanidine isothiocyonate solution (Trizol, Invitrogen, California) followed by isopropanol precipitation. Contaminating traces of DNA and protein will then be removed by silica-based affinity chromatography and Dnase treatment (NucleoSpin RNA Purification system, Clontech, California). Depending on the cell type, 5x105 cells will typically yield 5-15µg of high quality RNA (OD260/280=2.1 at pH7.5). In order to determine purity and molecular weight distribution, samples will be routinely tested on a Agilent Bioanalyzer. cDNA will be generated from total RNA and then be submitted to one round of amplification and labeling using the messageAmp system (Ambion). Probes are then hybridized to the Affymerix Rat 230 (A/B or 2.0) GeneChip expression array. Array data will be processed using the Micro Array Suite 5.0 (MAS, Affymetrix). Cel files containing normalized, relative expression data for all genes tested will be used for downstream statistical analysis.\nLast Updated (by provider): Jan 13 2006\nContributors:  Christoph Proschel
#> 4:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      The hypothalamic paraventricular (PVN) and supraoptic (SON) nuclei  are important integrative structures that regulate co-ordinated  responses to perturbations in cardiovascular homeostasis. Through  descending projections from parvocellular neurons to the brainstem,  the PVN acts as an autonomic 'premotor nucleus', regulating reflex  changes in sympathetic nerve activity that are involved in blood  pressure and blood volume regulation. Endocrine responses are  mediated through axonal projections from SON and PVN magnocellular  neurons (MCNs) to the capillaries of the posterior pituitary neural  lobe. In response to dehydration, a massive release of the  antidiuretic peptide hormone vasopressin (VP) into the circulation is  accompanied by a dramatic functional remodelling of the HNS. We have  used microarrays to comprehensively catalogue the genes expressed in  the PVN, the SON and the neurointermediate lobe (NIL) of the  pituitary gland. Further, we have identified transcripts that are  regulated as a consequence of dehydration, as well as RNAs that are  enriched in either the PVN or the SON. We suggest that these  differentially expressed genes represent candidate regulators and  effectors of HNS activity and remodelling.\nLast Updated (by provider): Jan 31 2006\nContributors:  GEMMA BEIGHTON JULIAN PATON SONG YAO CHARLES C HINDMARCH
#> 5:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      The hypothalamic paraventricular (PVN) and supraoptic (SON) nuclei  are important integrative structures that regulate co-ordinated  responses to perturbations in cardiovascular homeostasis. Through  descending projections from parvocellular neurons to the brainstem,  the PVN acts as an autonomic 'premotor nucleus', regulating reflex  changes in sympathetic nerve activity that are involved in blood  pressure and blood volume regulation. Endocrine responses are  mediated through axonal projections from SON and PVN magnocellular  neurons (MCNs) to the capillaries of the posterior pituitary neural  lobe. In response to dehydration, a massive release of the  antidiuretic peptide hormone vasopressin (VP) into the circulation is  accompanied by a dramatic functional remodelling of the HNS. We have  used microarrays to comprehensively catalogue the genes expressed in  the PVN, the SON and the neurointermediate lobe (NIL) of the  pituitary gland. Further, we have identified transcripts that are  regulated as a consequence of dehydration, as well as RNAs that are  enriched in either the PVN or the SON. We suggest that these  differentially expressed genes represent candidate regulators and  effectors of HNS activity and remodelling.\nLast Updated (by provider): Jan 31 2006\nContributors:  GEMMA BEIGHTON JULIAN PATON SONG YAO CHARLES C HINDMARCH
#> 6:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          Hypoxic-ischemic (HI) injury in the developing brain is a common cause of disability in children, and there are no effective treatments at this time.  Exposure to sublethal hypoxic conditions (hypoxic preconditioning) 24 hours prior to hypoxic-ischemic insult is protective in the developing rat model.  We have observed protective effects on brain histopathology and on long-term sensory-motor behavioral tasks.  Changes in gene expression are thought to underlie this protective effect.  By comparing gene expression in rats subjected to hypoxic preconditioning or sham conditioning at several time points from 0 to 24 hrs after preconditioning, we should gain insight into the mechanisms underlying these neuroprotective effects and may identify targets for therapeutic intervention. The aim of this study is to determine the effect of hypoxic preconditioning on global gene expression, and, in littermates, to examine the effect of hypoxic preconditioning 24 h prior to hypoxic-ischemic insult on brain histopathology. We hypothesize that changes in gene expression underlie the protective effect of hypoxic preconditioning against subsequent hypoxic-ischemic insult. Gene expression will be examined in two groups, 1) preconditioned and 2) sham controls, at 4 time points.  On postnatal day 6, preconditioned animals are exposed to normothermic hypoxia for 3 hrs (8.0% oxygen, 36 degrees C), and sham animals are simultaneouosly exposed to normoxia at 36 degrees C. Animals are then returned to their dams until euthanized at 4 time points (0h, 2h, 8h, and 24h later).  Five brains/group/timepoint will be used, with an equal number of males and females in each group.  Brains are removed and dissected on ice.  Cerebral cortex is dissected from both hemispheres and rapidly frozen on dry ice.  Total RNA is isolated using the QIAGEN RNeasy Protect Maxi Kit.  Littermates of these animals will be exposed to hypoxic preconditioning or sham preconditioning and subjected to hypoxic-ischemic injury 24 h later.  These animals are euthanized at postnatal day 14 for histopathologic evaluation of injury.\nNote: [] samples from this series, which appear in other Expression Experiments in Gemma, were not imported from the GEO source. The following samples were removed: \nLast Updated (by provider): May 01 2006\nContributors:  Mary-Ann Wilson
#>    experiment.Troubled experiment.Accession experiment.Database
#> 1:               FALSE              GSE2872                 GEO
#> 2:               FALSE              GSE2870                 GEO
#> 3:               FALSE              GSE4037                 GEO
#> 4:               FALSE              GSE3110                 GEO
#> 5:               FALSE              GSE3125                 GEO
#> 6:               FALSE              GSE4753                 GEO
#>                      experiment.URI experiment.SampleCount
#> 1: http://www.ncbi.nlm.nih.gov/geo/                     12
#> 2: http://www.ncbi.nlm.nih.gov/geo/                     18
#> 3: http://www.ncbi.nlm.nih.gov/geo/                     18
#> 4: http://www.ncbi.nlm.nih.gov/geo/                     10
#> 5: http://www.ncbi.nlm.nih.gov/geo/                     10
#> 6: http://www.ncbi.nlm.nih.gov/geo/                     40
#>    experiment.LastUpdated
#> 1:    2022-09-04 13:49:21
#> 2:    2022-08-25 21:15:14
#> 3:    2022-08-25 23:10:33
#> 4:    2022-08-19 23:36:07
#> 5:    2022-08-31 00:33:40
#> 6:    2022-08-23 01:17:50
#>                                         experiment.batchEffect
#> 1:                                        SINGLE_BATCH_SUCCESS
#> 2:                                        SINGLE_BATCH_SUCCESS
#> 3:                                        SINGLE_BATCH_SUCCESS
#> 4: This data set may have a batch artifact (PC 1), p=0.0098441
#> 5:                                               NO_BATCH_INFO
#> 6:                                        SINGLE_BATCH_SUCCESS
#>    geeq.batchCorrected geeq.batchConfound geeq.batchEffect geeq.rawData
#> 1:               FALSE                  1                1            1
#> 2:               FALSE                  1                1            1
#> 3:               FALSE                  1                1            1
#> 4:               FALSE                 -1                0            1
#> 5:               FALSE                  0                0           -1
#> 6:               FALSE                  1                1            1
#>    geeq.qScore geeq.sScore taxon.Name  taxon.Scientific taxon.ID taxon.NCBI
#> 1:   0.8552303       0.500        rat Rattus norvegicus        3      10116
#> 2:   0.5708228       0.500        rat Rattus norvegicus        3      10116
#> 3:   0.9964589       0.500        rat Rattus norvegicus        3      10116
#> 4:   0.5707173       0.750        rat Rattus norvegicus        3      10116
#> 5:   0.4267209       0.500        rat Rattus norvegicus        3      10116
#> 6:   0.8561355       0.625        rat Rattus norvegicus        3      10116
#>    taxon.Database.Name taxon.Database.ID
#> 1:                 rn6                86
#> 2:                 rn6                86
#> 3:                 rn6                86
#> 4:                 rn6                86
#> 5:                 rn6                86
#> 6:                 rn6                86