Dataset search

Does a search for datasets containing annotations, short name or full name matching the given string

searchDatasets(
  query = NA_character_,
  taxon = NA_character_,
  filter = NA_character_,
  offset = 0L,
  limit = 20L,
  sort = "+id",
  raw = getOption("gemma.raw", FALSE),
  async = getOption("gemma.async", FALSE),
  memoised = getOption("gemma.memoise", FALSE),
  file = getOption("gemma.file", NA_character_),
  overwrite = getOption("gemma.overwrite", FALSE)
)

Arguments

query

Required, defaults to empty. The search query. Either plain text ('traumatic'), or an ontology term URI ('http://purl.obolibrary.org/obo/UBERON_0002048'). Datasets that contain the given string in their short of full name will also be matched ('GSE201', 'Bronchoalveolar lavage samples'. Can be multiple identifiers separated by commas. When using in scripts, remember to URL-encode any forward slashes in the phenotype value URIs (see the compiled URL below).

taxon

Not required, part of the URL path. can either be Taxon ID, Taxon NCBI ID, or one of its string identifiers: scientific name, common name It is recommended to use Taxon ID for efficiency. Please note, that not all taxa have all the possible identifiers available. Use the 'All Taxa' endpoint to retrieve the necessary information. For convenience, below is a list of officially supported taxa: ID Comm.name Scient.name NcbiID

1 human Homo sapiens 9606 2 mouse Mus musculus 10090 3 rat Rattus norvegicus 10116 11 yeast Saccharomyces cerevisiae 4932 12 zebrafish Danio rerio 7955 13 fly Drosophila melanogaster 7227 14 worm Caenorhabditis elegans 6239

filter

Optional, defaults to empty. Filtering can be done on any* property or nested property that the appropriate object class defines or inherits (and that is mapped by hibernate). [These do not correspond to the properties of the objects returned by the API calls.].description-imp Class definitions:

Datasets: javaDoc gitHub

Platforms: javaDoc gitHub E.g: curationDetails or curationDetails.lastTroubledEvent.date.

Any property of a supported type. Currently supported types are:

String - property of String type, required value can be any String.

Number - any Number implementation. Required value must be a string parseable to the specific Number type.

Boolean - required value will be parsed to true only if the string matches 'true', ignoring case. Accepted operator keywords are:

'=' - equality

'!=' - non-equality

'<' - smaller than

'>' - larger than

'<=' - smaller or equal

'=>' - larger or equal

'like' - similar string, effectively means 'contains', translates to the sql 'LIKE' operator (given value will be surrounded by % signs) Multiple filters can be chained using AND and OR keywords. Leave space between the keywords and the previous/next word! E.g: ?filter=property1 < value1 AND property2 like value2 If chained filters are mixed conjunctions and disjunctions, the query must be in conjunctive normal form (CNF). Parentheses are not necessary

every AND keyword separates blocks of disjunctions. Example: ?filter=p1 = v1 OR p1 != v2 AND p2 <= v2 AND p3 > v3 OR p3 < v4 Above query will translate to: (p1 = v1 OR p1 != v2) AND (p2 <= v2) AND (p3 > v3 OR p3 < v4;) Breaking the CNF results in an error. Filter curationDetails.troubled will be ignored if user is not an administrator.

offset

Optional, defaults to 0. Skips the specified amount of objects when retrieving them from the database.

limit

Optional, defaults to 20. Limits the result to specified amount of objects. Use 0 for no limit.

sort

Optional, defaults to +id. Sets the ordering property and direction. Format is [+,-][property name]. E.g. -accession will translate to descending ordering by the 'Accession' property. Note that this does [not guarantee the order of the returned entities!].description-imp This is merely a signal to how the data should be pre-sorted before the limit and offset are applied. Nested properties are also supported (recursively). E.g: +curationDetails.lastTroubledEvent.date When using in scripts, remember to URL-encode the '+' plus character (see the compiled URL below).

raw

TRUE to receive results as-is from Gemma, or FALSE to enable parsing.

async

TRUE to run the API query on a separate worker, or FALSE to run synchronously. See the async package for details.

memoised

Whether or not to cache results so future requests for the same data will be faster. Use forgetGemmaMemoised to clear the cache.

file

The name of a file to save the results to, or NULL to not write results to a file. If raw == TRUE, the output will be a JSON file. Otherwise, it will be a RDS file.

overwrite

Whether or not to overwrite if a file exists at the specified filename.

Value

An array of datasets (expression experiment value objects) that are annotated with the given ontology terms, or, in case of plaintext query, experiments that contain the given words (name, short name, accession, tags) If an ontology term URI is given, the results will also include datasets that are associated with the descendants of the term. The search only only checks the annotations value, not the category (which is also an ontology term).

Examples

dat <- searchDatasets("bipolar")
str(dat)
#> Classes ‘data.table’ and 'data.frame':	20 obs. of  18 variables:
#>  $ ee.ShortName       : chr  "GSE5389" "GSE4030" "GSE5388" "GSE7036" ...
#>  $ ee.Name            : chr  "Adult postmortem brain tissue (orbitofrontal cortex) from subjects with bipolar disorder and healthy controls" "bunge-affy-arabi-162779" "Adult postmortem brain tissue (dorsolateral prefrontal cortex) from subjects with bipolar disorder and healthy controls" "Expression profiling in monozygotic twins discordant for bipolar disorder" ...
#>  $ ee.ID              : int  326 348 594 660 670 672 700 831 833 834 ...
#>  $ ee.Public          : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ ee.Troubled        : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ ee.Accession       : chr  "GSE5389" "GSE4030" "GSE5388" "GSE7036" ...
#>  $ ee.Database        : chr  "GEO" "GEO" "GEO" "GEO" ...
#>  $ ee.URL             : chr  "http://www.ncbi.nlm.nih.gov/geo/" "http://www.ncbi.nlm.nih.gov/geo/" "http://www.ncbi.nlm.nih.gov/geo/" "http://www.ncbi.nlm.nih.gov/geo/" ...
#>  $ ee.Samples         : int  21 8 61 6 23 66 8 50 102 47 ...
#>  $ ee.LastUpdated     : POSIXct, format: "2021-09-18 07:31:21" "2021-09-18 07:31:26" ...
#>  $ ee.Batch.P         : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ ee.Batch.PC        : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ geeq.batchCorrected: logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ geeq.qScore        : num  0.852 0.426 0.853 0.712 0.282 ...
#>  $ geeq.sScore        : num  0.875 0.838 0.625 0.838 0.25 ...
#>  $ taxon.Name         : chr  "human" "human" "human" "human" ...
#>  $ taxon.ID           : int  1 1 1 1 1 1 3 1 1 1 ...
#>  $ technology.Type    : logi  NA NA NA NA NA NA ...
#>  - attr(*, ".internal.selfref")=<externalptr>